% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.area.R
\name{aggregate.area}
\alias{aggregate.area}
\alias{aggregateArea}
\title{Aggregate Functions for Climate Data}
\usage{
\method{aggregate}{area}(
  x,
  ...,
  is = NULL,
  it = NULL,
  FUN = "sum",
  na.rm = TRUE,
  smallx = FALSE,
  verbose = FALSE,
  a = 6378,
  threshold = NULL
)
}
\arguments{
\item{x}{A \code{\link{station}} object or similar climate data object.}

\item{\dots}{Additional arguments passed to the function.}

\item{is}{A list or data.frame specifying the spatial index, see \code{\link{subset}}.}

\item{it}{A list or data.frame specifying the time index, see \code{\link{subset}}.}

\item{FUN}{A function to apply, such as 'sum' or 'mean'.}

\item{na.rm}{Logical; if TRUE, missing values (\code{NA}) are ignored, see \code{\link{mean}}.}

\item{smallx}{Logical; if TRUE, reduces the size of the data, default is FALSE.}

\item{verbose}{Logical; if TRUE, prints detailed progress messages for debugging.}

\item{a}{Radius of the Earth (in km), used for spatial weighting.}

\item{threshold}{Numeric; a threshold value used if \code{FUN} is 'area', 'exceedance', or 'lessthan'.}
}
\value{
Returns a \code{station} object or a similar climate data object with aggregated results.
}
\description{
These aggregation functions are based on the S3 method for \code{zoo} objects
and handle additional tasks such as maintaining attributes with metadata.
}
\details{
The function \code{aggregate.area} is specifically designed for aggregating
spatial statistics, such as calculating the global mean or the total area of a phenomenon.

The function \code{aggregateArea} is an alias for \code{aggregate.area}.
}
\examples{

## Aggregation for a station object
data(Svalbard)
x <- aggregate(Svalbard, month, FUN = 'mean', na.rm = TRUE)
plot(x)

## Aggregation for a field object
slp <- slp.DNMI()
y <- aggregate(slp, year, FUN = 'mean', na.rm = FALSE)

## Spatial aggregation
w <- aggregate.area(y)
plot(w)

}
\seealso{
\code{\link{aggregate}}, \code{\link{aggregate.size}}
}
\author{
R.E. Benestad
}
\keyword{utilities}
