% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.dsensemble.R
\name{aggregate.dsensemble}
\alias{aggregate.dsensemble}
\title{aggregate.dsensemble}
\usage{
\method{aggregate}{dsensemble}(
  x,
  ...,
  it = NULL,
  im = NULL,
  FUN = NULL,
  FUNX = "mean",
  verbose = FALSE,
  anomaly = FALSE,
  qc = TRUE,
  test = FALSE,
  eof = TRUE
)
}
\arguments{
\item{x}{an object of type 'dsensemble'}

\item{it}{time index (see \code{\link{subset}})}

\item{FUN}{function applied to aggregate in time}

\item{FUNX}{function applied aggregate ensemble members}

\item{verbose}{if TRUE print progress}

\item{anomaly}{if FALSE add the mean value stored as attribute in x}

\item{qc}{if TRUE perform quality control and exclude ensemble members with incomplete or suspect data}

\item{test}{if TRUE perform test on one GCM simulation}

\item{eof}{if TRUE and if the dsensemble object contains a spatially aggregated pattern, x$eof, 
expand with x$eof into an ensemble of field objects to which FUNX is applied. 
If FALSE, expand using the original pattern x$pca into an ensemble of station objects to be aggregated.}
}
\description{
The function \code{aggregate.dsensemble} is used to extract and aggregate information from PCA-based dsensemble-objects, 
to expand PCA into station data or EOFs into field data.
It applies operations through the argument 'FUNX' to estimate quantities such as the ensemble mean and 
takes care of extra house keeping, such as attributes with meta data. \code{aggregate.dsensemble}
is similar to \code{expandpca} but applies 'FUNX' for each time step after it has been expanded by matrix multiplication 
of the singular vectors (X = U %*% diag(W) %*% t(V)) and is therefore slower. 
\code{map.dsensemble} is a wrapper for \code{map} that uses one of the two former routines to distill selected
information.
}
\seealso{
aggregate expandpca map.dsensemble
}
