% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.field}
\alias{aggregate.field}
\title{Aggregate S3 method for a field object}
\usage{
\method{aggregate}{field}(
  x,
  by,
  FUN = "mean",
  na.rm = TRUE,
  regular = NULL,
  frequency = NULL,
  threshold = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{field}}.}

\item{by}{Index vector of the same length as \code{index(x)} which defines aggregation groups and the new index to be associated with each group. If \code{by} is a function, then it is applied to \code{index(x)} to obtain the aggregation groups.}

\item{FUN}{a function to compute the summary statistics which can be applied to all subsets. Always needs to return a result of fixed length (typically scalar) (default: 'mean', additional options: 'area','exceedance','lessthan').}

\item{na.rm}{TRUE: ignore NA - see see \code{\link{mean}}}

\item{regular}{logical. The default is \code{FALSE} for \code{"zoo"} series and \code{TRUE} for \code{"zooreg"} series.}

\item{frequency}{numeric indicating the frequency of the aggregated series (if a \code{"zooreg"} series should be returned. The default is to determine the frequency from the data if \code{regular} is \code{TRUE}. If \code{frequency} is specified, it sets \code{regular} to \code{TRUE}.}

\item{threshold}{threshold used if FUN is 'count', 'freq', 'wetfreq', or 'wetmean'}

\item{verbose}{if TRUE print progress}

\item{\dots}{additional arguments passed to \code{FUN}.}
}
\description{
Aggregate S3 method for a field object
}
\examples{

## S3 method for class 'field'
slp <- slp.DNMI()
y <- aggregate(slp, year, FUN='mean', na.rm=FALSE)

}
