% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.grid.R
\name{aggregate.grid}
\alias{aggregate.grid}
\title{aggregate.grid}
\usage{
\method{aggregate}{grid}(x, ..., is, FUN = "mean", na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A \code{field} object with finer grid mesh.}

\item{\dots}{additional arguments}

\item{is}{A list or \code{field} with the coarser grid with aggregated data from finer mesh.}

\item{FUN}{A function, e.g., 'sum' or 'mean'}

\item{na.rm}{a boolean; if TRUE ignore NA, see \code{\link{mean}}}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
The call returns a field object
}
\description{
The aggregation functions are based on the S3 method for \code{zoo} objects,
but takes care of extra house keeping, such as attributes with meta data.
Use given longitude-latitude coordinates to find the grid-boxes which are within these bins and
aggregate them as with a data.frame object. This is a bit similar to regrid, but does not 
use bilinear interpolation. but the mean of grid boxes within a larger grid-box. 
Rasmus E. Benestad
}
\details{
\code{aggregate.grid} is used for aggregating spatial statistics of a finer grid-mesh onto a coarser grid.
}
\examples{

test.aggregate.grid()

}
\seealso{
aggregate.area regrid
}
\author{
R.E. Benestad
}
\keyword{utilities}
