% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.size.R
\name{aggregate.size}
\alias{aggregate.size}
\alias{aggregateSize}
\alias{aggregateSize.matrix}
\alias{aggregateSize.field}
\title{aggregate}
\usage{
\method{aggregate}{size}(x, ...)
}
\arguments{
\item{x}{A \code{\link{station}} object}

\item{\dots}{additional arguments}

\item{x0}{threshold defining an event}

\item{plot}{a boolean; if TRUE display results as a plot}

\item{a}{radius of earth (unit: km)}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
The call returns a station object
}
\description{
The aggregation functions are based on the S3 method for \code{zoo} objects,
but takes care of extra house keeping, such as attributes with meta data.
}
\details{
\code{aggregate.size} is similar to \code{aggregate.area}, but returns the size statistics (square
meters) for individual events (defined as gridboxes touching each other).

The function \code{aggregateSize} is exactly the same as \code{aggregate.size}.
}
\seealso{
aggregate.area aggregate
}
\author{
R.E. Benestad
}
\keyword{utilities}
