% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.station}
\alias{aggregate.station}
\alias{aggregate}
\alias{aggregate.comb}
\alias{aggregate.field}
\title{aggregate}
\usage{
\method{aggregate}{station}(x, by, FUN = "mean", na.rm = TRUE,
  regular = NULL, ..., frequency = NULL, verbose = FALSE,
  threshold = 1)
}
\arguments{
\item{x}{A \code{\link{station}} object}

\item{by}{see \code{\link{aggregate.zoo}}}

\item{FUN}{a function; see \code{\link{aggregate.zoo}}. Additional options: 'area','exceedance','lessthan'.}

\item{na.rm}{TRUE: ignore NA - see see \code{\link{mean}}}

\item{regular}{see \code{\link{aggregate.zoo}}}

\item{frequency}{see \code{\link{aggregate.zoo}}}
}
\value{
The call returns a station object
}
\description{
The aggregation functions are based on the S3 method for \code{zoo} objects,
but takes care of extra house keeping, such as attributes with meta data.
}
\details{
\code{aggregate} calculates a time aggregate of an input object,
e.g, the mean seasonal cycle (if \code{by}=month and \code{FUN}="mean") or the
annual sum (if \code{by}=year and \code{FUN}="sum").

\code{aggregate.area} is used for aggregating spatial statistics, such as
the global mean or the global area of some phenomenon.
\code{aggregate.size} is similar to \code{aggregate.area}, but returns the size statistics (square
meters) for individual events (defined as gridboxes touching each other).
}
\examples{

## S3 method for class 'station'
data(Svalbard)
x <- aggregate(Svalbard, month, FUN='mean', na.rm=TRUE)
plot(x)

## S3 method for class 'field'
slp <- slp.DNMI()
y <- aggregate(slp, year, FUN='mean', na.rm=FALSE)

## Aggregate area
w <- aggregate.area(y)
plot(w)

}
\seealso{
aggregate.area aggregate.size aggregate.size.matrix aggregate.size.field
}
\author{
R.E. Benestad
}
\keyword{utilities}
