% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESD-package.R
\docType{data}
\name{albGIMMS}
\alias{albGIMMS}
\title{Schwaebische Alb GIMMS NDVI3g.v1}
\format{\code{raster::RasterBrick}}
\description{
Schwaebische Alb half-monthly GIMMS NDVI3g.v1 (2012 to 2015).
}
\details{
This dataset contains half-monthly GIMMS NDVI3g.v1 observations for
parts of the Schwaebische Alb, south-western Germany (Jan 2012 to Dec 2015).
Preprocessing included

\itemize{
\item{the conduction of a comprehensive quality control based on the
accompanying quality flags followed by}
\item{the application of a modified Whittaker smoother to account for
non-captured clouds and fill newly introduced data gaps.}
}
}
\references{
Pinzon JE, Tucker CJ (2014) A Non-Stationary 1981-2012 AVHRR NDVI3g Time
Series. Remote Sensing, 6(8), 6929-6960. Available online:
\url{http://www.mdpi.com/2072-4292/6/8/6929/htm}.
}

