% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgood.R
\name{allgood}
\alias{allgood}
\alias{is.T}
\alias{is.precip}
\alias{is.field(}
\alias{is.station}
\alias{is.eof}
\alias{is.pca}
\alias{is.cca}
\alias{is.trajectory}
\alias{is.daily}
\alias{is.monthly}
\alias{is.seasonal}
\alias{is.annual}
\alias{is.model}
\alias{is.wind}
\alias{is.direction}
\alias{is.pressure}
\title{Test for .}
\usage{
allgood(x, miss = 0.1, verbose = FALSE)
}
\arguments{
\item{x}{a data object}
}
\value{
Boolean
}
\description{
Computes different formulas
}
\examples{

data(ferder)
is.T(ferder)

}
\author{
R. Benestad, MET Norway
}
\keyword{parameter,element}
