% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.annual.R
\name{annual}
\alias{annual}
\alias{as.annual}
\alias{as.annual.default}
\alias{as.annual.numeric}
\alias{as.annual.integer}
\alias{as.annual.yearqtr}
\alias{as.annual.station}
\alias{as.annual.spell}
\alias{annual.zoo}
\alias{annual.default}
\alias{annual.dsensemble}
\alias{annual.station}
\alias{annual.spell}
\alias{annual.field}
\alias{as.monthly}
\alias{as.monthly.default}
\alias{as.monthly.station}
\alias{as.monthly.field}
\alias{as.4seasons}
\alias{as.4seasons.default}
\alias{as.4seasons.day}
\alias{as.4seasons.station}
\alias{as.4seasons.spell}
\alias{as.4seasons.field}
\alias{as.4seasons.dsensemble}
\alias{as.seasons}
\title{Conversion to esd objects.}
\usage{
annual(x, ...)
}
\arguments{
\item{x}{input object, e.g., a 'station' or 'field' object}

\item{\dots}{additional arguments}

\item{FUN}{a function}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{nmin}{minimum number of data points in a season}

\item{threshold}{threshold used if FUN is, e.g., 'exceedance' or}

\item{slow}{a boolean; if FALSE run a fast version that might not work on all data types?}

\item{dateindex}{a boolean; if TRUE transform index into date format}

\item{start}{first month and day of the season, e.g., '01-01' for January 1 (argument in \code{season})}

\item{end}{last month and day of the season, e.g., '12-31' for December 31 (argument in \code{season})}

\item{x}{an input object of, e.g., class 'station', 'field'}

\item{FUN}{a function, see \code{\link{aggregate.zoo}}}

\item{nmin}{Minimum number of data points (e.g. days or months) with valid
data accepted for annual estimate. NULL demands complete years.}

\item{format}{'numeric' or 'character'}

\item{na.rm}{a boolean; if TRUE, ignore NA - see see \code{\link{mean}}}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional argument}
}
\value{
Same class as x
}
\description{
\code{as.annual} and \code{annual} aggregates time series into annual values (e.g. means).
}
\details{
\code{as.monthly} aggregates time series into monthly values (e.g. means).

\code{as.4seasons} aggregates to four seasons ('djf': December-February, 'mam': March-May, 'jja': June-August, 'son': September-November)
and \code{as.seasons} aggregates to a user defined season (see arguments 'start' and 'end').
}
\examples{
data(ferder)
plot(annual(ferder,FUN="min"))
plot(annual(ferder,FUN="IQR",na.rm=TRUE))

data(bjornholt)
plot(as.4seasons(bjornholt,threshold=1,FUN="exceedance"))

}
\seealso{
aggregate

\code{\link{as.annual}} \code{\link{aggregate.station}} \code{\link{pentad}}
}
\keyword{utilities}
