% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.lonlat.R
\name{approx.lonlat}
\alias{approx.lonlat}
\title{Extention of \code{\link{approx}} for longitude and latitude}
\usage{
approx.lonlat(lon, lat, n = 20, a = 6378000, verbose = FALSE)
}
\arguments{
\item{lon}{longitudes}

\item{lat}{latitudes}

\item{n}{length of output}

\item{a}{the radius of the earth (unit: m)}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Linearly interpolate longitudes and latitudes to n equally spaced points spanning 
the interval (min(lon), max(lon)) and (min(lat), max(lat)).
}
