% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaClimateIndicator.R
\name{AreaClimateIndicator}
\alias{AreaClimateIndicator}
\title{Read daily station data of the Norwegian Meteorological Institute from thredds netCDF
using OpenDAP}
\usage{
AreaClimateIndicator(
  param = "precip",
  is = "Global",
  it = "annual",
  FUN = "mean",
  threshold = 1,
  plot = TRUE,
  verbose = FALSE,
  url = "https://thredds.met.no/thredds/dodsC/metusers/rasmusb/"
)
}
\arguments{
\item{param}{The element to read c('t2m','tmax','tmin','precip')}

\item{is}{the area or index space to select domain c('Global','50S-50N','Arctic')}

\item{it}{Index time to select times scale c('as.annual','as.4seasons','as.monthly','day')}

\item{FUN}{is the function used in aggregation}

\item{threshold}{select area of precipitation exceeding 1 mm/day or temperature above 0C (the freezing point)}
}
\description{
\code{AreaClimateIndicator} returns area-based climate indicators provided through the Met Norway thredds server,
some of which are described in Benestad et al (2022)  
"Global hydro-climatological indicators and changes in the global hydrological cycle and rainfall patterns", 
PLOS Climate, PCLM-D-21-00079R1, DOI: 10.1371/journal.pclm.0000029, and an extension of those to temperature and other thresholds.
}
\examples{
## Get the daily minimum temperature for Oslo-Blindern (station ID 18700)
\dontrun{
x <- AreaClimateIndicator()
y <- AreaClimateIndicator(param='t2m',is='Arctic',threshold=0)
z <- AreaClimateIndicator(param='precip',is='50S-50N',it='day',threshold=1)
w <- AreaClimateIndicator(param='t2m',it='max',threshold=40)
}
}
\seealso{
NINO3.4, NAO, GSL, QBO, CET, CO2, AMO, IOD,  aggregate
}
