% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.appended.R
\name{as.appended}
\alias{as.appended}
\alias{as.appended.ds.comb}
\alias{as.appended.eof.comb}
\alias{as.appended.field.comb}
\alias{as.calibrationdata}
\alias{as.calibrationdata.ds}
\alias{as.calibrationdata.station}
\alias{as.fitted.values}
\alias{as.fitted.values.ds}
\alias{as.fitted.values.station}
\alias{as.original.data}
\alias{as.original.data.ds}
\alias{as.original.data.station}
\alias{as.pattern}
\alias{as.pattern.ds}
\alias{as.pattern.corfield}
\alias{as.pattern.eof}
\alias{as.pattern.cca}
\alias{as.pattern.mvr}
\alias{as.pattern.field}
\title{Extract attributes}
\usage{
as.appended(x, ...)
}
\arguments{
\item{x}{input object with appendix (of class \code{comb})}

\item{\dots}{additional arguments}

\item{iapp}{index of appendix to extract}
}
\description{
\code{as.appended} extracts an appendix (e.g, attr(x,'appendix.1')) and copies the other attributes from the main object to the appendix.
}
\details{
\code{as.calibrationdata} extracts the calibration data from a DS object.
Note: if DS was performed with rm.trend=TRUE, the calibration data are detrended.

\code{as.fitted.values} extracts fitted values from a DS object.

\code{as.original.data} extracts the original data (the predictand) from a DS object.

\code{as.pattern} extracts a spatial pattern, e.g., from an 'eof' or 'cca' object.
}
