% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.comb.R
\name{as.comb}
\alias{as.comb}
\alias{as.comb.eof}
\title{Coerce input to a \code{comb} object}
\usage{
as.comb(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{the input object}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{...}{other arguments}
}
\value{
an \code{eof} object
}
\description{
Transform an \code{eof} object into the esd class \code{comb} by applying \code{eof2field} to the input object and its appendices.
}
\seealso{
eof2field
}
