% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{as.decimal}
\alias{as.decimal}
\alias{nv}
\alias{cv}
\alias{q5}
\alias{q95}
\alias{q975}
\alias{q995}
\alias{filt}
\alias{filt.default}
\alias{exit}
\alias{figlab}
\alias{ndig}
\alias{ensemblemean}
\alias{propchange}
\alias{stand}
\alias{rmse}
\alias{RMSE}
\alias{firstyear}
\alias{lastyear}
\alias{eofvar}
\alias{test.num.predictors}
\alias{arec}
\alias{arec.default}
\alias{arec.station}
\alias{lastrains}
\alias{lastdry}
\alias{lastelementrecord}
\alias{strstrip}
\alias{bin}
\alias{factor2numeric}
\alias{zeros}
\alias{missval}
\alias{leapyear}
\alias{distance2ocean}
\alias{difftime.month}
\title{Simple and handy functions}
\usage{
as.decimal(x = NULL)
}
\arguments{
\item{x}{A data.frame or a coredata zoo object.}

\item{na.rm}{If TRUE, remove NA's from data}

\item{type}{'ma' for moving average (box-car), 'gauss' for Gaussian, 'binom'
for binomial filter,' parzen' for Parzen filter, 'hanning' for Hanning
filter, or 'welch' for Welch filter.}

\item{lowpass}{True for smoothing, otherwise the highpass results is
returned}

\item{triangle}{a group of three stations with sea-level pressure', e.g.
from ECA\&D.}

\item{nbins}{number of bins/categories}
}
\value{
\item{as.decimal }{Decimal value} \item{trend.coef }{Linear trend
per decade}
}
\description{
\code{attrcp(x,y)} passes on attributes from \code{x} to \code{y} and
returns the \code{y} with new attributes.
}
\details{
\code{ensemblemean} returns the ensemble mean for dsensemble objects. The argument \code{FUN} can also be
used to estimate other statistics, e.g \code{FUN='q95'} where
\code{q95=function(x) apply(x,1,quantile,probs=0.95)}

\code{TGW} uses triangulation of pressure measurements to estimate geostrophic wind based on
Alexandersson et al. (1998), Glob. Atm. and Oce. Sys.

\code{stand} gives a standardised time series.

\code{zeros} counts the occurrence of zero values in a vector.

\code{nv} returns the number of valid points.

\code{missval} computes the percentage of missing data.

\code{arec} compares the number of record-breaking events to the number of
events for a time series of iid data (\code{sum(1/1:n)})

\code{strstrip} strips off spaces before and after text in a string.

\code{as.decimal} converts between degree-minute-second into decimal value.

\code{cv} computes the coefficient of variation.

\code{nv} count the number of valid data points.

\code{q5}, \code{q95}, \code{q975} and \code{q995} are shortcuts to the 5\%, 95\%, 97.5\% and 99.5\% percentiles.

\code{trend.coef} and \code{trend.pval} return the coefficient and the p-value of the linear trend.

\code{exit} is a handy function for exiting the R session without saving.

\code{figlab} is a handy function for labelling figures (e.g. 'Figure 1')

\code{ndig} estimates the number of digits for round(x,ndig), e.g. in scales for plotting.

\code{factor2numeric} transforms a factor to a numeric object

\code{rmse} and \code{RMSE} calculate the root-mean-square error

\code{leapyear} checks if a year is a leap year

\code{distance2ocean} estimates the distance from a point to the ocean

\code{lastyear} checks the availability of data of a station object and returns the last year(s) of valid data entries

\code{difftime.month} returns the number of months between two dates
}
\examples{

## Monthly mean temperature at 'Oslo - Blindern' station 
data(Oslo)
## Compute the linear trend and the p-value on annual aggregated values 
tr <- trend.coef(coredata(annual(Oslo)))
pval <- trend.pval(coredata(annual(Oslo)))
\dontrun{
pp <- station(param='slp',cntr='Denmark',src='ecad')
wind <- TGW(subset(pp,is=c(1,3,10))
plot(wind)
ws <- sqrt(wind[,1]^2 + wind[,2]^2)
plot(ws)
hist(ws)

## Estimate wind for a larger group of stations
wind <- geostrophicwind(pp,nmax=10)
u <- subset(wind,param='u')
v <- subset(wind,param='u')
ws <- sqrt(u^2+v^2)
ws <- attrcp(v,ws)
class(ws) <- class(v)
attr(ws,'variable')='windspeed'
attr(ws,'longname')='geostrophic wind speed'
map(ws,FUN='quantile',probs=0.98)

## Test firstyears on HadCRUT4
if (!file.exists('~/Downloads/HadCRUT.4.6.0.0.median.nc')) {
  print('Download HadCRUT4')
  download.file('https://crudata.uea.ac.uk/cru/data/temperature/HadCRUT.4.6.0.0.median.nc',
                dest='~/Downloads/HadCRUT.4.6.0.0.median.nc') 
}

Obs <- annual(retrieve('~/Downloads/HadCRUT.4.6.0.0.median.nc',param='temperature_anomaly'))
lons <- rep(lon(Obs),length(lat(Obs)))
lats <- sort(rep(lat(Obs),length(lon(Obs))))
fy <- firstyear(Obs)
map(subset(Obs,it=1))
points(lons[fy==1850],lats[fy==1850])
map(Obs,FUN='firstyear')
}

## Check for leap years
leapyear(1999)
leapyear(2000)

## Calculate the distance to the ocean from the point 10E/60N 
distance2ocean(10, 60)

}
\seealso{
attrcp
}
\author{
A. Mezghani
}
\keyword{rtools}
\keyword{~kwd2}
