% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.eof.R
\name{as.eof}
\alias{as.eof}
\title{Coerce input to an \code{eof} object}
\usage{
as.eof(x, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}

\item{iapp}{index of appendix}
}
\value{
an \code{eof} object
}
\description{
Transform an input object into the esd class \code{eof} (see \code{\link{EOF}}).
}
\details{
\code{as.eof} is an S3 method and will redirect to a fitting function depending on
the class of the input object.

\code{as.eof.dsensemble.pca} converts PCA-based DSensemble objects to EOF-based results (gridded)
}
\seealso{
as.eof as.eof.field as.eof.comb as.eof.list as.eof.zoo as.eof.ds
as.eof.dsensemble as.eof.appendix as.eof.dsensemble.pca
}
