% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.events.R
\name{as.events}
\alias{as.events}
\alias{as.events.default}
\alias{as.events.trajectory}
\alias{events}
\alias{trajectory2events}
\title{Transform an input object into an \code{events} object}
\usage{
as.events(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{input object}

\item{verbose}{if TRUE print progress}

\item{\dots}{additional arguments}
}
\description{
Transform an input object into an \code{events} object
}
