% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.comb}
\alias{as.field.comb}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{comb}(x, ..., iapp = NULL, verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{field} \code{comb}}

\item{...}{other arguments}

\item{iapp}{a numerical; an index representing the appendix to extract}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform a combined field object into a \code{field} object, either dropping the appendices (if ip=NULL) or selecting one of the appended fields.
}
\seealso{
as.field
}
