% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.ds}
\alias{as.field.ds}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{ds}(x, ..., iapp = NULL, verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{ds}}

\item{...}{other arguments}

\item{iapp}{a numerical; an index representing the appendix to extract}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform \code{ds} object (output of the function \code{DS}) into the esd class \code{field}.
}
\seealso{
as.field as.field.eof DS
}
