% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.dsensemble.eof}
\alias{as.field.dsensemble.eof}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{dsensemble.eof}(
  x,
  ...,
  is = NULL,
  ip = NULL,
  im = NULL,
  anomaly = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{the input object of class \code{ds}}

\item{...}{other arguments}

\item{is}{a list or data.frame providing a space index, e.g., station record or a lon(gitude) and lat(itude) range. If NULL include all.}

\item{ip}{a numerical or numerical vector with indices of the principle components to be included. If NULL include all.}

\item{im}{a numerical or numerical vector with indices of the ensemble members to be included. If NULL include all.}

\item{anomaly}{a boolean; if TRUE return anomalies (climatology is in attribute 'mean')}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{dsensemble} \code{field} \code{list} object, i.e., a list where each element is a downscaled field corresponding to an ensemble member
}
\description{
Transform a \code{dsensemble} \code{eof} object (output of the function \code{DS}) into the esd class \code{dsensemble} \code{field} \code{list}.
The function uses the downscaled principle components with the corresponding spatial EOF patterns and eigenvalues to calculate downscaled fields.
This is done for all selected ensemble members (see input \code{im}).
}
\seealso{
as.field DSensemble EOF
}
