% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.eof}
\alias{as.field.eof}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{eof}(x, ..., iapp = NULL, anomaly = FALSE,
  verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{eof}}

\item{...}{other arguments}

\item{iapp}{a numerical; an index representing the appendix to extract}

\item{anomaly}{a boolean; if TRUE return anomalies (climatology is in attribute 'mean')}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform an \code{eof} object into the esd class \code{field}. If the input is a \code{dsensemble} object, it will be redirected to \code{as.field.dsensemble}. Otherwise the object is transformed using the function \code{eof2field}.
}
\seealso{
as.field eof2field EOF as.field.dsensemble.eof
}
