% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field}
\alias{as.field}
\title{Coerce input to a \code{field} object}
\usage{
as.field(x, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}
}
\value{
a \code{field} object
}
\description{
Transform an input object into the esd class \code{field}. 
\code{as.field} is an S3 method and will redirect to a fitting function depending on the output. 
The way in which the transformation is performed depends on the type of input data.
}
\seealso{
as.field.default as.field.zoo as.field.eof as.field.comb as.field.field as.field.ds as.field.station as.field.events as.field.trajectory as.field.dsensemble.eof
}
