% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.station}
\alias{as.field.station}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{station}(x, ..., lon = NULL, lat = NULL, nx = 30,
  ny = 30, verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{ds}}

\item{...}{other arguments}

\item{lon}{a numerical vector of longitudes defining the grid of the output field}

\item{lat}{a numerical vector of latitudes defining the grid of the output field}

\item{nx}{number of grid points in the longitude direction - only used if lon=NULL}

\item{ny}{number of grid points in the latitude direction - only used if lat=NULL}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform a \code{station} object (output of the function \code{DS}) into the esd class \code{field} using the function \code{regrid}.
}
\seealso{
as.field regrid
}
