% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field.zoo}
\alias{as.field.zoo}
\title{Coerce input to a \code{field} object}
\usage{
\method{as.field}{zoo}(
  x,
  ...,
  lon,
  lat,
  param,
  unit,
  longname = NA,
  quality = NA,
  src = NA,
  url = NA,
  reference = NA,
  info = NA,
  calendar = "gregorian",
  greenwich = TRUE,
  method = NA,
  type = NA,
  aspect = NA,
  verbose = FALSE
)
}
\arguments{
\item{x}{the input object of class \code{zoo} typically containing data from one or several measurement stations}

\item{...}{other arguments}

\item{lon}{longitude(s), a numerical or numerical vector}

\item{lat}{latitudes(s), a numerical or numerical vector}

\item{param}{short name of variable}

\item{unit}{unit of variable, e.g., 't2m'}

\item{longname}{long name of variable, e.g, 'temperature at 2m'}

\item{quality}{quality flag}

\item{src}{source of data}

\item{url}{url to website where data can be downloaded}

\item{reference}{reference describing data set}

\item{info}{additional information}

\item{calendar}{calendar type}

\item{greenwich}{a boolean; if TRUE center map on the Greenwich line (0E)}

\item{method}{method applied to data}

\item{type}{type of data}

\item{aspect}{aspect describing data, e.g., 'original', 'anomaly', 'climatology'}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform a \code{zoo} object into a \code{field} object
}
\seealso{
as.field
}
