% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.pca.R
\name{as.pca.ds}
\alias{as.pca.ds}
\title{Coerce input to a \code{pca} object}
\usage{
\method{as.pca}{ds}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{the input object}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{...}{other arguments}
}
\value{
a \code{pc} object
}
\description{
Coerce a \code{ds} \code{pca} object into a \code{pca} object by replacing the class.
}
\seealso{
as.pca DS
}
