% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.pca.R
\name{as.pca.station}
\alias{as.pca.station}
\title{Coerce input to a \code{pca} object}
\usage{
\method{as.pca}{station}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}
}
\value{
a \code{pc} object
}
\description{
Transform a \code{station} object into a \code{pca} object using the function \code{PCA}.
}
\seealso{
as.pca PCA
}
