% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.residual.R
\name{as.residual}
\alias{as.residual}
\alias{as.residual.ds}
\alias{as.residual.station}
\alias{as.residual.field}
\title{Calculate residual}
\usage{
as.residual(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a 'ds' object}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}
}
\description{
Caluculate the residual of a 'ds' object, i.e., the original data minus the fitted values.
}
