% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.stand.R
\name{as.stand}
\alias{as.stand}
\alias{as.stand.station}
\alias{as.original}
\alias{as.original.station}
\title{Normalise data}
\usage{
as.stand(x, ...)
}
\arguments{
\item{x}{a station object}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{na.rm}{a boolean; if TRUE remove NA values}
}
\description{
\code{as.stand} returns normalised values:
If the input contains precipitation data the data are normalised by the mean value.
If the input contains temperature data, the data are stanardised by subtracting the mean and dividing by the standard deviation.
\code{as.original} transforms normalised data to its original values.
}
