% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.ds}
\alias{as.station.ds}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{ds}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{ds}}

\item{...}{other arguments}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform a \code{ds} object (empirical-statistical downscaling output of the \code{DS} function) into a \code{station} object
}
\seealso{
as.station DS
}
