% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.dsensemble.pca}
\alias{as.station.dsensemble.pca}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{dsensemble.pca}(x, ..., is = NULL, ip = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{dsensemble} \code{pca}}

\item{...}{other arguments}

\item{is}{a list or data.frame providing a space index, e.g., station record or a lon(gitude) and lat(itude) range}

\item{ip}{a numerical or numerical vector with indices of the principle components to be included}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{dsensemble} \code{list} or \code{dsensemble} \code{station} object
}
\description{
Transform a \code{dsensemble} \code{pca} object to a \code{dsensemble} \code{list} or \code{dsensemble} \code{station} object
holding the downscaled ensemble for each station, calculated based on the corresponding ensemble of downscaled principle components
}
\seealso{
as.station DSensemble PCA
}
