% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.dsensemble.station}
\alias{as.station.dsensemble.station}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{dsensemble.station}(x, ..., is = NULL, it = NULL,
  FUN = "mean", verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{dsensemble}}

\item{...}{other arguments}

\item{is}{a list or data.frame providing a space index, e.g., station record or a lon(gitude) and lat(itude) range}

\item{it}{a list providing a time index such as months (c("dec","jan","feb")), seasons ("djf"), a range of years or dates, or a logical or numerical vector providing the elements to be selected (1:10 for the first ten elements)}

\item{FUN}{a function, applied to the downscaled ensemble}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform a \code{dsensemble} \code{pca} object into a \code{station} object holding the ensemble mean
of the downscaled results for each station (or another statistical characteristic of the ensemble, see input argument \code{FUN}).
}
\seealso{
as.station DSensemble
}
