% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.eof}
\alias{as.station.eof}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{eof}(x, ..., ip = 1:10, verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{eof}}

\item{...}{other arguments}

\item{ip}{a numerical or numerical vector with indices of principle components to be included}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform an \code{eof} object (output of the \code{EOF} function) into a \code{station} object.
The function \code{as.station.EOF} represents the various principle components of the EOF as stations.
}
\seealso{
as.station EOF
}
