% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.field}
\alias{as.station.field}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{field}(x, ..., is = NULL, verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{field}}

\item{...}{other arguments}

\item{is}{a list or data.frame providing space index, e.g. station record or a lon(gitude) and lat(itude) range}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object where every grid box is represented as one station
}
\description{
Transform a \code{field} object into a \code{station} object.
}
\seealso{
as.station
}
