% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.list}
\alias{as.station.list}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{list}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{list} containing downscaled results e.g. for various months or seasons}

\item{...}{other arguments}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform a \code{list} of downscaled results into a \code{station} object. 
Note that this function will not work on lists containing other types of data.
}
\seealso{
as.station as.station.ds DS
}
