% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station}
\alias{as.station}
\alias{as.station.zoo}
\alias{as.station.data.frame}
\alias{as.station.ds}
\alias{as.station.pca}
\alias{as.station.list}
\alias{as.station.spell}
\alias{as.station.events}
\alias{as.station.dsensemble.station}
\alias{as.station.eof}
\alias{as.station.field}
\title{Coerce input to a \code{station} object}
\usage{
as.station(x, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}

\item{loc}{location(s), e.g, "Manchester" or c("Oslo","Bergen")}

\item{param}{short name of variable}

\item{unit}{unit of variable, e.g., 't2m'}

\item{lon}{longitude(s), a numerical or numerical vector}

\item{lat}{latitudes(s), a numerical or numerical vector}

\item{alt}{altitude(s), a numerical or numerical vector}

\item{cntr}{country, a character string or vector of character strings}

\item{longname}{long name of variable, e.g, 'temperature at 2m'}

\item{calendar}{calendar type}

\item{stid}{station id, a numerical or numerical vector}

\item{quality}{quality flag}

\item{src}{source of data}

\item{url}{url to website where data can be downloaded}

\item{reference}{reference describing data set}

\item{info}{additional information}

\item{method}{method applied to data}

\item{type}{type of data}

\item{aspect}{aspect describing data, e.g., 'original', 'anomaly', 'climatology'}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform an input object into the esd class \code{station}. 
\code{as.station} is an S3 method and will redirect to a fitting function depending on the type of input data.
}
\details{
\code{as.station.zoo} and \code{as.station.data.frame} adds attributes and changes the class to transform the input to a 'station' object.

\code{as.station.field} returns an object where every grid box is represented as one station.

\code{as.station.pca} transforms a 'pca' object (see \code{\link{PCA}}) to a 'station' object using the method \code{\link{pca2station}}.

\code{as.station.eof} represents the principle components of an 'eof' object as different stations.

\code{as.station.dsensemble} transform a \code{dsensemble} object to a \code{station} object in one of two ways:
i) If the input is of class \code{dsensemble} \code{pca}, you are redirected to \code{as.station.dsensemble.pca}
which calculates the downscaled ensemble for each station based on the downscaled ensemble of principle components,
returning a \code{dsensemble} \code{station} or \code{dsensemble} \code{list} object.
ii) If the input is a \code{dsensemble} \code{station} or \code{dsensemble} \code{list} object, you are redirected to \code{as.station.dsensemble.station}
which returns a \code{station} object holding the ensemble mean (or another statistical characteristic of the ensemble, see input argument \code{FUN})
of the downscaled results for each station.

\code{as.station.events} and \code{as.station.trajectory} aggregate an 'event' or 'trajectory' object to a 'station' object by
aggregating some aspect of the cyclones/anti-cyclones (or other type of event). By default, the total number of events/trajectories per month is calculated 
but the method can also estimate some other characteristic, e.g., the monthly mean sea level pressure at the center of the cyclones ().
}
\examples{
# How to generate a new 'station' object
data <- round(matrix(rnorm(20*12),20,12),2)
colnames(data) <- month.abb
x <- data.frame(year=1981:2000,data)
X <- as.station(x,loc="",param="noise",unit="none")

# Transform a field object into a station object
slp.field <- slp.DNMI(lon=c(-20,20), lat=c(50,70)) # get example SLP data
slp.station <- as.station(slp.field) # coerce SLP field to a station object
cb <- list(pal="burd", breaks=seq(1000,1020,2)) # specify color bar for maps
map(slp.field, FUN="mean", colbar=cb) # show map of SLP field
map(slp.station, FUN="mean", colbar=cb) # show map of SLP as station data

}
\seealso{
as.station.dsensemble as.station.dsensemble.pca as.station.dsensemble.station
}
