% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.spell}
\alias{as.station.spell}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{spell}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{spell}}

\item{...}{other arguments}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform a \code{spell} object (containing information about e.g. warm or cold spells) into a \code{station} object.
}
\seealso{
as.station
}
