% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.zoo}
\alias{as.station.zoo}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{zoo}(x, ..., loc = NA, param = NA, unit = NA,
  lon = NA, lat = NA, alt = NA, cntr = NA, longname = NA,
  calendar = NA, stid = NA, quality = NA, src = NA, url = NA,
  reference = NA, info = NA, method = NA, type = NA, aspect = NA,
  verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{zoo} typically containing data from one or several measurement stations}

\item{loc}{location(s), e.g, "Manchester" or c("Oslo","Bergen")}

\item{param}{short name of variable}

\item{unit}{unit of variable, e.g., 't2m'}

\item{lon}{longitude(s), a numerical or numerical vector}

\item{lat}{latitudes(s), a numerical or numerical vector}

\item{alt}{altitude(s), a numerical or numerical vector}

\item{cntr}{country, a character string or vector of character strings}

\item{longname}{long name of variable, e.g, 'temperature at 2m'}

\item{calendar}{calendar type}

\item{stid}{station id, a numerical or numerical vector}

\item{quality}{quality flag}

\item{src}{source of data}

\item{url}{url to website where data can be downloaded}

\item{reference}{reference describing data set}

\item{info}{additional information}

\item{method}{method applied to data}

\item{type}{type of data}

\item{aspect}{aspect describing data, e.g., 'original', 'anomaly', 'climatology'}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object
}
\description{
Transform a \code{zoo} object into a \code{station} object
}
\seealso{
as.station
}
