% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trajectory.R
\name{as.trajectory}
\alias{as.trajectory}
\alias{as.trajectory.default}
\alias{as.trajectory.events}
\alias{events2trajectory}
\alias{trajectory}
\title{Transform an input object into a \code{trajectory} object}
\usage{
as.trajectory(x, ...)
}
\arguments{
\item{x}{input object}

\item{\dots}{additional arguments}
}
\description{
Transform an input object into a \code{trajectory} object
}
