% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balls.R
\name{balls}
\alias{balls}
\title{Extention of points}
\usage{
balls(x, y = NULL, col = NULL, cex.max = 2, n = 20)
}
\arguments{
\item{x}{input vector}

\item{y}{input vector of same length as x}

\item{col}{vector of colors}

\item{cex.max}{maximum size of balls}

\item{n}{length of color scale}
}
\description{
Extention of points
}
