% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.station.R
\name{barplot.station}
\alias{barplot.station}
\title{create a barplot}
\usage{
\method{barplot}{station}(x, threshold = 0, ...)
}
\arguments{
\item{x}{a 'station' object}

\item{threshold}{threshold - midline of plot}
}
\description{
create a barplot
}
