% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasfix.R
\name{biasfix}
\alias{biasfix}
\title{Bias correct}
\usage{
biasfix(x, verbose = FALSE)
}
\description{
Bias correction as described in  Imbert & Benestad (2005), Theor. Appl. Clim., DOI: 10.1007/s00704-005-0133-4
}
