% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{C.C.eq}
\alias{C.C.eq}
\title{Various formulas, equations and transforms.}
\usage{
C.C.eq(x)
}
\arguments{
\item{x}{an object containing air temperature data}
}
\value{
the saturation vapor pressure
}
\description{
\code{C.C.eq}: Clapeyron-Clausius equation (saturation evaporation pressure)
where \code{x} is a data object holding the temperature.
}
\examples{

t2m <- t2m.DNMI(lon=c(-70,-10),lat=c(20,60))
es <- C.C.eq(t2m)
map(es)

}
\references{
'An Introduction to Atmospheric Physics' by RG. Fleagle, JA. Businger Academic Press, 9. jan. 1981, eq. 2.89, p. 72.
}
\seealso{
precip.vul t2m.vul precip.rv precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
