% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cbind.field}
\alias{cbind.field}
\title{Extension of rbind for field objects}
\usage{
\method{cbind}{field}(..., verbose = FALSE)
}
\arguments{
\item{\dots}{input arguments}

\item{verbose}{if TRUE print information on progress}
}
\description{
Extension of rbind for field objects
}
