% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{CCA}
\alias{CCA}
\alias{CCA.default}
\alias{CCA.eof}
\alias{CCA.pca}
\alias{CCA.field}
\alias{predict.cca}
\title{Canonical correlation analysis}
\usage{
CCA(Y, X, ...)
}
\arguments{
\item{Y}{An object with climate data: field, eof, pca.}

\item{X}{Same as Y.}

\item{...}{Other arguments.}

\item{ip}{Which EOFs to include in the CCA.}

\item{newdata}{The same as X.}

\item{verbose}{If TRUE print information about progress.}
}
\value{
A CCA object: a list containing a.m, b.m, u.k, v.k, and r,
describing the Canonical Correlation variates, patterns and correlations.
a.m and b.m are the patterns and u.k and v.k the vectors (time evolution).
}
\description{
Applies a canonical correlation analysis (CCA) to two data sets. The CCA
here can be carried out based on an \code{\link{svd}} based approach (after
Bretherton et al. (1992), J. Clim. Vol 5, p. 541, also documented in
Benestad (1998): "Evaluation of Seasonal Forecast Potential for Norwegian
Land Temperatures and Precipitation using CCA", DNMI KLIMA Report 23/98 at
\url{http://met.no/english/r_and_d_activities/publications/1998.html}) or
ii) a covariance-eigenvalue approach (after Wilks, 1995, "Statistical
methods in the Atmospheric Sciences", Academic Press, p. 401).
}
\details{
The analysis can also be applied to either EOFs or fields.
}
\examples{

# CCA with two eofs
slp <- slp.NCEP(lat=c(-40,40),anomaly=TRUE)
sst <- sst.NCEP(lat=c(-40,40),anomaly=TRUE)
eof.1 <- EOF(slp, it='Jan')
eof.2 <- EOF(sst, it='Jan')
cca <- CCA(eof.1,eof.2)
plot(cca)

# CCA with PCA and EOF:
\dontrun{
NACD <- station.nacd()
plot(annual(NACD))
map(NACD,FUN="sd")
pca <- PCA(NACD)
plot(pca)
naslp <- slp.NCEP(lon=c(-30,40),lat=c(30,70),anomaly=TRUE)
map(naslp)
eof <- EOF(naslp,it='Jan')
nacca <- CCA(pca,eof)
plot(nacca)
cca.pre <- precit.cca(nacca)
}

}
