% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.bad.dates.R
\name{check.bad.dates}
\alias{check.bad.dates}
\title{Check dates for impossibilities}
\usage{
check.bad.dates(yyyy, mm, dd)
}
\arguments{
\item{yyyy}{year(s)}

\item{mm}{month(s)}

\item{dd}{day(s)}
}
\description{
Check dates for impossibilities
}
