% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.station.R
\name{clean.station}
\alias{clean.station}
\title{Remove stations with a lot of missing data}
\usage{
clean.station(x, miss = 0.1, verbose = TRUE)
}
\arguments{
\item{x}{a \code{station} object}

\item{miss}{fraction of data that are allowed to be missing,
e.g., miss=.1 means that stations with more than 10\% missing data
will be excluded}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Remove stations with a lot of missing data
}
