% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim2pca.R
\name{clim2pca}
\alias{clim2pca}
\alias{clim2pca.default}
\alias{clim2pca.month}
\alias{clim2pca.day}
\title{PCA analysis of the seasonal cycle}
\usage{
clim2pca(x, verbose = FALSE, ...)
}
\description{
Express station data as \code{\link{PCA}} where each of the EOFs
(attribute 'pattern' of output object)
represent one year and the PCs
(main part of ouput object) describe the seasonal variations.
}
