% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip.R
\name{cmip3.model_id}
\alias{cmip3.model_id}
\alias{cmip5.model_id}
\title{Clean up CMIP3 and CMIP5 meta data}
\usage{
cmip3.model_id(txt)
}
\arguments{
\item{txt}{meta data}
}
\description{
Clean up CMIP3 and CMIP5 meta data
}
