% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale.R
\name{col.bar}
\alias{col.bar}
\alias{colbar}
\title{Display a color bar object on an existing plot.}
\usage{
col.bar(
  xleft,
  ybottom,
  xright,
  ytop,
  breaks,
  horiz = TRUE,
  srt = 45,
  pch = 15,
  v = 1,
  h = 1,
  col = col,
  cex = 5,
  cex.lab = 0.6,
  cex.axis = 0.9,
  type = "r",
  verbose = FALSE,
  vl = 0.5,
  border = "black",
  ...
)
}
\arguments{
\item{breaks}{A numeric vector of breakpoints for the colours}

\item{horiz}{a boolean; if TRUE add horizontal color bar, else add vertical color bar}

\item{srt}{Angle of tick labels}

\item{pch}{see \code{\link{par}}}

\item{v}{Vertical space between color bar points}

\item{h}{horizontal space between color bar points}

\item{col}{see \code{\link{par}}}

\item{cex}{A numerical value giving the amount by which plotting text
and symbols should be magnified relative to the default (see \code{\link{par}})}

\item{cex.lab}{Magnification factor for x and y labels (see \code{\link{par}})}

\item{cex.axis}{Magnification factor for axis annotations (see \code{\link{par}})}

\item{type}{r : rectangular shape , p : for points}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{vl}{a numerical specifying the relative placement of the vertical lines}

\item{border}{a boolean; if TRUE show color bar borders}

\item{\dots}{Additional graphical parameters to be passed on}
}
\description{
Add a color bar or color points into an exisiting plot or map.
For a description of palette choices, see \code{\link{colscal}}.
}
\seealso{
colbar.ini colscal
}
