% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale.R
\name{colbar.ini}
\alias{colbar.ini}
\title{Display a color bar object on an existing plot.}
\usage{
colbar.ini(x, FUN = NULL, colbar = NULL, verbose = FALSE)
}
\arguments{
\item{x}{an input object, e.g., a 'zoo', 'station' or 'field' object or numerical vector}

\item{FUN}{a function}

\item{colbar}{a list}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{a}{list: colbar = list(col, breaks, n, type, cex, h, v, pos, show, rev)
where 
\code{col} is a vector containing the colors corresponding to the values 
specified in the numerical vector \code{breaks},
\code{n} is the number of breaks (used only if breaks are not specified),
\code{show} if TRUE show color bar,
\code{rev} if TRUE reverse color scale,
\code{cex} see \code{\link{par}},
\code{h}, \code{v}, \code{type}: see \code{\link{col.bar}}
\code{pos} not in use?}
}
\description{
Generate a color bar list and add information about the breaks of the color scale based on the numerical range of the input data.
}
\seealso{
col.bar colbar
}
