% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale.R
\name{colscal}
\alias{colscal}
\title{Generate a color scale}
\usage{
colscal(n = 14, pal = "t2m", rev = FALSE, alpha = NULL,
  test = FALSE, verbose = FALSE)
}
\arguments{
\item{n}{length of color vector}

\item{pal}{color palette: "bwr",rwb","faint.bwr","faint.rwb","rainbow","gray.colors","heat.colors","terrain.colors","topo.colors","cm.colors","grmg","brbu","budor","budrd","bugr","bugy","buor","buorr","bu","rd","cat","cold", or "warm"}

\item{rev}{a boolean; if TRUE reverse color scale}

\item{alpha}{factor defining transparency of color}

\item{test}{a boolean; if TRUE show a sample of the color scale}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Generate a vector of colors
}
\details{
Palette options are as follows:
't2m': blue-yellow-red (from seNorge), 
'precip', 'mu' and 'fw': white-blue (from seNorge),
'cold': a cold color scale (from seNorge),
'warm': a warm color scale (from seNorge),
'bwr' (blue-white-red),
'slp' and 'mslp' (same as 'bwr'),
'rwb' (red-white-blue), 
'faint.rwb': fainter version of 'rwb',
'faint.bwr': fainter version of 'bwr', 
'grmg': green-magenta, 
'brbu': brown-blue,
'budor': blue-orange,
'budrd': blue-red,
'bugr': blue-green,
'bugy': blue-gray,
'buor': blue-orange (brighter colors than 'budor'),
'buorr': blue-green (brighter and more yellow than 'buor' and 'budor'),
'bu': blues,
'rd': reds,
'cat': categorical color scale,
 color scales from grDevices: 'rainbow', 'gray.colors', 'heat.colors', 'terrain.colors',
'topo.colors', and 'cm.colors'.
}
\seealso{
colbar col.bar colbar.ini
}
