% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\alias{combine.default}
\alias{combine.station}
\alias{combine.stations}
\alias{combine.zoo}
\alias{combine.ds}
\alias{combine.ds.comb}
\alias{combine.ds.station}
\alias{combine.ds.station.eof}
\alias{combine.ds.station.field}
\alias{combine.ds.pca}
\alias{combine.station.eof}
\alias{combine.field}
\alias{combine.filed.station}
\alias{combine.station.field}
\alias{sp2np}
\alias{combine.trajectory}
\alias{combine.events}
\alias{combine.field.station}
\alias{combine.list}
\title{Combine}
\usage{
combine(...)
}
\arguments{
\item{x}{station, eof, or field object}

\item{all}{See \code{link{merge.zoo}}}

\item{orig.format}{TRUE: the result will the formatted the same way as the
input.}

\item{dimension}{Which dimension to combine - in time or in space}

\item{approach}{How to combine}

\item{greenwich}{TRUE: center map on the Greenwich line (0E)}

\item{SP2NP}{TRUE: order from south pole (bottom of plot) to north pole (top
of plot)}

\item{ignore}{List of attributes to ignore.}
}
\value{
A field object
}
\description{
\code{combine} is a S3 method for combining esd objects, e.g. into groups of
stations, stations and eof object, or fields. The function is based on
\code{\link[zoo]{merge.zoo}}, and is also used to synchronise the esd objects.
}
\details{
For fields, \code{combine.field} is used to append different data sets, e.g.
for the purpose of computing common EOFs (seeo \code{\link{EOF}} or for
mixing fields (coupled EOFs).

For stations, \code{combine.station} can work tow ways: (1) to combine a set
of stations and group them into one data object; (2) combine series with
different monthly values for one specific site into one record for the
monthly data. E.g. January, February, ..., December months can be combined
into one complete series of monthly data.

For DS-results, \code{combine.ds} is based on \code{combine.station}, but
also takes care of the additional meta data (the original series and
predictor patterns). For instance, this method can combine seperate
downscaled results for each calendar months at a single location into one
complete time series.

\code{g2dl} transform objects between grid starting at the grenwich
(\code{greenwich=TRUE}) and the data line (\code{greenwich=FALSE}).

\code{sp2np} re-arranges field objects accroding to a grid going from 90S
(South Pole) to 90N (Noth Pole) for \code{SP2NP=TRUE}. Otherwise, the object
is arranged from 90N to 90S.

Other operations, such as \code{c(...)}, \code{rbind(...)} (combine along
the time dimension), and \code{cbind(...)} (combine along the space
dimension) also work.
}
\examples{
T2m_DNMI <- t2m.DNMI(lon=c(-40,40),lat=c(30,70))
T2m_NorESM <- t2m.NorESM.M(lon=c(-40,40),lat=c(30,70))

# Combine in time to compute common EOFs:
X <- combine(T2m_DNMI,T2m_NorESM)
ceof <- EOF(X,it="Jan")
plot(ceof)

# Use combine to synchronise field and station data:
data("Oslo")
y <- combine(Oslo,T2m_DNMI)
plot(y$y)

}
\keyword{utilities}
