% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ncfiles.R
\name{combine_nc_stid}
\alias{combine_nc_stid}
\title{Combine or append netCDF files for stations into one. 
Rasmus.Benestad@met.no & chatGPT, , 2025-10-09}
\usage{
combine_nc_stid(
  input_files,
  output_file = "combined.nc",
  stid_dim = "stid",
  time_dim = "time",
  verbose = FALSE
)
}
\arguments{
\item{input_files}{a string of names of netCDF station files to combine}

\item{output_file}{netCDF file with the combined station data}

\item{stid_dim}{identifier for station IDs}

\item{verbose}{used for debugging and diagnostics of executing the code}

\item{time_dimidentifier}{for time dimension}
}
\description{
Combine or append netCDF files for stations into one. 
Rasmus.Benestad@met.no & chatGPT, , 2025-10-09
}
\examples{
\dontrun{
Files <- list.files("~/Downloads/ecad2nc4/", pattern = "\\\\.nc$", full.names = TRUE)

params <- c('t2m','tmax','tmin','precip','sd')
for (param in params) { 
  files <- Files[grep(param,Files)]
  print(files)
  output <- paste0(param,".ecad_blend.nc")
  if (file.exists(output)) file.remove(output)
  combine_nc_stid(files, output)
}
}

}
\seealso{
write2ncdf4.station as.station retrieve.station
}
\keyword{utilities}
