% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatESGF.R
\name{concat.ESGF}
\alias{concat.ESGF}
\alias{}
\title{concat.ESGF
Concatenate netcdf files from ESGF into a combined netcdf file of historical + projected 
climate simulations using CDO command line}
\usage{
concat.ESGF(
  meta,
  param = "tas",
  path = "/lustre/storeB/project/CMIP/CMIP6.monthly/from_synda/CMIP6",
  expid = "ssp585",
  use = "CDO",
  verbose = FALSE
)
}
\arguments{
\item{meta}{Output data.frame from meta.ESGF() function. 
The full path to the netCDF file can either be included in 'ncfile' or entered as a separate input ('path').}

\item{param}{Climate parameter as defined in ESGF.}

\item{path}{Path to netcdf files}
}
\value{
NetCDF file output containing both historical and future climate simulations.
}
\description{
This function assumes that the data has been downloaded locally using the functions \code{meta.ESGF} and \code{retrieve.ESGF}.
}
\examples{

\dontrun{
}



}
\seealso{
meta.ESGF retrieve.ESGF
}
