% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ncfiles.R
\name{concat_nc_along_stid}
\alias{concat_nc_along_stid}
\title{Combine or append netCDF files for stations into one. 
Rasmus.Benestad@met.no & chatGPT, , 2025-10-09}
\usage{
concat_nc_along_stid(input_files, output_file, verbose = TRUE)
}
\arguments{
\item{input_files}{a string of names of netCDF station files to combine}

\item{output_file}{netCDF file with the combined station data}

\item{verbose}{used for debugging and diagnostics of executing the code}
}
\description{
Combine or append netCDF files for stations into one. 
Rasmus.Benestad@met.no & chatGPT, , 2025-10-09
}
\examples{
\dontrun{
concat_nc_along_stid(ncfiles, paste0('~/Downloads/',variable,'.ecad_blend.nc'),verbose=TRUE)
}

}
\seealso{
write2ncdf4.station as.station retrieve.station
}
\keyword{utilities}
