% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfield.R
\name{corfield}
\alias{corfield}
\alias{corfield.default}
\alias{corfield.zoo}
\alias{corfield.field}
\alias{corfield.field.station}
\alias{corfield.station}
\alias{corfield.eof}
\alias{corfield.trajectory}
\title{Correlation}
\usage{
corfield(x, y, ...)
}
\arguments{
\item{x}{data object}

\item{y}{data object}

\item{plot}{TRUE: plot the results}

\item{use}{see \code{\link{cor}}}

\item{new}{see \code{link{map}}}

\item{ip}{index EOF pattern}
}
\value{
Map of correlation
}
\description{
Compute the correlation between field objects and station/field.
}
\examples{
x <- t2m.DNMI(lon=c(-40,30),lat=c(0,50))
y <- t2m.NorESM.M(lon=c(-40,30),lat=c(0,50))
r <- corfield(annual(x),annual(y))

data(Oslo)
t2m <- t2m.DNMI()
x <- subset(Oslo,it='january')
y <- subset(t2m,it='january')
r <- corfield(x,y)

}
\keyword{manip}
