% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.events.R
\name{count.events}
\alias{count.events}
\title{Count the number of events per month}
\usage{
count.events(x, by.trajectory = TRUE, FUN = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{input object of class 'events'}

\item{by.trajectory}{if TRUE count every trajectory once, otherwise count every time step separately}

\item{verbose}{if TRUE print progress}

\item{\dots}{additional arguments}
}
\description{
Count the number of events per month
}
