% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{count}
\alias{count}
\title{Count statistics}
\usage{
count(x, threshold = 1, fraction = FALSE, ...)
}
\arguments{
\item{x}{station or field object}

\item{threshold}{threshold value}

\item{fraction}{TRUE: divide the number of counts by number of samples}

\item{upper}{upper limit for maximum length - ignore any above this because they are likely erronoeus}

\item{higher}{argument of count, if TRUE: count values above threshold, FALSE: count values below threshold}

\item{FUN}{function}
}
\value{
Station or field objects
}
\description{
Returns a count of cases with values over a threshold value.
}
\examples{

# number of warm days:
data(ferder)
plot(as.seasons(ferder,FUN='count',threshold=20), new=FALSE)

# Mild winter days - number of days in the winter season with
# above freezing temperatures
try(coldwinterdays(ferder))

# Quantile-quantile plot
qqgeom(ferder, treshold=1, pois=TRUE)

}
\seealso{
spell exceedance nevents wetfreq hotsummerdays coldwinterdays nwetdays plot
}
\keyword{utilities}
