% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumugram.R
\name{cumugram}
\alias{cumugram}
\title{InfoGraphics}
\usage{
cumugram(
  x,
  it = NULL,
  start = "-01-01",
  prog = FALSE,
  plot = TRUE,
  verbose = FALSE,
  FUN = "mean",
  main = NULL,
  ...
)
}
\arguments{
\item{x}{an input object of class 'station'}

\item{it}{A list or data.frame providing time index, e.g. month}

\item{start}{year and month, e.g., '-01-01' to start in january}

\item{prog}{a boolean; if TRUE show prognosis for end of year in cumugram}

\item{plot}{a boolean; if TRUE show the plot}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{FUN}{a function}

\item{main}{main title}

\item{\dots}{additional arguments}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{cumugram} shows the running cumulative mean or sum of a time series
}
\examples{
data(bjornholt)
cumugram(bjornholt)

}
\seealso{
wheel graph visprob conf vis diagram scatter plot map
}
