% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{datafrequency}
\alias{datafrequency}
\title{Calculate the frequency}
\usage{
datafrequency(data = NULL, unit = NULL, verbose = FALSE)
}
\arguments{
\item{data}{input object containing a time index}

\item{unit}{unit of time index in 'data'}
}
\description{
Calculate the frequency
}
