% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{day2IDF}
\alias{day2IDF}
\alias{as.residual}
\alias{as.residual.ds}
\alias{as.residual.station}
\alias{as.residual.field}
\title{Calculate approximate intensity-duration-frequency (IDF) curves from daily data through
a crude downscaling in timescales. 
Caluculate the residual of a 'ds' object, i.e., the original data minus the fitted values.}
\usage{
day2IDF(
  x,
  L = c(1, 2, 3, 6, 12, 24),
  tau = 10,
  zeta = NULL,
  n0 = 365.25,
  alpha = c(1.256, 0.064),
  verbose = FALSE
)
}
\arguments{
\item{x}{a daily station object for rainfall}

\item{L}{timescale in hours}

\item{tau}{return interval in years}

\item{n0}{number of observations per year}

\item{alpha}{best fit constant and slope for scaling factor between exponential distribution and empirical 
distribution taken from https://doi.org/10.1088/1748-9326/ab2bb2}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{exponent}{for the power-law dependency between return values and timescales}
}
\description{
Calculate approximate intensity-duration-frequency (IDF) curves from daily data through
a crude downscaling in timescales. 
Caluculate the residual of a 'ds' object, i.e., the original data minus the fitted values.
}
\examples{
y <- station(stid=18700,src='metnod.thredds',param='precip')
z <- IDF(y)
View(z)

}
