% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density2count.R
\name{density2count}
\alias{density2count}
\title{Estimate the storm count based on cyclone density}
\usage{
density2count(y, it = NULL, is = NULL, verbose = TRUE)
}
\description{
Estimate the storm count based on cyclone density
}
