% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Perform EOT-Based Spatial Downscaling}
\usage{
downscale(pred, resp, neot = 1L, newdata, nprd = 1L, var = NULL,
  filename = "", format = "GTiff", overwrite = TRUE, ...)
}
\arguments{
\item{pred, resp}{\code{Raster*} objects used as predictor and response 
domains during EOT analysis.}

\item{neot}{\code{integer}. Number of EOT modes to calculate.}

\item{newdata}{\code{Raster*} object passed to \code{\link[remote]{predict}}.}

\item{nprd}{\code{integer}. Number of EOT modes used for prediction.}

\item{var}{\code{numeric}. Minimum amount of variance to be explained by the 
EOT modes. If specified, 'nprd' is adjusted automatically.}

\item{filename, format, overwrite}{Arguments passed to 
\code{\link[remote]{predict}} and its call to \code{\link{writeRaster}}.}

\item{...}{Additional arguments passed to \code{\link{eot}}.}
}
\value{
A \code{Raster*} object.
}
\description{
Perform EOT-based spatial downscaling based on user-defined predictor and 
response domains.
}
\examples{
## raw data
plot(albGIMMS[[49:52]])

## downscaled data
prd <- downscale(pred = albGIMMS[[1:48]], resp = albMODIS[[1:48]], neot = 10L,
                 newdata = albGIMMS[[49:96]], var = 0.85)
plot(prd[[1:4]])                  

}
\references{
Appelhans T, Detsch F, Nauss T (2015) \strong{remote}: Empirical Orthogonal 
Teleconnections in R. \emph{Journal of Statistical Software} 65(10), 1-19, 
doi:10.18637/jss.v065.i10. Available online: 
\url{https://www.jstatsoft.org/article/view/v065i10}.
}
\author{
Florian Detsch
}
