% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ele2param}
\alias{ele2param}
\alias{esd2ele}
\alias{param2ele}
\title{Dictionary and conversion tools between esd element identifier and variables
names and specifications.}
\usage{
ele2param(ele = NULL, src = NULL)
}
\arguments{
\item{ele}{element identifier}

\item{src}{a character string for the acronym of the data source}

\item{param}{parameter identifier}
}
\value{
A meta data matrix object with the glossary of the different
variables or element identifiers as originally defined by each data source
}
\description{
Converts between esd element/parameter identifier and variable names from
different data sources.
}
\examples{
# Display the glossary of paramerters or element identifiers for 'GHCND' data source.
print(ele2param(ele=NULL,src='GHCND'))
# Display the glossary of parameters or element identifiers for all data sources. 
print(ele2param())
# Convert mean temperature parameter (param) to esd element (ele).
ele <- esd2ele(param='t2m')
print(ele)


}
\keyword{parameter,element}
