% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.events.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate trajectories}
\usage{
enumerate(x, param = "trajectory", verbose = FALSE)
}
\arguments{
\item{x}{an \code{events} object}

\item{param}{Name of the column holding the trajectory numbers (default: "trajectory")}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
an \code{events} object
}
\description{
The function enumerates the trajectories ("trajectory") of an event object 
for which the events have been tracked using the function \code{track}. 
\code{enumerate} is called from \code{trackstats} and \code{as.trajectory}
and can also be applied after selecting cyclones using \code{subset} to give 
the trajectories new numbers.
}
