% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF.mvcomb.R
\name{EOF.mvcomb}
\alias{EOF.mvcomb}
\title{Multivariate Empirical Orthogonal Functions (EOFs).}
\usage{
\method{EOF}{mvcomb}(
  X,
  Y = NULL,
  it = NULL,
  is = NULL,
  n = 20,
  lon = NULL,
  lat = NULL,
  geoweighting = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{An \code{mvcomb} object or \code{field} object}

\item{Y}{A field object. If two \code{field} objects X and Y are provided, \code{mvcombine} is applied to combine the two.}

\item{it}{A list or data.frame providing time index, e.g. a range of years like c(1979,2010), a season ('djf'), or a month ('dec' or 'december').}

\item{is}{A list or data.frame providing space index, e.g. a list of longitude and latitude range like list(lon=c(0,60), lat=c(35,60)).}

\item{n}{Number of EOFs.}

\item{lon}{set longitude range - see \code{\link{t2m.NCEP}}}

\item{lat}{set latitude range.}

\item{verbose}{If TRUE print diagnostics}

\item{\dots}{Additional arguments}
}
\description{
Computes EOFs (a type of principal component analysis) for an \code{mvcomb} object which is
the combination of two fields of different variables and output of the \code{mvcombine} function.
}
\seealso{
EOF
}
