% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF.R
\name{EOF}
\alias{EOF}
\alias{EOF.default}
\alias{EOF.field}
\alias{EOF.comb}
\alias{eof2field}
\alias{PCA}
\alias{PCA.default}
\alias{PCA.station}
\alias{PCA.matrix}
\alias{pca2station}
\title{Empirical Orthogonal Functions (EOFs).}
\usage{
EOF(
  X,
  ...,
  it = NULL,
  is = NULL,
  n = 20,
  lon = NULL,
  lat = NULL,
  verbose = FALSE,
  anomaly = TRUE
)
}
\arguments{
\item{X}{a 'field' or 'pca' object}

\item{\dots}{additional arguments}

\item{it}{see \code{\link{subset}}}

\item{is}{Spatial subsetting - see \code{\link{subset.eof}}}

\item{n}{number of EOFs}

\item{lon}{set longitude range - see \code{\link{t2m.NCEP}}}

\item{lat}{set latitude range}

\item{verbose}{TRUE - clutter the screen with messages}

\item{anomaly}{When TRUE, subtract the mean before SVD.}
}
\value{
File containing an 'eof' object which is based on the 'zoo' class.
}
\description{
Computes EOFs (a type of principal component analysis) for combinations of
data sets, typically from gridded data, reanalysis and climate models
results.
}
\details{
[ref: Benestad (2001), "A comparison between two empirical downscaling
strategies", \emph{Int. J. Climatology}, \bold{vol 21}, Issue 13,
pp.1645-1668. DOI 10.1002/joc.703]. and \code{mixFields} prepares for
mixed-field EOF analysis [ref. Bretherton et al. (1992) "An Intercomparison
of Methods for finding Coupled Patterns in Climate Data", \emph{J. Climate},
\bold{vol 5}, 541-560; Benestad et al. (2002), "Empirically downscaled
temperature scenarios for Svalbard", \emph{Atm. Sci. Lett.},
doi.10.1006/asle.2002.0051].

Uncertainty estimates are computed according to North et al. (1982),
"Sampling Errors in the Estimation of Empirical Orthogonal Functions",
\emph{Mon. Weather Rev.}, \bold{vol 110}, 699-706.

The EOFs are based on \code{\link{svd}}.

See the course notes from Environmental statistics for climate researchers
\url{http://www.gfi.uib.no/~nilsg/kurs/notes/course.html} for a discussion
on EOF analysis.

The method \code{PCA} is similar to EOF, but designed for parallel station
series (e.g. grouped together with \code{\link{merge}}). PCA does not assume
gridded values and hence does not weigth according to grid area. PCA is
useful for downscaling where the spatial covariance/coherence is important,
e.u involving different variables from same site, same variable from
different sites, or a mix between these. For instance, PCA can be applied to
the two wind components from a specific site and hence extract the most
important wind directions/speeds. \code{PCA.matrix} is just a wrapper function 
for \code{svd} that makes sure that the dimensions of the input are in order.
}
\examples{

# Simple EOF for annual mean SST:
sst <- sst.NCEP(lon=c(-90,20),lat=c(0,70))
SST <- aggregate(sst, year, mean, na.rm = FALSE)
eof.sst <- EOF(SST)
plot(eof.sst)

# EOF of July SST:
eof.sst7 <- EOF(sst,it="Jul")
plot(eof.sst7)

# common EOF for model
# Get some sample data, extract regions:
GCM <- t2m.NorESM.M()
gcm <- subset(GCM,is=list(lon=c(-50,60),lat=c(30,70)))
t2m.dnmi <- t2m.DNMI()
dnmi <- subset(t2m.dnmi,is=list(lon=c(-50,60),lat=c(30,70)))

OBS <-  aggregate(dnmi, by=year, mean, na.rm = FALSE)
GCM <- aggregate(gcm, by=year, mean, na.rm = FALSE)
OBSGCM <- combine(OBS,GCM,dimension='time')

ceof <- EOF(OBSGCM)
plot(ceof)

# Example for using PCA in downscaling
## nacd <- station(src='nacd')
## X <- annual(nacd)
X <- station(src='nacd')
nv <- function(x) sum(is.finite(x))
ok <- (1:dim(X)[2])[apply(X,2,nv) == dim(X)[1]]
X <- subset(X,is=ok)
pca <- PCA(X)
map(pca)

slp <- slp.NCEP(lon=c(-20,30),lat=c(30,70))
eof <- EOF(slp,it="Jan")
ds <- DS(pca,eof)
# ds is a PCA-object
plot(ds)

# Recover the station data:
Z <- pca2station(pca)
plot(Z,plot.type='multiple')

}
\seealso{
as.eof
}
\author{
R.E. Benestad
}
\keyword{multivariate}
\keyword{spatial}
\keyword{ts}
