% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESD-package.R
\docType{package}
\name{ESD-package}
\alias{ESD-package}
\alias{ESDpackage}
\title{EOT-Based Spatial Downscaling}
\description{
EOT-Based Spatial Downscaling
}
\details{
We provide a set of functions to spatially downscale arbitrary \code{Raster*}
time series based on the empirical orthogonal teleconnections (EOT) method.
}
\author{
Florian Detsch, Thomas Nauss
}
\keyword{package}

