% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCycle.R
\name{evaluateCycle}
\alias{evaluateCycle}
\title{Evaluation Cycle of EOT-Based NDVI Resampling}
\usage{
evaluateCycle(pred, resp, training, n = 10L, var = NULL,
  deseasoning = FALSE, cores = 1L, ...)
}
\arguments{
\item{pred}{\code{Raster*} object. Predictor data for EOT analysis.}

\item{resp}{\code{Raster*} object. Response data for EOT analysis.}

\item{training}{\code{integer}. Indices of training layers in 'pred' and
'resp'. All non-included layers are automatically used for model testing.}

\item{n}{\code{integer}. Number of EOT modes to calculate.}

\item{var}{\code{numeric}. Minimum amount of variance to be explained. If
\code{NULL} (default), all calculated EOT modes are used for
\code{\link[remote]{predict}}-ion.}

\item{deseasoning}{\code{logical}. If \code{TRUE}, \code{\link{deseason}}-ing
is applied prior to the actual EOT procedure.}

\item{cores}{\code{integer}. Number of cores for parallel processing.}

\item{...}{Additional arguments passed to \code{\link{eot}}.}
}
\value{
A \code{list} of length 2. The content of the first (second) slot is a
\code{data.frame} (\code{RasterLayer}) with spatial (temporal) error and
regression metrics (RMSE).
}
\description{
Evaluate the performance of EOT-based spatial resampling based on selected
error and regression metrics, including mean error (ME), mean absolute error
(MAE), root-mean-square-error (RMSE) as well as correlation coefficient
(\eqn{r}) and coefficient of determination (\eqn{R^2}).
}
\author{
Florian Detsch
}
\references{
Detsch F, Otte I, Appelhans T, Hemp A, Nauss T (2016) Seasonal and long-term
vegetation dynamics from 1-km GIMMS-based NDVI time series at Mt.
Kilimanjaro, Tanzania. Remote Sensing of Environment 178, 70-83,
doi:10.1016/j.rse.2016.03.007.
}
\seealso{
\code{\link{deseason}}, \code{\link{eot}}, \code{\link{nXplain}}
}

