% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events2field.R
\name{events2field}
\alias{events2field}
\alias{as.field.events}
\alias{events2density}
\title{Transform an input object into a \code{field} object}
\usage{
events2field(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a \code{trajectory} object}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}

\item{dt}{frequency of output: 'month', 'season', 'quarter' (same as 'season') or 'year'}

\item{dx}{resolution in longitude direction (unit: degrees)}

\item{dy}{resolution in latitude direction (unit: degrees)}

\item{plot}{if TRUE show plot of results}

\item{radius}{radius within which to look for trajectories for each grid point (unit: m)}

\item{it}{a time index, e.g., a range of years: c(1984,2019)}

\item{is}{a spatial index, e.g., a list with longitude and latitude ranges: list(lon=c(0,45), lat=c(45,70))}

\item{param}{parameter to calculate field of; if NULL calculate track density (for other options, see \code{colnames(x)})}

\item{type}{'track', 'genesis', or 'lysis'}

\item{longname}{name of variable}
}
\value{
a \code{field} object
}
\description{
Transform a \code{trajectory} object into a \code{field} object by aggregating it in time and space.
}
\seealso{
as.field CCI track.events
}
