% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events2station.R
\name{events2station}
\alias{events2station}
\title{Transform an 'event' object into a 'station' object}
\usage{
events2station(
  x,
  param = "count",
  FUN = "mean",
  verbose = TRUE,
  longname = NULL,
  unit = NULL,
  ...
)
}
\arguments{
\item{x}{input object of class 'events'}

\item{param}{parameter to aggregate, e.g., 'count' or some characteristic such as 'pcent' or 'radius' (for options, see names(x))}

\item{FUN}{a function, e.g., 'mean' or 'sum'}

\item{verbose}{a boolean; If TRUE print information about progress}

\item{longname}{long name of variable}

\item{unit}{name of unit}

\item{\dots}{additional arguments}
}
\value{
a 'station' object
}
\description{
Aggregate some aspects of an 'events' object in time and space and transform the time series into a 'station' object.
}
