% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratories-package.R
\docType{package}
\name{exploratories-package}
\alias{exploratories-package}
\alias{exploratoriespackage}
\title{Vegetation Analysis in the Biodiversity Exploratories}
\description{
Vegetation Analysis in the Biodiversity Exploratories
}
\details{
We provide a set of functions to analyze long-term vegetation dynamics in the
Biodiversity Exploratories.
}
\author{
Florian Detsch, Thomas Nauss
}
\keyword{package}

