% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.class.R
\name{file.class}
\alias{file.class}
\title{Used to check the contents in netCDF file.}
\usage{
file.class(ncfile)
}
\arguments{
\item{ncfile}{filename of netcdf file}
}
\description{
Assumes that empty class attribute means a field object
}
\seealso{
check.ncdf4 retrieve
}
