% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\docType{data}
\name{frequency.abb}
\alias{frequency.abb}
\title{Abbreviated time frequency names}
\format{
An object of class \code{character} of length 7.
}
\usage{
frequency.abb
}
\description{
Abbreviated time frequency names
}
\keyword{datasets}
