% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\docType{data}
\name{frequency.name}
\alias{frequency.name}
\title{Full time frequency names}
\format{
An object of class \code{character} of length 7.
}
\usage{
frequency.name
}
\description{
Full time frequency names
}
\keyword{datasets}
