% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frostAPI.R
\name{frostAPI}
\alias{frostAPI}
\title{Function to read METNO data from Frost}
\usage{
frostAPI(
  param = "mean(air_temperature P1D)",
  stid = 18700,
  type = "observations",
  keyfile = "~/.FrostAPI.key",
  it = NULL,
  browser = "firefox",
  verbose = FALSE,
  level = "2",
  fields = "referenceTime\%2Cvalue"
)
}
\arguments{
\item{param}{variable name}

\item{stid}{station id number}

\item{type}{type of data, e.g., observations}

\item{keyfile}{location of FrostAPI key}

\item{it}{time index, e.g., a range of years or dates}

\item{browser}{internet browes, e.g., firefox or chrome}

\item{verbose}{if TRUE print information on progress}

\item{level}{don't know}

\item{fields}{don't know}
}
\description{
Read data from the METNO server Frost. This function requires an API-key that can be obtained by contacting METNO?
}
