% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_varlabel.R
\name{generate_varlabel}
\alias{generate_varlabel}
\title{Generate label for maps and plots}
\usage{
generate_varlabel(
  x = NULL,
  variable = NULL,
  unit = NULL,
  method = NULL,
  period = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{The object to be plotted. If x is provided and the variable and unit names are not explicitly stated, this information is extracted from the attributes of x.}

\item{unit}{Unit name. If provided, unit is used rather than the attr(x, 'unit')}

\item{method}{Method. If provided, this information is added to the label}

\item{period}{Time period. If provided, this information is added to the label}

\item{verbose}{If TRUE, print information about the progress}

\item{variabel}{Variable name. If provided, variable is used rather than the attr(x, 'variable')}
}
\value{
An expression to be used as a label in maps and plots
}
\description{
Create a label containing the variable name and unit and optional additional information to be used in maps and plots.
}
\examples{

data(bjornholt)
label <- generate_varlabel(bjornholt)
print(label)

}
\keyword{label}
