% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getatt.R
\name{getatt}
\alias{getatt}
\title{Reads and extracts the attribute information in a netCDF files}
\usage{
getatt(filename)
}
\arguments{
\item{filename}{Name of netCDF file including path}
}
\description{
Reads and extracts the attribute information in a netCDF files
}
