% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getncid.R
\name{getncid}
\alias{getncid}
\title{Generic function to retrieve information from a netcdf file}
\usage{
getncid(filename, path = NULL, verbose = FALSE)
}
\arguments{
\item{filename}{Names of files}

\item{path}{Path to files, if not provided in filename}

\item{verbose}{if TRUE print information about progress}
}
\description{
Generic function to retrieve information from a netcdf file
}
