% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnd.R
\name{ghcnd.meta}
\alias{ghcnd.meta}
\alias{ghcnd.data}
\alias{ghcnm.meta}
\alias{ghcnm.data}
\title{Functions to fetch data from the Global Historical Climatology Network (GHCN) data base}
\usage{
ghcnd.meta(
  param = NULL,
  src = "ghcnd",
  path = "data.GHCND",
  url = "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily",
  save.file = FALSE,
  verbose = TRUE,
  force = TRUE
)
}
\arguments{
\item{param}{climate variable}

\item{src}{source of data}

\item{path}{path to directory where to save data}

\item{url}{url to data source}

\item{save.file}{a boolean; if TRUE save data or metadata}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{force}{a boolean; if TRUE overwrite old file}
}
\description{
\code{ghcnd.meta} and \code{ghncm.meta} read and organize metadata of daily (ghcnd) and monthly (ghcnm) GHCN data.
\code{ghncd.data} and \code{ghcnd.data} read daily and monthly mean GHCN data from NOAA (\url{ftp.ncdc.noaa.gov}).
}
