% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridbox.R
\name{gridbox}
\alias{gridbox}
\title{Draw a gridbox}
\usage{
gridbox(x, col, verbose = FALSE)
}
\arguments{
\item{x}{location of gridbox: c(x0, x1, x2, x3, y1, y2, y3, y4, i)
where x0,x1,x2,x3 and y0,y1,y2,y3 are the four corners of the box on the x- and y-axes,
and i is an index specifying which element of \code{col} to use}

\item{col}{colors}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Draw a gridbox of color \code{col} in a location specified by \code{x}.
}
