% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstation.R
\name{gridstation}
\alias{gridstation}
\title{Function for gridding station data Y.}
\usage{
gridstation(Y, i = 1, verbose = FALSE, xlim = NULL, ylim = NULL)
}
\arguments{
\item{Y}{a \code{station} object}

\item{i}{index}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{xlim}{longitude range}

\item{ylim}{latitude range}
}
\value{
a \code{field} object
}
\description{
\code{gridstation} transforms a \code{station} object into a \code{field} object
by interpolation, using the package 'LatticeKrig'.
}
