% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadCRUT4.R
\name{HadCRUT4}
\alias{HadCRUT4}
\title{Download HadCRUT4 temperature data from UK MetOffice}
\usage{

  HadCRUT4(url = "http://www.metoffice.gov.uk/hadobs/hadcrut4/data/current/time_series/HadCRUT.4.6.0.0.monthly_ns_avg.txt",
  plot = FALSE)
}
\arguments{
\item{url}{url}

\item{plot}{a boolean; if TRUE show resutls in plot}
}
\description{
Download HadCRUT4 temperature data from UK MetOffice
}
