% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDD.R
\name{HDD}
\alias{HDD}
\alias{GDD}
\alias{CDD}
\title{Heating, growing and cooling degree days}
\usage{
HDD(x, x0 = 18, na.rm = TRUE)
}
\arguments{
\item{x}{station or field object}

\item{na.rm}{TRUE - remove NAs.}

\item{threshold}{threshold value}
}
\value{
Station or field objects
}
\description{
Functions to estimate \code{HDD}, \code{GDD}, and \code{CDD}.
}
\examples{

# Growing degree days:
data(ferder)
plot(as.seasons(ferder,FUN='GDD'), new=FALSE)

}
\seealso{
hotsummerdays coldwinterdays coldspells heatwavespells nwetdays plot count spell
}
\keyword{utilities}
