% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visprob.R
\name{histwet}
\alias{histwet}
\title{Compute a histogram of the wet-day mean precipitation}
\usage{
histwet(x, breaks = NULL, threshold = 1)
}
\arguments{
\item{x}{input object}

\item{breaks}{breaks of histogram}

\item{threshold}{threshold}
}
\description{
Compute a histogram for values above \code{threshold}
}
