% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howsimilar.R
\name{howsimilar}
\alias{howsimilar}
\title{aggregate}
\usage{
howsimilar(x, y, plot = "cor", FUN = NULL, regress = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A \code{\link{field}} object representing the reanalysis}

\item{y}{A \code{\link{station}} object}

\item{plot}{a boolean; if TRUE ignore NA, see \code{\link{mean}}}

\item{regress}{a boolean defaulting to FALSE}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
The call returns a station object
}
\description{
The aggregation functions are based on the S3 method for \code{zoo} objects,
but takes care of extra house keeping, such as attributes with meta data.
}
\details{
\code{howsimilar} is used for assessing reanalyses against station data.
}
\examples{
test.howsimilar()

}
\seealso{
regrid
}
\author{
R.E. Benestad
}
\keyword{utilities}
