% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.plot.R
\name{image.plot}
\alias{image.plot}
\title{Plot image}
\usage{
\method{image}{plot}(..., add = FALSE, nlevel = 64,
  horizontal = FALSE, legend.shrink = 0.9, legend.width = 1.2,
  legend.mar = ifelse(horizontal, 3.1, 5.1), legend.lab = NULL,
  legend.line = 2, graphics.reset = FALSE, bigplot = NULL,
  smallplot = NULL, legend.only = FALSE, col = NULL, pal = "heat",
  lab.breaks = NULL, axis.args = NULL, legend.args = NULL,
  midpoint = FALSE, border = NA, lwd = 1, verbose = FALSE)
}
\description{
Plot image
}
