% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.T}
\alias{is.T}
\alias{is.precip}
\alias{is.field}
\alias{is.station}
\alias{is.eof}
\alias{is.pca}
\alias{is.cca}
\alias{is.trajectory}
\alias{is.daily}
\alias{is.monthly}
\alias{is.seasonal}
\alias{is.annual}
\alias{is.model}
\alias{is.wind}
\alias{is.direction}
\alias{is.pressure}
\alias{is.dates}
\alias{is.ds}
\alias{is.dsensemble}
\alias{is.events}
\alias{is.months}
\alias{is.seasons}
\alias{is.url}
\alias{is.years}
\title{Test object type}
\usage{
is.T(x)
}
\arguments{
\item{x}{a data object}
}
\value{
a boolean
}
\description{
Test if an object is of a certain class or contains some variable
}
\examples{
data(ferder)
is.T(ferder)
is.precip(ferder)

}
\keyword{parameter,element}
