% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools_water.R
\name{kge}
\alias{kge}
\title{Estimate the Kling-Gupta model efficiency}
\usage{
kge(x, y, return_all = FALSE, scc = 1, sv = 1, sm = 1)
}
\arguments{
\item{x}{is a 1-D zoo object corresponding to the prediction}

\item{y}{is a 1-D zoo object corresponding to the reference}

\item{scc, }{sv, sm are weights within (0,1) allowing an informed change from the default weight of 1.}
}
\description{
This function returns the 2012 Kling-Gupta model efficiency using CV for the variability of \bold{x} in predicting \bold{y}.
An efficiency of one corresponds to a perfect match, while the lowest score is -\eqn{\inf}.
}
\details{
This functions return the dimensionless, multi-objective Kling-Gupta efficiency metric
, KGE, which is based on the correlation, variability ratio, and mean ratio
between the prediction and reference objects and is a modified verision of the
Nash-Sutcliffe efficiency (which again corresponds to the unbiased \eqn{R^2}).
The weight of the sub-metrics (scc, sv, sm) may be adjusted from 1 in the arguments
\deqn{kge= 1 - sqrt( (scc(cc-1)^2 + sv(CV_m/CV_o - 1)^2 + sm(mean(mod)/mean(obs)-1)^2 )) }
It is calculated for pairwise complete observations
}
\section{Warning}{

 The mean ratio may not be approriate for Celcius (can blow up around 0)
 or other variables that may have a mean within (-1,1). Though the metric is dimensionless, keep in mind that
 variables with highere avarages (e.g. temperature in Kelvin, longwave radiation) will be less penalized
 for mean devaitons than variables with lower averages (temp. in Celsius, winter SW radiation)
}
\author{
Helene B. Erlandsen
}
\references{
Gupta HV, Kling H, Yilmaz KK and Martinez GF (2009).,
\emph{"Decomposition of the mean squared error and NSE performance criteria: Implications for improving hydrological modelling."}, Journal of Hydrology, 377(1), pp. 80–91.
}
\seealso{
\code{nse}
}

