% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{lag.station}
\alias{lag.station}
\alias{lag.field}
\title{Wrap-around for lag.zoo to work on station and field objects}
\usage{
\method{lag}{station}(x, ...)
}
\description{
Wrap-around for lag.zoo to work on station and field objects
}
