% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.transform.R
\name{LatLon2UTM}
\alias{LatLon2UTM}
\title{Coordinate transformations}
\usage{
LatLon2UTM(lat, lon, zone, verbose = FALSE)
}
\arguments{
\item{lat}{A vector containing latitudes (unit: degrees east)}

\item{lon}{A vector containing longitude (unit: degrees north/south)}

\item{zone}{UTM zone}

\item{verbose}{If TRUE, print out diagnosics}
}
\description{
Transform latitude and longitude to UTM (Universal Transverse Mercator)
}
\seealso{
UTM2LatLon
}
\author{
K. Tunheim
}
