% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesfromthredds.R
\name{list_thredds}
\alias{list_thredds}
\title{A function that lists files on thredds catalogue.}
\usage{
list_thredds(
 
    caturl = "https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/seasonal_RCM/catalog.html",
  extension = ".nc",
  pattern = NULL,
  verbose = FALSE
)
}
\arguments{
\item{caturl}{url of the thredds ctalogue containing the files}

\item{extension}{the extension ot the netCDF files (default is ".nc")}

\item{pattern}{text pattern to filter file names}

\item{verbose}{if true promt test lines}
}
\description{
A function that lists files on thredds catalogue.
}
\examples{

list_thredds()
list_thredds(pattern=c("rcp45","r1i1p1","RACMO"))
list_thredds(caturl="http://esgf3.dkrz.de/thredds/catalog/esgcet/12/CMIP6.ScenarioMIP.DKRZ.MPI-ESM1-2-HR.ssp585.r1i1p1f1.day.pr.gn.v20190710.html")
\dontrun{
caturl <- "https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/bias_corrected/3DBC/cross-validation/noresm-r1i1p1-remo/tasmin/catalog.html"
extension <- '.nc4'
list_thredds(caturl,extension)
}
}
\seealso{
retrieve
}
