% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattributes.R
\name{lon}
\alias{lon}
\alias{longitude}
\alias{lat}
\alias{latitude}
\alias{alt}
\alias{altitude}
\alias{stid}
\alias{qual}
\alias{quality}
\alias{calendar}
\alias{cntr}
\alias{country}
\alias{loc}
\alias{location}
\alias{varid}
\alias{variable}
\alias{unit}
\alias{src}
\alias{aspect}
\alias{ref}
\alias{reference}
\alias{info}
\alias{information}
\alias{history.esd}
\alias{pattern}
\alias{ele}
\alias{element}
\alias{err}
\title{Shortcuts to attributes}
\usage{
lon(x)
}
\arguments{
\item{x}{input object}
}
\description{
Fast access to attributes, e.g, \code{lon(x)} gives you the same output as \code{attr(x,"longitude")}.
}
