\name{manual}
\alias{manual}
\alias{ABC4ESD}
\alias{esd.tips}
\alias{esd.issues}
\docType{data}
\title{Manual for esd}
\description{
  Help, assistance and manuals for \code{esd} and empirical-statistical
  downsclaing (ESD). 
}
\usage{
manual(url='https://ndownloader.figshare.com/files/2126237',browser='firefox')
ABC4ESD(url='http://www.gvc2.gu.se/ngeo/rcg/edu/esd.pdf',browser='firefox')
esd.tips(url='https://github.com/metno/esd/wiki',browser='firefox')
esd.issues(url='https://github.com/metno/esd/issues',browser='firefox')
}
\source{

}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
manual()
}
\keyword{datasets}
