% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.dsensemble.R
\name{map.dsensemble}
\alias{map.dsensemble}
\title{expandpca}
\usage{
\method{map}{dsensemble}(
  x,
  it = c(2000, 2099),
  is = NULL,
  im = NULL,
  ip = NULL,
  colbar = list(pal = NULL, rev = FALSE, n = 10, breaks = NULL, pos = 0.05, show =
    TRUE, type = "p", cex = 2, h = 0.6, v = 1),
  FUN = "mean",
  FUNX = "mean",
  verbose = FALSE,
  anomaly = FALSE,
  test = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of type 'pca'}

\item{it}{time index (see \code{\link{subset}})}

\item{FUN}{function applied to aggregate in time}

\item{FUNX}{function applied aggregate ensemble members}

\item{verbose}{if TRUE print progress}

\item{anomaly}{if FALSE add the mean value stored as attribute in x}

\item{test}{if TRUE perform test on one GCM simulation}
}
\description{
\code{map.dsensemble} is a wrapper for \code{map} that distills ensemble statistics from PCA-based dsensemble objects. 
The argument 'FUNX' is the function that is applied to the ensemble. If 'FUNX' is 'mean', the function \code{expandpca} is 
used to aggregate the data, otherwise the function \code{aggregate.dsensemble} is employed. The difference is that 
\code{expandpca} applies 'FUNX' to the PCs before expanding the data, while \code{aggregate.dsensemble} expands the data
for all ensemble members before aggregating them which is more suitable when calculating e.g. the .
}
\seealso{
map aggregate.dsensemble expandpca
}
