% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mask}
\alias{mask}
\title{Function that masks either ocean or land}
\usage{
mask(x, land = FALSE)
}
\arguments{
\item{x}{a \code{field} object}

\item{land}{a boolean; if TRUE mask land, if FALSE mask ocean}
}
\description{
Uses topography from \code{\link{etopo5}} to mask either land or ocean.
}
