% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchdate.R
\name{matchdate}
\alias{matchdate}
\alias{matchdate.list}
\alias{matchdate.default}
\title{Match date of one object with another object or date string}
\usage{
matchdate(x, it, verbose = FALSE, ...)
}
\arguments{
\item{x}{input object (e.g., \code{station}, \code{field} or \code{zoo}) with a date index}

\item{it}{a character string with dates or an object (e.g., \code{station}, \code{field} or \code{zoo}) with a date index}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
the part of \code{x} that matches the dates provided in \code{it}
}
\description{
Match date of one object with another object or date string
}
\seealso{
subset
}
