% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meps.thredds.R
\name{meps}
\alias{meps}
\title{Function to read weather forecast data for the Nordic domain from thredds.met.no}
\usage{
meps(url = "https://thredds.met.no/thredds/catalog/metpplatest",
  type = "forecast", param = "slp", lon = c(9.5, 11.5), lat = c(59,
  61), it = "latest", dt = 50, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{url}{URL for the data on thredds.met.no}

\item{param}{Variable name (rr = precipitation, tg = temperature)}

\item{lon}{Longitude selection (=NULL reads all)}

\item{lat}{Latitude selection (=NULL reads all)}

\item{it}{Index time - a range of dates or years to select (e.g., it=as.Date(c("2010-01-01","2010-01-31")))}

\item{dt}{Number of time steps to access at a time (default: 50). 
A smaller value can be set when requesting a large spatial domain as the server 
doesn't like to deal with large data amounts.}

\item{verbose}{write out diagnostics}

\item{plot}{plot the results while reading.}

\item{FUN}{Function for daily aggregation. =NULL gives raw data}

\item{ncfile}{netCDF file to access (path to a file or url)}

\item{path}{path to ncfile (use only if ncfile doesn't contain the full path)}
}
\value{
A "zoo" "station" object with additional attributes used for further
processing.
}
\description{
The data are post processed forecasts or analysis based on the MetCoOp Ensemble Prediction System (MEPS).
As the resolution is high both in space (1x1km) and time (1 hours), it may be necessary
to limit the spatial and/or temporal scope (see input options 'it', 'lon', 'lat')
as you will otherwise run out of memory.
}
\examples{
it <- "latest"
lon <- c(-2,15)
lat <- c(55,63)
slp <- meps(param="slp", lon=lon, lat=lat, it=it, verbose=TRUE)
map(slp, FUN="mean")

}
\seealso{
senorge, station.thredds
}
