% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.ESGF.R
\name{meta.ESGF}
\alias{meta.ESGF}
\title{Read metadata from ESGF}
\usage{
meta.ESGF(
  url = "https://esgf-data.dkrz.de/esg-search/search/",
  mip = "CMIP6",
  param = "tas",
  freq = "mon",
  expid = "ssp585",
  verbose = FALSE,
  n = NULL
)
}
\arguments{
\item{url}{URL}

\item{mip}{CMIP5 or CMIP6}

\item{param}{meteorological variable name, e.g., "tas"}

\item{freq}{temporal frequency, e.g., "mon"=monthly}

\item{expid}{scenario (ssp for CMIP6, rcp for CMIP5)}

\item{verbose}{if TRUE print progress}

\item{n}{number of datasets}
}
\description{
Read metadata from ESGF
}
\seealso{
retrieve.ESGF
}
