% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesfromthredds.R
\name{meta.s_enda}
\alias{meta.s_enda}
\title{A function that provides metadata for S-ENDA data on thredds.}
\usage{
meta.s_enda(
 
    caturl = "https://thredds.met.no/thredds/catalog/met.no/observations/surface/catalog.html",
  verbose = FALSE
)
}
\arguments{
\item{caturl}{url of the thredds ctalogue containing the files}

\item{verbose}{if true promt test lines}
}
\description{
A function that provides metadata for S-ENDA data on thredds.
}
\examples{

meta <- meta.s_enda()
}
\seealso{
retrieve list_thredds station.s_enda
}
