% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaextract.R
\name{metaextract}
\alias{metaextract}
\alias{metaextract.cordex}
\alias{metaextract.cmip}
\title{Function to extract the metadata from local NetCDF files}
\usage{
metaextract(
  files.in,
  file.out = "meta.rda",
  path.in = NULL,
  path.out = NULL,
  add = TRUE,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{files.in}{Filenames of netcdf files}

\item{file.out}{Filename of .rda file where the metadata will be saved}

\item{path.in}{Path to files.in}

\item{path.out}{Path to file out}

\item{add}{If TRUE, add metadata to the existing metadata file. If FALSE create a new metadata file.}

\item{force}{If TRUE, replace metadata of files that that are already in the existing metadata. 
If FALSE only add metadata for new files.}

\item{verbose}{if TRUE print information about progress}
}
\description{
The function \code{metaextract} extracts metadata from local NetCDF files containing climate model data 
from the CORDEX and CMIP projects. The function \code{metaextract} redirects to \code{metaextract.cordex}
or \code{metaextract.cmip} depending on the project id.
}
\seealso{
getncid
}
