% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frostAPI.R
\name{metafrostAPI}
\alias{metafrostAPI}
\title{Retrieve metadata of MetNo data from frost.met.no}
\usage{
metafrostAPI(keyfile = "~/.FrostAPI.key", verbose = FALSE,
  fields = "id,name,masl,county,municipality,wmoid,geometry,type",
  url = "frost.met.no/sources/v0.jsonld?country=NO")
}
\arguments{
\item{keyfile}{location of API key}

\item{fields}{fields to be extracted}

\item{url}{url of data}
}
\description{
Retrieve metadata of MetNo data from frost.met.no
}
