% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metno.frost.data.R
\name{metno.frost.data}
\alias{metno.frost.data}
\alias{metno.frost.station}
\title{Retrieve records from frost.met.no, for one element and one or more stations}
\usage{
metno.frost.data(
  keyfile = "~/.FrostAPI.key",
  url = "https://frost.met.no/auth/requestCredentials.html",
  stid = NULL,
  param = NULL,
  it = NULL,
  lon = NULL,
  lat = NULL,
  loc = NULL,
  alt = NULL,
  cntr = NULL,
  timeresolutions = "P1M",
  levels = "default",
  timeoffsets = "default",
  performancecategories = "A,B,C",
  exposurecategories = "1,2",
  qualities = "0,1,2,3,4,5",
  fetch.meta = TRUE,
  path = NULL,
  browser = "firefox",
  save2file = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{keyfile}{The path where the frost.met.no credentials keyfile is stored locally.}

\item{url}{The URL of the data portal or webpage for requesting new client credentials.}

\item{stid}{A string of characters as an identifier of the weather/climate
station.}

\item{param}{Parameter or element type or variable identifier. There are
several core parameters or elements as well as a number of additional
parameters. The parameters or elements are: precip = Precipitation (mm) tas,
tavg = 2m-surface temperature (in degrees Celcius) tmax, tasmax = Maximum
temperature (in degrees Celcius) tmin, tasmin = Minimum temperature (in
degrees Celcius)}

\item{it}{A vector of two dates (from and to) in the form "2014-01-01".}

\item{lon}{Numeric value of longitude (in decimal degrees East) for the
reference point (e.g. weather station) as a single value or a vector
containing the range of longitude values in the form of c(lon.min,lon.max)}

\item{lat}{Numeric value of latitude for the reference point (in decimal
degrees North) or a vector containing the range of latitude values in the
form of c(lat.min,lat.max)}

\item{loc}{A string of characters as the name of the location
(weather/climate station) or an object of class "stationmeta".}

\item{alt}{Numeric value of altitude (in meters a.s.l.) used for selection.
Positive value, select all stations above this altitude; for negative
values, select all stations below this latitude.}

\item{cntr}{A string or a vector of strings of the full name of the country:
Select the stations from a specified country or a set of countries.}

\item{timeresolutions}{A string of the desired time resolution. Accepted strings
are the ISO 8601 durations P1M, P1D and PT1M, as well as the strings MONTHLY,
MONTH, DAILY, DAY, MINUTE and MIN. If no value is set, this defaults to P1M.}

\item{levels}{Numeric value of sensor height used for selection. If no value
is set, then standard heights will be selected.}

\item{timeoffsets}{A string of the ISO 8601 duration desired for the offset
value of timeseries. If no value is set, then the best available option will 
be selected.}

\item{performancecategories}{A string with a comma-separated list of categories
to accept for sensor performance. If no value is set, then the default A,B,C
will be used.}

\item{exposurecategories}{A string with a comma-separated list of categories to
accept for siting quality of the sensor. If no value is set, then the default
1,2 will be used.}

\item{qualities}{A string with a comma-separated list of quality codes to
accept for the individual measurements. If no value is set, then the default
0,1,2,3,4,5 will be used.}

\item{fetch.meta}{Logical value defaulting to TRUE. If TRUE, download timeseries
metadata from frost.met.no. If FALSE, try to load data file station.meta instead.}

\item{path}{The path where the data are stored. Can be a symbolic link.}

\item{browser}{A string signifying which internet browser to open to ask
user to create new credentials for frost.met.no}

\item{save2file}{Logical value defaulting to FALSE. If TRUE, save a file
with the output data.}

\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{\dots}{additional arguments}
}
\value{
A time series of "zoo" "station" class with additional attributes
used for further processing.
}
\description{
Retrieve records from frost.met.no, for one element and one or more stations
}
\examples{

\dontrun{ 
metno.frost.data(param='t2m', stid=18700, it=c('2020-01-01','2020-02-01'), timeresolutions='P1D')
}

}
\author{
K. Tunheim
}
\keyword{parameter,data,metno,norway,frost}
