% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcombine.R
\name{mvcombine}
\alias{mvcombine}
\title{Multivariate combination of fileds}
\usage{
mvcombine(X, Y, is = NULL, verbose = FALSE)
}
\arguments{
\item{X}{A field object.}

\item{Y}{A field object.}

\item{is}{A list or data.frame providing space index, e.g. a list of longitude and latitude range like list(lon=c(0,60), lat=c(35,60)).}

\item{verbose}{If TRUE, print out diagnosics.}
}
\description{
\code{mvcombine} combines two fields with different variables by attaching them along the spatial axis.
Before combining the two fields, the same time slices are selected from both fields and
the data are normalized by subtracting the mean and dividing by the standard deviation.
The output of mvcombine is an 'mvcomb' object to which you can apply a multivariate
Empirical Orthogonal Function (EOF) using the function \code{EOF}.
}
\seealso{
EOF.mvcomb EOF
}
