% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateindex.R
\name{NAO}
\alias{NAO}
\alias{NINO3.4}
\alias{SOI}
\alias{GSL}
\alias{GSL.nasa}
\alias{QBO}
\alias{CET}
\alias{CO2}
\alias{AMO}
\alias{IOD}
\alias{Sunspots}
\title{Functions to read climate indices from sources on the internet}
\usage{
NAO(freq = "monthly", url = NULL, header = FALSE, verbose = FALSE)
}
\arguments{
\item{freq}{frequency}

\item{url}{a URL or web address to location of data}

\item{header}{a boolean, indicating if file includes a header or not}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
NAO: Daily North Atlantic Oscillation index from NCEP/NOAA
}
\details{
NINO3.4: Daily Nino3.4 index provided by NOAA/NCDC downloaded from KNMI Climate Explorer

SOI: Souther Oscillation index from The Australian Bureau fo Meteorology (bom.gov.au)

GSL: Global Average Absolute Sea level Change, 1880-2015,
from EPA's Climate Change Indicators in the United States: www.epa.gov/climate-indicators

GSL.nasa: Global Average Sea level from NASA
GSL.aviso: Global Average Sea level from AVISO

QBO: Quasi-Biennial Oscillation. Calculated at NOAA/ESRL/PSD from the zonal average
of the 30mb zonal wind at the equator as computed from the NCEP/NCAR Reanalysis.

CET: Central England Temperature from the Hadley Center

CO2: Carbon dioxide at Mauna Loa, Hawaii, from NOAA ESRL.
Reference: 'C.D. Keeling, R.B. Bacastow, A.E. Bainbridge, C.A. Ekdahl, P.R. Guenther, and L.S. Waterman, (1976),
Atmospheric carbon dioxide variations at Mauna Loa Observatory, Hawaii, Tellus, vol. 28, 538-551'

AMO: Atlantic Multidecadal Oscillation, unsmoothed calculated from the Kaplan SST V2 at NOAA/ESRL/PSD1

IOD: Indian Ocean Dipole index

Sunspots: updated monthly sunspot number
}
