% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addstations2nc.R
\name{nc4add.stations}
\alias{nc4add.stations}
\title{Combine or append netCDF files for stations into one. Used with write2ncdf4.station to bypass memory demand/limitations.
The \code{write2ncdf4.station} with the argument append=TRUE also saves data chunk-wise. Another function is 
\code{ncrcat} which reads all data and attributes in listed files and sacve them into a new file for all the 
inputs combined. 
Rasmus.Benestad, met.no, 2023-07-31}
\usage{
nc4add.stations(x, fname, verbose = FALSE)
}
\arguments{
\item{x}{a station object}

\item{fname}{list with file names of netCDF files with station data}

\item{verbose}{used for debugging and diagnostics of executing the code}

\item{out}{name of the output file}

\item{namelength}{- see write2ncdf4.station}

\item{prec}{precision}

\item{offset}{offset value for the data}

\item{scale}{scaling value for the data}

\item{missval}{missing value}

\item{torg}{time origin}

\item{it}{index time - selection of time interval unless NULL}

\item{stid}{selection of station IDs unless set to NULL}
}
\description{
Combine or append netCDF files for stations into one. Used with write2ncdf4.station to bypass memory demand/limitations.
The \code{write2ncdf4.station} with the argument append=TRUE also saves data chunk-wise. Another function is 
\code{ncrcat} which reads all data and attributes in listed files and sacve them into a new file for all the 
inputs combined. 
Rasmus.Benestad, met.no, 2023-07-31
}
\examples{
\dontrun{
  ncfiles <- list.files(path='~/Downloads/ecad2nc4',pattern='.ecad.nc',full.names=TRUE)
  nc4combine.stations(ncfiles,out=paste0('~/Downloads/',variable,'.ecad.nc'))
}

}
\seealso{
write2ncdf4.station as.station retrieve.station
}
\keyword{utilities}
