% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{NE}
\alias{NE}
\title{Various formulas, equations and transforms.}
\usage{
NE(p)
}
\arguments{
\item{p}{a probability?}
}
\value{
The right hand side of the equation
}
\description{
\code{NE}: predicts the number of events given the probability Pr.
}
\seealso{
C.C.eq precip.vul t2m.vul precip.Pr t2m.Pr
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
