% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{\code{nearest} selects the time series in \code{x} that are closest to the locations specified in \code{is}}
\usage{
nearest(x, is)
}
\arguments{
\item{x}{a \code{station} or \code{field} object (if x is a field it will first be transformed to a station object)}

\item{is}{a spatial index, either a list or data frame containing coordinates (e.g., list(lon=c(...), lat=c(...))) or
a \code{station} or \code{field} object whos longitudes and latitudes will be used}
}
\description{
\code{nearest} selects the time series in \code{x} that are closest to the locations specified in \code{is}
}
