% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools_water.R
\name{nse}
\alias{nse}
\title{Estimate the Nash-Sutcliffe model efficiency}
\usage{
nse(x, y)
}
\arguments{
\item{x}{is a 1-D zoo object corresponding to the prediction}

\item{y}{is a 1-D zoo object corresponding to the reference}
}
\description{
This function returns the dimensionless Nash Sutcliffe model efficiency, evalutaing \bold{x} as a prediction of \bold{y}.
An efficiency of one corresponds to a perfect match, while the lowest score is -\eqn{\inf}.
}
\details{
This functions return the multi-objective Nash-Sutcliffe efficiency metric
, NSE, which corresponds to the unbiased \eqn{R^2}. The metric is scaled by the observed variance. The metric is given as
 \deqn{nse= 1 - mean_square_error/observed_variance }
 and varies thus form 1 to minus infinity. It is calculated for pairwise complete observations
}
\section{Warning}{

Keep in mind that in regions with higher variance (e.g. seasonality) equal absolute deviations will be penalized less than for regions with a lower variance
}
\author{
Helene B. Erlandsen
}
\references{
Nash JE and Sutcliffe JV (1970),
\emph{“River flow forecasting through conceptual models part I—A discussion of principles.”},
 Journal of hydrology, 10(3), pp. 282–290.,

 Gupta HV, Kling H, Yilmaz KK and Martinez GF (2009).,
\emph{"Decomposition of the mean squared error and NSE performance criteria: Implications for improving hydrological modelling."}, Journal of Hydrology, 377(1), pp. 80–91.
}
\seealso{
\code{kge}
}

