% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.events.R
\name{param.events}
\alias{param.events}
\title{Aggregate an \code{events} object in time and space}
\usage{
param.events(x, param = "count", FUN = "mean", verbose = TRUE,
  longname = NULL, unit = NULL, ...)
}
\arguments{
\item{x}{the input object of class \code{events}}

\item{param}{parameter to return, e.g., "count" or "pcent"}

\item{FUN}{a function}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{longname}{long variable name of output, e.g., "number of cyclones per month" or "mean sea level pressure"}

\item{unit}{unit of output}

\item{...}{other arguments}
}
\value{
a \code{station} object
}
\description{
Transform an \code{events} object (output of \code{CCI} including information about cyclones or anticyclones) 
into a \code{station} object by aggregating some characteristic of the events in time and space. 
By default the argument \code{param} is set to count 'count' and \code{param.events} returns the number of 
events per month (calculated with \code{count.events}). Depending on the input parameters \code{param} and \code{FUN}, 
the function can also return some other characteristic of the events, e.g., the monthly mean sea level pressure 
at the center of the cyclones (\code{param}='pcent' and \code{FUN}='mean').
}
\seealso{
as.station.events count.events CCI
}
