% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diagnose.R
\name{plot.diagnose}
\alias{plot.diagnose}
\alias{plot.diagnose.comb.eof}
\alias{plot.diagnose.dsensemble}
\alias{plot.diagnose.matrix}
\title{Diagnosis plots}
\usage{
\method{plot}{diagnose}(x, ...)
}
\description{
Produce plots to diagnose and examine different data types and checks for consistency. The method
plot.diagnose.matrix has an additional agrument 'calibrate' that can make the target plot presentation
stricter or leaner (default is 100).
}
\seealso{
diagnose
}
