% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dsensemble.multi}
\alias{plot.dsensemble.multi}
\title{Plot multiple stations/spatially aggregated field.}
\usage{
\method{plot}{dsensemble.multi}(
  x,
  it = c(2000, 2099),
  FUNX = "mean",
  verbose = FALSE,
  anomaly = FALSE,
  test = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{input object of class 'dsensemble'}

\item{it}{a time index, see \code{\link{subset}}}

\item{FUNX}{a function}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{anomaly}{a boolean; if TRUE show anomalies}

\item{test}{a boolean; if TRUE perform some test?}

\item{plot}{a boolean; if TRUE show plot}

\item{\dots}{additional arguments}
}
\description{
Plot multiple stations/spatially aggregated field.
}
\seealso{
\code{\link{plot}} plot.station plot.ds plot.eof
}
