% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dsensemble.one}
\alias{plot.dsensemble.one}
\title{Plot one station}
\usage{
\method{plot}{dsensemble.one}(
  x,
  pts = FALSE,
  it = 0,
  envcol = rgb(1, 0, 0, 0.2),
  legend.show = TRUE,
  ylab = NULL,
  obs.show = TRUE,
  target.show = TRUE,
  map.show = TRUE,
  map.type = NULL,
  map.insert = TRUE,
  new = FALSE,
  xrange = NULL,
  yrange = NULL,
  alpha = 0.5,
  alpha.map = 0.7,
  mar = c(5.1, 4.5, 4.1, 2.1),
  cex.axis = 1,
  cex.lab = 1.2,
  cex.main = 1.2,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{input object of class 'dsensemble'}

\item{pts}{a boolean; if TRUE show points}

\item{envcol}{color of envelope}

\item{obs.show}{if TRUE show observations}

\item{target.show}{if TRUE show diagnosis as target plot}

\item{map.type}{type of map, 'points' or 'rectangle'}

\item{map.insert}{if TRUE show map as insert, else show in new window}

\item{alpha}{transparancy factor of main plot}

\item{alpha.map}{transparancy factor of map}

\item{mar}{see \code{\link[graphics]{par}}}

\item{cex.axis}{see \code{\link[graphics]{par}}}

\item{cex.lab}{see \code{\link[graphics]{par}}}

\item{cex.main}{see \code{\link[graphics]{par}}}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}

\item{anomaly}{a boolean; if TRUE show anomalies}

\item{test}{a boolean; if TRUE perform some test?}

\item{plot}{a boolean; if TRUE show plot}
}
\description{
Plot one station
}
