% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dsensemble}
\alias{plot.dsensemble}
\title{Plot esd objects}
\usage{
\method{plot}{dsensemble}(x, verbose = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{the object to be plotted}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{plot}{a boolean; if TRUE show plot}

\item{\dots}{additional arguments}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\details{
\code{plot.dsensemble} plots a downscaled GCM ensemble.
}
\seealso{
\code{\link{plot}}
}
