% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eof}
\alias{plot.eof}
\title{Plot esd objects}
\usage{
\method{plot}{eof}(x, ..., new = FALSE, xlim = NULL, ylim = NULL,
  ip = 1, what = c("pc", "eof", "var"), colbar = list(pal = NULL, rev
  = FALSE, n = 10, alpha = 0.8, breaks = NULL, type = "p", cex = 2, show =
  TRUE, h = 0.6, v = 1, pos = 0.05), verbose = FALSE, is = NULL,
  it = NULL)
}
\arguments{
\item{x}{the object to be plotted}

\item{\dots}{additional arguments}

\item{new}{if TRUE plot in new window}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{ip}{Which EOF/CCA pattern (mode) to plot}

\item{what}{Indicate what to plot. For \code{plot.eof}, c('pc', 'eof', 'var') is the default setting which means that
the plot will include the principle components, EOF patterns and explained variance. 'field' expands eof to field before
plotting}

\item{colbar}{a list, see \code{\link{colbar}}}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{is}{For subsetting in space - See \code{\link{subset}}. Can also be
a station value and if provided, the plotting will involve an interpolation
to the same coordinates as defined by \code{is}.}

\item{it}{For subsetting in time - See \code{\link{subset}}.}

\item{plot.type}{"single"}

\item{lwd}{width of line}

\item{type}{type of plot: 'l' = line, 'p' = point, 'b' = both}

\item{pch}{type of marker}

\item{main}{main title}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{errorbar}{if TRUE show errorbar}

\item{legend.show}{if TRUE show legendp}

\item{map.show}{show map of stations}

\item{map.type}{'points' to show stations on map, 'rectangle' to show area}

\item{map.insert}{if TRUE show map as insert, else show map in new window}

\item{cex}{magnification factor, see \code{\link[graphics]{par}}}

\item{cex.axis}{see \code{\link[graphics]{par}}}

\item{cex.lab}{see \code{\link[graphics]{par}}}

\item{cex.main}{see \code{\link[graphics]{par}}}

\item{mar}{see \code{\link[graphics]{par}}}

\item{fig}{coordinates of figure region, see \code{\link[graphics]{par}}}

\item{alpha}{transparency factor for main plot}

\item{alpha.map}{transparency factor for map}

\item{col}{Colour see \code{\link[graphics]{par}}}

\item{lwd}{width of line}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\examples{

# Example: use aggregate to compute annual mean temperature for Svalbard:
data(Svalbard)
y <- aggregate(Svalbard, by=year(Svalbard), FUN='mean', na.rm = FALSE) 
plot(y)

# Example with downscaling:
lon <- c(-12,37)
lat <- c(52,72)
t2m <- t2m.DNMI(lon=lon,lat=lat)
data(Oslo)
ds <- DS(Oslo,t2m)

# Plot the results for January month
# plot(subset(ds,it='Jan'))

# Plot the residuals:
residual <- as.residual(ds)
obs <- as.anomaly(as.calibrationdata(ds))

plot.zoo(obs,lwd=2)
lines(residual,col="red")

print("Global climate model simulation NorESM")
T2m <- t2m.NorESM.M(lon=lon,lat=lat)

# Plot the global mean of the field:
plot(T2m)
# Plot area mean of a sub region
plot(T2m,is=list(lon=c(0,10),lat=c(60,70)))

# Plot interpolated results corresponding to ferder
data(ferder)
plot(T2m,ferder)

# Plot Hovmuller diagram: Not working ...
## plot(T2m,is=list(lon=0)) 

print("Extract a subset - the January month")
x <- subset(t2m,it="jan")
X <- subset(T2m,it="jan")

print("Combine the fields for computing common EOFs:")
XX <- combine(x,X)

print("Compute common EOFs")
eofxx <- EOF(XX)
plot(eofxx)

print("Downscale the January mean temperature") 
ds.jan <- DS(Oslo,eofxx)
plot(ds.jan)

}
\keyword{graphics}
\keyword{plot}
