% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nevents}
\alias{plot.nevents}
\title{Plot esd objects}
\usage{
\method{plot}{nevents}(
  x,
  verbose = FALSE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  ...
)
}
\arguments{
\item{x}{the object to be plotted}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{main}{main title}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{col}{Colour see \code{\link[graphics]{par}}}

\item{\dots}{additional arguments}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\seealso{
\code{\link{plot}}
}
