% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.station}
\alias{plot.pca}
\alias{vis.map}
\alias{plot.eof}
\alias{plot.eof.field}
\alias{plot.eof.var}
\alias{plot.eof.comb}
\alias{plot.field}
\alias{plot.spell}
\alias{plot.cca}
\alias{plot.ds}
\alias{plot.ds.pca}
\alias{plot.ds.eof}
\alias{plot.dsx}
\alias{plot.dsensemble}
\alias{plot.diagnose}
\alias{plot.xval}
\alias{plot.diagnose.comb.eof}
\alias{plot.diagnose.matrix}
\alias{plot.diagnose.dsensemble}
\alias{plot.nevents}
\alias{plot.ssa}
\title{Plot esd objects}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{the object to be plotted}

\item{ip}{Which EOF/CCA pattern (mode) to plot}

\item{col}{Colour}

\item{is}{For subsetting in space - See \code{link{subset}}, but can also be
a station value and if provided, the plotting will involve an interploation
to the same coordinates as defined by \code{is}.}

\item{it}{For subsetting in time - See \code{link{subset}}.'it=0' returns
the annual means (mean of DJF + MAM + JJA + SON)}

\item{FUN}{function}

\item{obs.show}{Show the observations}

\item{target.show}{Show a 'target plot' for quality}

\item{map.show}{Show a small map of the location}

\item{map.type}{type of map: 'rectangle' to show area or 'points' to show points}

\item{what}{Indicate what to plot. 'field' expands eof to field before
plotting}

\item{new}{if TRUE plot in new window}
}
\value{
A field object
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\examples{

# Example: use aggregate to compute annual mean temperature for Svalbard:
data(Svalbard)
y <- aggregate(Svalbard, by=year(Svalbard), FUN='mean', na.rm = FALSE) 
plot(y)

# Example with downscaling:
lon <- c(-12,37)
lat <- c(52,72)
t2m <- t2m.DNMI(lon=lon,lat=lat)
data(Oslo)
ds <- DS(Oslo,t2m)

# Plot the results for January month
# plot(subset(ds,it='Jan'))

# Plot the residuals:
residual <- as.residual(ds)
obs <- as.anomaly(as.calibrationdata(ds))

plot.zoo(obs,lwd=2)
lines(residual,col="red")

print("Global climate model simulation NorESM")
T2m <- t2m.NorESM.M(lon=lon,lat=lat)

# Plot the global mean of the field:
plot(T2m)
# Plot area mean of a sub region
plot(T2m,is=list(lon=c(0,10),lat=c(60,70)))

# Plot interpolated results corresponding to ferder
data(ferder)
plot(T2m,ferder)

# Plot Hovmuller diagram: Not working ...
## plot(T2m,is=list(lon=0)) 

print("Extract a subset - the January month")
x <- subset(t2m,it="jan")
X <- subset(T2m,it="jan")

print("Combine the fields for computing common EOFs:")
XX <- combine(x,X)

print("Compute common EOFs")
eofxx <- EOF(XX)
plot(eofxx)

print("Downscale the January mean temperature") 
ds.jan <- DS(Oslo,eofxx)
plot(ds.jan)

}
\seealso{
\code{\link{plot}}
}
\keyword{hplot}
