% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.spell}
\alias{plot.spell}
\title{Plot esd objects}
\usage{
\method{plot}{spell}(x, ..., xlim = NULL, ylim = NULL)
}
\arguments{
\item{x}{the object to be plotted}

\item{\dots}{additional arguments}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\details{
\code{plot.spell} plots wet/dry or cold/warm spells.
}
\seealso{
\code{\link{plot}} plot.station plot.ds plot.eof plot.field
}
