% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ssa}
\alias{plot.ssa}
\title{Plot esd objects}
\usage{
\method{plot}{ssa}(x, ..., main = "SSA analysis", sub = "", verbose = FALSE)
}
\arguments{
\item{x}{the object to be plotted}

\item{\dots}{additional arguments}

\item{main}{main title}

\item{sub}{subtitle}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\seealso{
\code{\link{plot}} plot.station plot.ds plot.eof plot.field
}
