% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.station}
\alias{plot.station}
\title{Plot esd objects}
\usage{
\method{plot}{station}(
  x,
  ...,
  plot.type = "single",
  new = TRUE,
  lwd = 3,
  type = "l",
  pch = 0,
  main = NULL,
  col = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  ylab = NULL,
  errorbar = TRUE,
  legend.show = FALSE,
  map.show = TRUE,
  map.type = NULL,
  map.insert = TRUE,
  xrange = NULL,
  yrange = NULL,
  cex.axis = 1.2,
  cex.lab = 1.2,
  cex.main = 1.2,
  mar = NULL,
  alpha = 0.5,
  alpha.map = 0.7,
  verbose = FALSE
)
}
\arguments{
\item{x}{the object to be plotted}

\item{\dots}{additional arguments}

\item{plot.type}{"single"}

\item{new}{if TRUE plot in new window}

\item{lwd}{width of line}

\item{type}{type of plot: 'l' = line, 'p' = point, 'b' = both}

\item{pch}{type of marker}

\item{main}{main title}

\item{col}{Colour see \code{\link[graphics]{par}}}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{errorbar}{if TRUE show errorbar}

\item{legend.show}{if TRUE show legendp}

\item{map.show}{show map of stations}

\item{map.type}{'points' to show stations on map, 'rectangle' to show area}

\item{map.insert}{if TRUE show map as insert, else show map in new window}

\item{cex.axis}{see \code{\link[graphics]{par}}}

\item{cex.lab}{see \code{\link[graphics]{par}}}

\item{cex.main}{see \code{\link[graphics]{par}}}

\item{mar}{see \code{\link[graphics]{par}}}

\item{alpha}{transparency factor for main plot}

\item{alpha.map}{transparency factor for map}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{it}{For subsetting in time - See \code{\link{subset}}.}

\item{is}{For subsetting in space - See \code{\link{subset}}. Can also be
a station value and if provided, the plotting will involve an interpolation
to the same coordinates as defined by \code{is}.}

\item{ip}{Which EOF/CCA pattern (mode) to plot}

\item{cex}{magnification factor, see \code{\link[graphics]{par}}}

\item{fig}{coordinates of figure region, see \code{\link[graphics]{par}}}

\item{what}{Indicate what to plot. For \code{plot.eof}, c('pc', 'eof', 'var') is the default setting which means that
the plot will include the principle components, EOF patterns and explained variance. 'field' expands eof to field before
plotting}

\item{colbar}{a list, see \code{\link{colbar}}}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\examples{

# Example: use aggregate to compute annual mean temperature for Svalbard:
data(Svalbard)
y <- aggregate(Svalbard, by=year(Svalbard), FUN='mean', na.rm = FALSE) 
plot(y, new=FALSE)

# Example with downscaling:
lon <- c(-12,37)
lat <- c(52,72)
t2m <- t2m.DNMI(lon=lon,lat=lat)
data(Oslo)
ds <- DS(Oslo,t2m)

# Plot the results for January month
# plot(subset(ds,it='Jan'))

# Plot the residuals:
residual <- as.residual(ds)
obs <- as.anomaly(as.calibrationdata(ds))

plot.zoo(obs[,1], lwd=2, new=FALSE)
lines(residual, col="red")

print("Global climate model simulation NorESM")
T2m <- t2m.NorESM.M(lon=lon,lat=lat)

# Plot the global mean of the field:
plot(T2m, new=FALSE)
# Plot area mean of a sub region
plot(T2m,is=list(lon=c(0,10),lat=c(60,70)), new=FALSE)

# Plot interpolated results corresponding to ferder
data(ferder)
plot(T2m,ferder, new=FALSE)

print("Extract a subset - the January month")
x <- subset(t2m,it="jan")
X <- subset(T2m,it="jan")

print("Combine the fields for computing common EOFs:")
XX <- combine(x,X)

print("Compute common EOFs")
eofxx <- EOF(XX)
plot(eofxx, new=FALSE)

print("Downscale the January mean temperature") 
ds.jan <- DS(Oslo,eofxx)
plot(ds.jan, new=FALSE)

}
\keyword{graphics}
\keyword{plot}
