% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trajectory}
\alias{plot.trajectory}
\title{Plot esd objects}
\usage{
\method{plot}{trajectory}(
  x,
  it = NULL,
  is = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  pch = 0,
  type = "l",
  lwd = 3,
  xlab = "",
  ylab = NULL,
  new = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{the object to be plotted}

\item{is}{For subsetting in space - See \code{\link{subset}}. Can also be
a station value and if provided, the plotting will involve an interpolation
to the same coordinates as defined by \code{is}.}

\item{main}{main title}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{col}{Colour see \code{\link[graphics]{par}}}

\item{lwd}{width of line}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}

\item{legend.show}{if TRUE show legendp}
}
\value{
None
}
\description{
These plot functions are S3 methods for esd objects, based on \code{plot}.
}
\details{
\code{plot.trajectory} plots the number of events per year.
}
\examples{
data(imilast.M03)
plot(imilast.M03)

}
\seealso{
\code{\link{plot}}
}
