% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.xval}
\alias{plot.xval}
\title{plot cross-validation}
\usage{
\method{plot}{xval}(x, ..., new = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{input object to be shown in plot}

\item{new}{a boolean; if TRUE plot in new window}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
plot cross-validation
}
