% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegressionStats.R
\name{plotRegressionStats}
\alias{plotRegressionStats}
\title{Visualize Error and Regression Metrics}
\usage{
plotRegressionStats(metrics, xlim_err = NULL, xlim_rsq = c(0, 1.05),
  left = TRUE, add = FALSE, ...)
}
\arguments{
\item{metrics}{\code{data.frame}. Error and regression metrics as returned by 
\code{\link{evaluate}}.}

\item{xlim_err, xlim_rsq}{\code{numeric}. x-axis range for the lower (errors; 
see '...') and upper (\eqn{R^2}) panel.}

\item{left}{\code{logical}, defaults to \code{TRUE}. Determines position of 
the created plot in a multi-panel figure (i.e., if \code{FALSE}, the left 
y-axis is removed).}

\item{add}{\code{logical}, defaults to \code{FALSE}. Determines whether the 
created plot is added to an already existing multi-panel figure.}

\item{...}{Additional argument passed to \code{\link{round}}. If 'xlim_err' 
is missing, the lower axis range is calculated automatically.}
}
\value{
A \code{trellis} object.
}
\description{
\strong{ESD} standard method for visualizing calculated error and regression 
metrics.
}
\examples{
\dontrun{
## evaluate performance of EOT-based spatial downscaling
metrics <- evaluate(albGIMMS, albMODIS, size = 2)

## visualize results
p <- plotRegressionStats(metrics, xlim_err = c(-.012, 0.052), xlim_rsq = c(0.75, 1.05))
p
}

}
\author{
Florian Detsch
}
