% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{precip.Pr}
\alias{precip.Pr}
\title{Various formulas, equations and transforms.}
\usage{
precip.Pr(x, x0 = 10)
}
\arguments{
\item{x}{a data object}

\item{x0}{a threshold value}
}
\value{
A probability
}
\description{
\code{precip.Pr}: rough estimate of the probability of more than x0 of rain
based on an exponential distribution.
}
\seealso{
C.C.eq precip.vul t2m.vul precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
