% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{precip.rv}
\alias{precip.rv}
\title{Various formulas, equations and transforms.}
\usage{
precip.rv(x, tau = 10)
}
\arguments{
\item{x}{a data object}

\item{tau}{time scale (years)}
}
\value{
An estimate of the return value for the precipitation
}
\description{
\code{precip.rv}: a rough estimate of the return value for precipitation
under the assumption that it is exponentially distributed. Gives apprximate
answers for low return levels (less than 20 years). Advantage, can be
predicted given wet-day mean and frequency.
}
\seealso{
C.C.eq precip.vul t2m.vul precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
