% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{precip.vul}
\alias{precip.vul}
\title{Various formulas, equations and transforms.}
\usage{
precip.vul(x)
}
\arguments{
\item{x}{a data object}
}
\value{
an index for the vulerability to precipitation
}
\description{
\code{precip.vul}: an index for the vulerability to precipitation defined
as wetmean(x)/wetfreq(x). High when the mean intensity is high and/or the
frequency is low (it rains seldom, but when it rains, it really pours down).
}
\seealso{
C.C.eq t2m.vul precip.rv precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
