% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ds}
\alias{predict.ds}
\alias{predict.ds.eof}
\alias{predict.ds.comb}
\alias{predict.mvr}
\alias{predict.cca}
\alias{project.ds}
\title{Prediction based on DS or CCA model}
\usage{
\method{predict}{ds}(x, newdata = NULL, ..., addnoise = FALSE,
  n = 100, verbose = FALSE)
}
\arguments{
\item{x}{A ds object}

\item{newdata}{An eof object containing the new data sets on which the prediction is made.}

\item{addnoise}{If TRUE, will add an attribute called "noise" to the ouput based on WG}

\item{n}{Number of runs to be generated, used only if addnoise is set to TRUE}
}
\value{
Predicted ds values.
}
\description{
Apply an empirical-statistical downscaling model to new data
}
\details{
\code{predict} is similar to the predict function in R

\code{project} returns projection of climate
}
\examples{

# Get predictor
## Get reanalysis
X <- t2m.DNMI(lon=c(-40,50),lat=c(40,75))
## Get Gcm output
Y <- t2m.NorESM.M(lon=c(-40,50),lat=c(40,75))
## Combine
XY <- combine(X,Y)
# Compute common eof for January
ceof <- EOF(XY,it='jan')
# Get predictand
data(Oslo)
# Do the downscaling
ds <- DS(Oslo,ceof)
# Plot ds results
plot(ds)
# Do the prediction based on the calibration (or the fitted values)
ds.pre <- predict(ds)
# Plot predicted results based on ds object
plot(ds.pre)
# Display the attribute "aspect"
attr(ds.pre, "aspect")
## Extract the projected results
plot(project.ds(ds))

}
\seealso{
\code{\link{DS}}
}
