% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictorpatterns.R
\name{predictorpatterns}
\alias{predictorpatterns}
\title{Extract predictor patterns from a dse-object and apply a common EOF evaluation}
\usage{
predictorpatterns(x, ip = 1, plot = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{A DSensemble object.}

\item{plot}{Plot intermediate results if TRUE.}

\item{verbose}{TRUE for checking and debugging the functions.}
}
\value{
A common EOF object of predictor patterns for different models
}
\description{
Extract predictor patterns from a dse-object and apply a common EOF evaluation
}
\examples{

\dontrun{
}

}
\keyword{manip}
