% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessMODIS.R
\name{preprocessMODIS}
\alias{preprocessMODIS}
\title{MODIS Preprocessing Chain}
\usage{
preprocessMODIS(dsn = getwd(), product = c("MOD13Q1", "MYD13Q1"),
  ext = NULL, cores = 1L, ...)
}
\arguments{
\item{dsn}{\code{character}. Target folder for file output.}

\item{product}{\code{character}, defaults to the Terra (\code{"MOD13Q1"}) and
Aqua-MODIS (\code{"MYD13Q1"}) 250-m NDVI products. See also
\code{\link{getProduct}}.}

\item{ext}{\code{Extent}, or any object from which an \code{Extent} can be
extracted, see \code{\link[raster]{crop}}.}

\item{cores}{\code{integer}. Number of cores for parallel processing.}

\item{...}{Additional arguments passed to \code{\link{whittaker.raster}}
(except for 'vi' and 'outDirPath').}
}
\value{
A preprocessed \code{Raster*} object.
}
\description{
MODIS preprocessing chain including layer clipping, two-fold quality-control,
fortnightly composite creation, and subsequent combination of Terra and Aqua
imagery.
}
\author{
Florian Detsch
}
\seealso{
\code{\link[raster]{crop}}, \code{\link{whittaker.raster}}.
}

