% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processChain.R
\name{processChain}
\alias{processChain}
\title{MODIS and GIMMS Preprocessing Chain}
\usage{
processChain(type = c("MODIS", "GIMMS"), ...)
}
\arguments{
\item{type}{\code{character}. Currently available options are "MODIS"
(default) and "GIMMS".}

\item{...}{Additional arguments passed to the underlying preprocessing
functions for MODIS (\code{\link{processChainMODIS}}) and GIMMS
\code{\link{processChainGIMMS}} depending on 'type'.}
}
\value{
A preprocessed \code{Raster*} object.
}
\description{
This is one of the core functions of the \strong{exploratories} package as it
invokes the preprocessing chains of MODIS or GIMMS in preparation for the
application of the EOT algorithm.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{processChainMODIS}}, \code{\link{processChainGIMMS}}
}

