% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processChainMODIS.R
\name{processChainMODIS}
\alias{processChainMODIS}
\title{MODIS Preprocessing Chain}
\usage{
processChainMODIS(dsn = getwd(), product = c("MOD13Q1", "MYD13Q1"),
  ext = NULL, cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{dsn}{\code{character}. Target folder for file output.}

\item{ext}{\code{Extent}, or any object from which an \code{Extent} can be
extracted, see \code{\link[raster]{crop}}.}

\item{cores}{\code{integer}. Number of cores for parallel processing.}

\item{...}{Additional arguments passed to \code{\link{runGdal}}.}
}
\value{
A preprocessed \code{Raster*} object.
}
\description{
MODIS preprocessing chain including layer clipping, two-fold quality-control,
fortnightly composite creation, and subsequent combination of Terra and Aqua
imagery.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{runGdal}}, \code{\link[raster]{crop}}.
}

