% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattributes.R
\name{provenance}
\alias{provenance}
\title{A function that extracts the history/provenance of an object. Eg. the list
of steps in a chain of analysis.}
\usage{
provenance(x, what = "call")
}
\arguments{
\item{x}{any esd-object}

\item{what}{What to return from the history attribute}
}
\description{
A function that extracts the history/provenance of an object. Eg. the list
of steps in a chain of analysis.
}
\examples{
data(Oslo)
provenance(Oslo)

}
\seealso{
history.stamp
}
