% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.test.R
\name{qp.test}
\alias{qp.test}
\title{quantile-quantile plot}
\usage{
qp.test(x, p = c(seq(0.1, 0.95, 0.05), 0.97, 0.98, 0.99),
  threshold = 1, ...)
}
\arguments{
\item{x}{input object}

\item{p}{percentiles to show in plot, a numeric vector}

\item{threshold}{threshold used only if input is precipitation data - only values above threshold are included in analysis}
}
\description{
quantile-quantile plot
}
