% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.thredds.met.no.R
\name{radar}
\alias{radar}
\title{Function to read radar data from thredds.met.no}
\usage{
radar(
  url = "https://thredds.met.no/thredds/catalog/remotesensingradaraccr/",
  lons = c(9.5, 11.5),
  lats = c(59, 61),
  param = "lwe_precipitation_rate",
  FUN = "sum",
  it = 2010:2019,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{url}{URL for the data on thredds.met.no}

\item{lons}{Longitude selection - if NULL read all}

\item{lats}{Latitude selection - if NULL read all}

\item{param}{Variable name}

\item{FUN}{Function for daily aggregation. =NULL gives raw data}

\item{it}{Intex time - the years to select}

\item{verbose}{write out diagnostics}

\item{plot}{plot the results while reading.}
}
\description{
Function to read radar data from thredds.met.no
}
\examples{
Z <- radar(lons = c(5,15), lats = c(55,60), it=2010)
map(Z)

y <- station.thredds(stid=18700,param='precip')
x <- regrid(Z,is=y)
plot(combine.stations(subset(y,it=x),x),new=FALSE)
plot(as.monthly(combine.stations(subset(y,it=x),x),FUN='sum'),new=FALSE)

}
\seealso{
station.thredds, meta.thredds
}
