% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.best.track.R
\name{read.best.track}
\alias{read.best.track}
\title{Read tropical storm trajectory data from \url{http://www.usno.navy.mil/NOOC/nmfc-ph/RSS/jtwc/best_tracks/}}
\usage{

  read.best.track(url = "http://www.usno.navy.mil/NOOC/nmfc-ph/RSS/jtwc/best_tracks/",
  domain = "io", start = 1945, end = 2014, n = 20,
  verbose = TRUE)
}
\arguments{
\item{url}{url}

\item{start}{first year}

\item{end}{last year}

\item{n}{length to which trajectories are interpolated}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{domain:}{'io'='Northern Indian Ocean',
         'sh'='Southern Hemisphere',
'wp'='Northwestern Pacific'}
}
\description{
Read tropical storm trajectory data from \url{http://www.usno.navy.mil/NOOC/nmfc-ph/RSS/jtwc/best_tracks/}
}
