% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.ESGF.R
\name{retrieve.ESGF}
\alias{retrieve.ESGF}
\alias{meta.ESGF}
\title{Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api
meta.ESGF returns a data.frame with the model metadata and the OpenDAP URL that 
can be used with retrieve.ESGF.}
\usage{
\method{retrieve}{ESGF}(im = 1, meta = NULL, verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{param}{Name of parameter}

\item{url}{The base URL for ESGF}

\item{CMIP}{Which generation of CMIP}

\item{MIP}{Name of activity}

\item{inst}{Institution ID}

\item{model}{Name of climate model}

\item{scen}{Name of scenario}

\item{ID}{The name of the run}

\item{type}{The table ID}

\item{grid}{Grid ID}

\item{version}{Version}

\item{period}{The time segment of the file}
}
\value{
A "zoo" "field" object with additional attributes used for further
processing.
}
\description{
Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api
meta.ESGF returns a data.frame with the model metadata and the OpenDAP URL that 
can be used with retrieve.ESGF.
}
\examples{
\dontrun{
meta <- meta.ESGF()
X <- retrieve.ESGF(im=2,lon=c(-30,40),lat=c(50,70),meta=meta)
map(X)
}
}
\seealso{
retrieve
}
