% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sametimescale.R
\name{sametimescale}
\alias{sametimescale}
\title{Function to ensure that station y has the same time scale as X}
\usage{
sametimescale(y, X, FUN = "mean", verbose = FALSE)
}
\arguments{
\item{y}{an input object, e.g., of class \code{station} or \code{field}}

\item{X}{a second input object, e.g., of class \code{station} or \code{field}}

\item{FUN}{a function}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
input object y aggregated to the same time scale as X
}
\description{
Function to ensure that station y has the same time scale as X
}
