% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Advanced scatter plots}
\usage{
scatter(x, y, type = "heat", verbose = FALSE, ...)
}
\arguments{
\item{x}{x-variable}

\item{y}{y-variable}

\item{type}{Type of scatter cplot \code{c('heat','sunflower','hexbin'}}

\item{verbose}{TRUE for diagnosing the internals of the function.}
}
\value{
Graphics and visualisation only.
}
\description{
Various functions that display bi-variate data in different ways. The
default presents the number of points falling into pixles of a 2D grid.
}
\details{
The 'heat' type produces a heat map type display whereas 'sunflower' and
'hexbin' present alternative infographics
}
\examples{

scatter(rnorm(10000),rnorm(10000),dig=1,log=TRUE)
scatter(rnorm(10000),rnorm(10000),type='sunflower',petalsize=7,dx=NULL,dy=NULL,
                              xgrid=NULL,ygrid=NULL,xlim=NULL,ylim=NULL,
                              xlab=NULL,ylab=NULL,main=NULL,leg=TRUE,rotate=TRUE,
                              alpha=0.6,leg.loc=2,new=TRUE, verbose=FALSE)
scatter(rnorm(10000),rnorm(10000),type='hexbin',new=TRUE,Nmax=NULL,
                           dx=NULL,dy=NULL,xgrid=NULL,ygrid=NULL,
                           xlim=NULL,ylim=NULL,
                           xlab=NULL,ylab=NULL,main=NULL,
                           leg=TRUE,col='blue',border='white',
                           colmap='heat.colors',
                           scale.col=TRUE,scale.size=FALSE, verbose=FALSE)

}
\seealso{
\code{\link{vis}},\code{\link{map}},\code{\link{plot}}
}
\author{
Kajsa M. Parding and Rasmus E. Benestad
}
\keyword{graphics}
\keyword{infographics}
