% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasevol.R
\name{seasevol}
\alias{seasevol}
\title{Visualise the seasonal evolution of a daily time series}
\usage{
seasevol(x, nv = 25, verbose = FALSE, ...)
}
\arguments{
\item{x}{as \code{station} object with daily data}

\item{nv}{number of steps in color scale}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Visualise the daily values of time series as a color scale on a plot
with the julian day on the y-axis and year on the x-axis.
}
\examples{
data(ferder)
seasevol(ferder)

}
