% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.abb.R
\name{season.abb}
\alias{season.abb}
\title{Season abbreviation}
\usage{
\method{season}{abb}()
}
\value{
a list with season abbreviations and their corresponding months: list("annual"=1:12, "djf"=c(12,1,2),...)
}
\description{
Season abbreviation
}
