% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{season.default}
\alias{season.default}
\alias{season}
\title{Conversion to esd objects.}
\usage{
\method{season}{default}(x, format = "character")
}
\arguments{
\item{x}{an object of, e.g., class 'station', 'field', or 'zoo', or a date}

\item{format}{for season, set the format of the output 'character' or 'numeric'}
}
\value{
a numeric or character
}
\description{
Used to estimate Dec-Feb, Mar-May, Jun-Aug, and Sep-Nov statistics
Manipulate the zoo-object by shifting the year/chonology so that
zoo thinks the year defined as December-November is January-December.
}
\details{
\code{season} return the seasons associated with the data.
}
\examples{
data(bjornholt)
year(bjornholt)
month(bjornholt)
day(bjornholt)
season(bjornholt)
season(bjornholt, format="numeric")

}
\seealso{
year month
}
\keyword{utilities}
