% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.trajectory.R
\name{season.trajectory}
\alias{season.trajectory}
\alias{anomaly.trajectory}
\alias{polyfit.trajectory}
\alias{count.trajectory}
\alias{param.trajectory}
\alias{sort.trajectory}
\alias{anomaly2trajectory}
\alias{fnlon}
\alias{trajectory2density}
\title{Functions to process and analyse storm trajectories}
\usage{
\method{season}{trajectory}(x, format = "character", verbose = FALSE)
}
\arguments{
\item{x}{A trajectory object}

\item{verbose}{if TRUE print information about progress}

\item{\dots}{Other arguments}

\item{type}{type of anomaly: 'first' gives you the spatial anomaly with
regards to the first time step of the trajectory and 'mean' centers the
trajectories around the mean longitude and latitude}

\item{param}{parameters to calculate anomaly of}
}
\value{
A trajecory object
}
\description{
Functions to process and analyse storm trajectories
}
\examples{

# Load trajectory example data
data('imilast.M03')
# Calculate anomaly
x <- anomaly(imilast.M03)
# Show maps of original trajectories and spatial anomalies
map(imilast.M03, new=FALSE)
map(x, new=FALSE)
# Transform trajectory anomalies back to regular trajectories
y <- anomaly2trajectory(x)
# Print longitudes of first trajectory
imilast.M03[1,1:12]
x[1,1:12]
y[1,12]

# Fit polynomial to trajectories
y <- polyfit.trajectory(x)
# Show coefficients of first trajectory
print(attr(y,"coefs")[,1])
# Plot original trajectory and polynomial fit
ilon <- colnames(x)=="lon"
ilat <- colnames(x)=="lat"
plot(x[1,ilon], x[1,ilat], col="black", pch=1)
lines(y[1,ilon], y[1,ilat], col="blue", lty=2)

}
\author{
Kajsa M. Parding
}
