% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select from meta data base}
\usage{
select(stid = NULL, param = NULL, lon = NULL, lat = NULL,
  alt = NULL, cntr = NULL, ...)
}
\description{
Function that searches the meta data base for the requested station data
Search priority: ID, name, coordinates, altitude, country,...
Can return several matches
}
