% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senorge.thredds.R
\name{senorge}
\alias{senorge}
\alias{senorge,}
\alias{retrieve.senorge}
\title{Function to read seNorge daily temperature and precipitation data from thredds.met.no}
\usage{
senorge(
  url = "https://thredds.met.no/thredds/catalog/senorge/seNorge_2018/Archive",
  param = "rr",
  lon = c(9.5, 11.5),
  lat = c(59, 61),
  it = 2010:2019,
  dt = 50,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{url}{URL for the data on thredds.met.no}

\item{param}{Variable name (rr = precipitation, tg = temperature)}

\item{lon}{Longitude selection (=NULL reads all)}

\item{lat}{Latitude selection (=NULL reads all)}

\item{it}{Index time - a range of dates or years to select (e.g., it=as.Date(c("2010-01-01","2010-01-31")))}

\item{dt}{Number of time steps to access at a time (default: 50). 
A smaller value can be set when requesting a large spatial domain as the server 
doesn't like to deal with large data amounts.}

\item{verbose}{write out diagnostics}

\item{plot}{plot the results while reading.}

\item{FUN}{Function for daily aggregation. =NULL gives raw data}

\item{ncfile}{netCDF file to access (path to a file or url)}

\item{path}{path to ncfile (use only if ncfile doesn't contain the full path)}
}
\value{
A "zoo" "station" object with additional attributes used for further
processing.
}
\description{
The seNorge data set offers gridded fields of meteorological variables at a resolution of 1x1km. 
The 'senorge' function reads the seNorge data directly from the metno thredds server. 
As the resolution is very high, it may be necessary to limit the spatial and/or temporal scope 
(see input options 'it', 'lon', and 'lat') as you will otherwise run out of memory.
}
\examples{
\dontrun{
it <- as.Date(c("1962-12-01","1962-12-31"))
lon <- c(-2,15)
lat <- c(55,63)
rr <- senorge(param="rr", lon=lon, lat=lat, it=it, verbose=TRUE)
map(rr, FUN="mean")
}
}
\seealso{
meta.thredds, station.thredds
}
