% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showmaps.R
\name{showmaps}
\alias{showmaps}
\title{combined maps with northern and southern stereographic projections respectively.}
\usage{
showmaps(
  x,
  FUN = "mean",
  colbar = list(pal = "bwr", show = FALSE),
  nnum = 11,
  ...
)
}
\arguments{
\item{x}{a 'field' object}

\item{FUN}{a function, see \code{\link{aggregate.zoo}}}

\item{colbar}{is a list - see \code{\link{map}}}

\item{nnum}{how many numbers to be shown along the colour bar}
}
\value{
list with map output
}
\description{
\code{as.annual} and \code{annual} aggregates time series into annual values (e.g. means).
}
\details{
\code{as.monthly} aggregates time series into monthly values (e.g. means).

\code{as.daily} aggregates time series into daily values (e.g. means).

\code{showmaps} presents a montage of two maps with stereographic projections, one for the northa nd one for the south.
}
\examples{
t2m <- t2m.NCEP()
period <- paste(range(year(t2m)),collapse=' - ')
attr(t2m,'sub') <- paste0('Average over: ',period,'; source: NCEP')
showmaps(t2m,FUN='mean',colbar=list(breaks=seq(-60,20,by=5),show=FALSE))

attr(t2m,'sub') <- paste0('Variability (stdv) over: ',period,'; source: NCEP')
showmaps(t2m,FUN='sd')

attr(t2m,'sub') <- paste0('Trend over: ',period,'; source: NCEP')
showmaps(t2m,FUN='trend',colbar=list(breaks=seq(-1,1,by=0.1),show=FALSE))
}
\seealso{
map
}
\keyword{utilities}
