% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softattr.R
\name{softattr}
\alias{softattr}
\title{Get names of attributes}
\usage{
softattr(x, ignore = NULL)
}
\arguments{
\item{x}{input object with attributes}

\item{ignore}{names to ignore}
}
\value{
character string or vector with names of attributes
}
\description{
Get names of attributes
}
\seealso{
attrcp
}
