% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{sort.station}
\alias{sort.station}
\title{Routine for sorting the order of station series.}
\usage{
\method{sort}{station}(x, decreasing = TRUE, ..., is = NULL)
}
\description{
Routine for sorting the order of station series.
}
