% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere2cartesian.R
\name{sphere2cartesian}
\alias{sphere2cartesian}
\title{Convert rotated spherical coordinates back to Cartesian coordinates}
\usage{
sphere2cartesian(X, Y, Z, lonR, latR, verbose = FALSE)
}
\arguments{
\item{X}{A numeric vector of X coordinates, where X is the radial distance}

\item{Y}{A numeric vector of Y coordinates, where Y is the polar angle}

\item{Z}{A numeric vector of Z coordinates, where Z is the azimuthal angle}

\item{lonR}{Center longitude of viewing angle}

\item{latR}{Center latitude of viewing angle}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
A list(Lon=longitudes, Lat=latitudes)
}
\description{
Convert rotated spherical coordinates back to Cartesian coordinates
}
