% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_contour.R
\name{spherical_contour}
\alias{spherical_contour}
\title{Function to plot contours on spherical grid}
\usage{
spherical_contour(
  field,
  FUN = "mean",
  lonR = NULL,
  latR = NULL,
  nx = 100,
  ny = 100,
  breaks = NULL,
  zlim = NULL,
  nlevels = 5,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  add = FALSE,
  col = "grey70",
  drawlabels = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{field}{A field object or a matrix with longitude and latitude attributes which is the output of esd::map.field}

\item{lonR}{Center longitude of viewing angle}

\item{latR}{Center latitude of viewing angle}

\item{nx}{Length of output grid along x-direction. Default: nx=100}

\item{ny}{Length of output grid along y-direction. Default: ny=100}

\item{breaks}{Contour levels. If the argument 'breaks' is not defined, breaks = pretty(zlim, n=nlevels)}

\item{zlim}{Range of contours. If the arguments 'breaks' and 'zlim' are not defined, zlim = range(field, na.rm=TRUE)}

\item{nlevels}{Number of contour levels. If the arguments 'breaks' and 'nlevels' are not defined, nlevels=5}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{main}{main title}

\item{add}{a boolean; If TRUE, add contour lines to an existing plot. If FALSE, create new plot with contour lines}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{lat}{A numeric vector of latitudes}
}
\description{
Function to plot contours on spherical grid
}
