% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTindex.R
\docType{methods}
\name{SSTindex}
\alias{SSTindex}
\alias{SSTindex,character-method}
\alias{SSTindex,missing-method}
\title{Get Sea Surface Temperature Indices}
\usage{
\S4method{SSTindex}{character}(index = "dmi", quiet = FALSE, ...)

\S4method{SSTindex}{missing}()
}
\arguments{
\item{index}{\code{character}, defaults to \code{"dmi"} (Dipole Mode Index). 
SST index under investigation, run \code{SSTindex()} for a full list of 
currently available options.}

\item{quiet}{\code{logical}, defaults to \code{FALSE}. Determines whether 
status messages and warnings are suppressed.}

\item{...}{Additional arguments passed to \code{\link{download.file}} (except 
for 'destfile' and 'quiet'), ignored if 'index' is missing.}
}
\value{
If 'index' is missing, a \code{list} of available indices divided 
by region of interest (Pacific, Atlantic, Indian), else a 2-column 
\code{data.frame} with SST index values per time step.
}
\description{
Retrieve selected surface ocean indices provided by the Ocean Observations 
Panel for Climate (OOPC). If an active internet connection is available, the 
latest online version of the respective index is used instead of the 
built-in static data sets (end 27 February 2017).
}
\examples{
## list available indices
SSTindex()

## nino 3.4 region
oni <- SSTindex(index = "nino34", quiet = TRUE)
plot(oni[, 1], oni[, 2], type = "l", xlab = "Time", ylab = "ONI",
     main = expression(bold("Oceanic Ni" * tilde(n) * "o Index") ~ 
                       "(Ni" * tilde(n) * "o 3.4 region)"))

}
\references{
Detsch F, Otte I, Appelhans T, Hemp A, Nauss T (2016) Seasonal and long-term 
vegetation dynamics from 1-km GIMMS-based NDVI time series at Mt. 
Kilimanjaro, Tanzania. \emph{Remote Sensing of Environment} \strong{178}, 
70-83, \url{http://dx.doi.org/10.1016/j.rse.2016.03.007}.

Otte I, Detsch F, Mwangomo E, Hemp A, Appelhans T, Nauss T (2017) 
Multidecadal Trends and Interannual Variability of Rainfall as Observed from 
Five Lowland Stations at Mt. Kilimanjaro, Tanzania. \emph{Journal of 
Hydrometeorology} \strong{18}, 349-361, 
\url{http://dx.doi.org/10.1175/JHM-D-16-0062.1}.
}
\seealso{
Ocean Observations Panel for Climate (2017) The State of the Ocean Climate. 
Available online: \url{http://stateoftheocean.osmc.noaa.gov/all/}.
}
