% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.GHCND.R
\name{station.GHCND}
\alias{station.GHCND}
\title{Retrieve station record from a given data source.}
\usage{
\method{station}{GHCND}(
  x = NULL,
  cntr = NULL,
  param = NULL,
  lon = NULL,
  lat = NULL,
  url =
    "https://www.ncei.noaa.gov/data/global-historical-climatology-network-daily/access",
  sep = ",",
  verbose = FALSE
)
}
\arguments{
\item{x}{a data.frame with GHCND metadata}

\item{cntr}{A string or a vector of strings of the full name of the country:
Select the stations from a specified country or a set of countries.}

\item{param}{Parameter or element type or variable identifier. There are
several core parameters or elements as well as a number of additional
parameters. The parameters or elements are: precip = Precipitation (mm) tas,
tavg = 2m-surface temperature (in degrees Celcius) tmax, tasmax = Maximum
temperature (in degrees Celcius) tmin, tasmin = Minimum temperature (in
degrees Celcius)}

\item{lon}{Numeric value of longitude (in decimal degrees East) for the
reference point (e.g. weather station) as a vector
containing the range of longitude values in the form of c(lon.min,lon.max)}

\item{lat}{Numeric value of latitudes for the range of latitude values in the
form of c(lat.min,lat.max)}

\item{url}{URL of the GHCND database}

\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{stid}{A string of characters as an identifier of the weather/climate
station.}

\item{alt}{Numeric value of altitude (in meters a.s.l.) used for selection.
Positive value, select all stations above this altitude; for negative
values, select all stations below this latitude.}

\item{plot}{Logical value. If, TRUE provides a plot.}
}
\value{
A time series of "zoo" "station" class with additional attributes
used for further processing.
}
\description{
Retrieve station record from a given data source.
}
\examples{

## It may take a little time reading all the metadata over the Internet
## First show all the locations with available data:
zz <- meta.GHCND()
esd::map(zz)

## Read rain gauge data from Mozambique
meta.GHCND(cntr='Mozambique',verbose=TRUE) -> mz
Y <- station.GHCND(mz,param='precip')
plot(Y)

## Get data from several countries and all available variables:
meta.GHCND(cntr=c('Ghana','Togo','Benin'),verbose=TRUE) -> gz
## Get all available variables: returned as a list object where 
X <- station.GHCND(gz)
print(names(gz))
plot(X$precip)

## Extract data from a geographical region defined by longitude, latitude and altitude:
## (stations higher than 300 m above sea level)
nz <- meta.GHCND(lon=c(5,10),lat=c(60,65),alt=500)
esd::map(nz)

}
\seealso{
[func()] station station.default station.ecad station.nacd
station.narp station.nordklim station.metnod station.metnom station.ghcnd
station.ghcnm station.ghcnm station.sonel station.gloss station.newlyn
station.giss clean.station allgood station.thredds map.station select.station
}
\author{
Rasmus Benestad
}
\keyword{station.GHCND}
