% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesfromthredds.R
\name{station.s_enda}
\alias{station.s_enda}
\title{A function that provides reads S-ENDA data from thredds.}
\usage{
\method{station}{s_enda}(
  stid = 18700,
  param = "precip",
  freq = "day",
  FUN = "default",
  start.precip = 7,
  ...,
  caturl = "https://thredds.met.no/thredds/dodsC/met.no/observations/surface/",
  verbose = FALSE
)
}
\arguments{
\item{stid}{station ID or the S-ENDA metadata object}

\item{param}{parameter ('precip'|'t2m')}

\item{freq}{frequency of the data}

\item{FUN}{function for aggregating data (e.g. sub-daily for generating daily data)}

\item{start.precip}{Hour of daily start of the accumulated rainfall}

\item{caturl}{url of the thredds ctalogue containing the files}

\item{verbose}{if true promt test lines}
}
\description{
A function that provides reads S-ENDA data from thredds.
}
\examples{

x <- station.s_enda()
x <- station.s_enda(param='t2m')
## Using the S-ENDA metadata (it takes some time to fetch themetadata...)
meta <- meta.s_enda()
x <- station.s_enda(meta[1:2,])
}
\seealso{
retrieve list_thredds meta.s_enda
}
