% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.thredds.R
\name{station.thredds}
\alias{station.thredds}
\alias{meta.thredds}
\title{Read daily station data of the Norwegian Meteorological Institute from thredds netCDF
using OpenDAP}
\usage{
\method{station}{thredds}(
  param = "t2m",
  is = NULL,
  stid = NULL,
  loc = NULL,
  lon = NULL,
  lat = NULL,
  it = NULL,
  alt = NULL,
  cntr = NULL,
  start.year.before = NULL,
  end.year.after = NULL,
  nmin = NULL,
  verbose = FALSE,
  onebyone = FALSE,
  ...
)
}
\arguments{
\item{param}{The element to read c('t2m','tmax','tmin','precip','slp','sd','fx','fg','dd')}

\item{is}{Index space to select station (list)}

\item{stid}{Station ID to select}

\item{loc}{Name of location to select}

\item{lon}{Range of longitudes to select}

\item{lat}{Range of latitudes to select}

\item{it}{Range of times to select}

\item{alt}{Range of altitudes to select of stations above (positive) or below (negative) a threshold}

\item{start.year.before}{Select stations with record starting before a given year}

\item{end.year.after}{Select stations with record ending after a given year}

\item{nmin}{Select stations with minimum number of valid data points}

\item{verbose}{If TRUE print information about progress.}

\item{onebyone}{If many stations, select them first individually and then combine}

\item{...}{Other arguments.}

\item{cntrs}{Countries to select}
}
\value{
A station object.
}
\description{
\code{station.thredds} is a wrapper that uses retrieve.station combined with information about the
files stored on Thredds. The analysis can also be applied to either EOFs or fields.
}
\details{
\code{meta.thredds} retrieves meta data from Thredds.
}
\examples{
## Get the daily minimum temperature for Oslo-Blindern (station ID 18700)
tmin <- station.thredds(param='tmax',stid=18700)

meta <- meta.thredds(param='precip')
precip <- station.thredds(meta[1,])

}
\seealso{
retrieve.station, station, radar
}
