% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station2field.R
\name{station2field}
\alias{station2field}
\title{Transform \code{station} to \code{field}}
\usage{
station2field(x, verbose = FALSE)
}
\arguments{
\item{x}{a \code{station} object}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Function for converting station data to field data vie the computation of PCAs
gridding to EOFs and then transforming the EOFs to field object.
}
\seealso{
pca2eof eof2field as.field
}
