% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station2sphere.R
\name{station2sphere}
\alias{station2sphere}
\title{Map station data on a spherical projection}
\usage{
station2sphere(
  x,
  FUN = "mean",
  lonR = NULL,
  latR = NULL,
  axiR = 0,
  stereographic = FALSE,
  mask_horizon = TRUE,
  xlim = NULL,
  ylim = NULL,
  gridlines = TRUE,
  cex = 0.2,
  cex.axis = 1,
  cex.lab = 1,
  cex.main = 1.5,
  pch = 21,
  colbar = list(pal = "t2m", col = NULL, rev = FALSE, n = 10, breaks = NULL, type = "p",
    cex = 2, h = 0.6, v = 1, pos = 0.1, show = TRUE),
  new = TRUE,
  add = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{an input object of class 'station'}

\item{FUN}{a function, see \code{\link{aggregate.zoo}}. Default: "mean"}

\item{lonR}{Center longitude of viewing angle}

\item{latR}{Center latitude of viewing angle}

\item{stereographic}{If TRUE use a stereographic projection, otherwise use rotated Cartesian coordinates}

\item{mask_horizon}{if TRUE mask areas beyond the horizon so that only one hemisphere can be shown at once. This is only 
relevant to the stereographic projection which can expand beyond the hemisphere and show the whole globe
(with a distorted perspective far from the center)}

\item{xlim}{range of map in the longitude direction}

\item{ylim}{range of map in the latitude direction}

\item{cex}{magnification factor for point symbols}

\item{cex.axis}{magnification factor for axis text}

\item{cex.lab}{magnification factor for label text}

\item{cex.main}{magnification factor for main title}

\item{pch}{point symbol}

\item{colbar}{List containing colorbar specifications: 
list(pal='t2m',rev=FALSE,n=10,col=NULL,breaks=NULL,type="p",cex=2,h=0.6,v=1,pos=0.1,show=TRUE).
See \link{col.bar} for details.}

\item{new}{If new create new plot window}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional argument}
}
\value{
A visualization of the data on spherical coordinates and a data frame
}
\description{
The function sphere is used inside map.station to plot station data on a spherical projection
}
\seealso{
map.station
}
\keyword{utilities}
