% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.dsensemble}
\alias{subset.dsensemble}
\alias{subset}
\alias{subset.station}
\alias{subset.eof}
\alias{subset.pca}
\alias{subset.cca}
\alias{subset.events}
\alias{subset.trajectory}
\alias{subset.trend}
\alias{subset.corfield}
\alias{subset.ds}
\alias{subset.comb}
\alias{subset.field}
\alias{subset.spell}
\alias{subset.zoo}
\alias{station.subset}
\alias{subset.stationmeta}
\alias{subset.matrix}
\alias{subset.pattern}
\alias{subset.Default}
\alias{subset.mvcomb}
\title{Subsetting esd objects}
\usage{
\method{subset}{dsensemble}(
  x,
  ...,
  it = NULL,
  is = NULL,
  ip = NULL,
  im = NULL,
  ensemble.aggregate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{'dsensemble' object (output of \code{DSensemble}) from which the subset is taken}

\item{\dots}{additional arguments}

\item{it}{A list or data.frame providing time index, e.g. a range of years like c(1979,2010), a season ('djf'), or a month ('dec' or 'december').}

\item{is}{a list or data.frame providing a space index, e.g., station record or a lon(gitude) and lat(itude) range. If NULL include all.}

\item{ip}{a numerical or numerical vector with indices of the principle components to be included. If NULL include all.}

\item{im}{a numerical or numerical vector with indices of the ensemble members to be included. If NULL include all.}

\item{ensemble.aggregate}{If TRUE, call \code{subset.dsensemble.multi} if
appropriate.}

\item{verbose}{If TRUE, print out diagnosics}
}
\value{
An object of the same class as the input object
}
\description{
The subset method tries to be 'intelligent', and if the list has no names,
then the list contains two vectors of length 2, then this is interpreted as
a region, e.g. argument \code{is = list(
c(lon.min,lon.max),c(lat.min,lat.max) )}. If, on the other hand, \code{is =
list( lon=1:50,lat=55:65 )}, then the function picks the longitudes and
latitudes which match these. This makes it flexible so that one can pick any
irregular sequence.
}
\seealso{
matchdate sort.station
}
\author{
R.E. Benestad, A. Mezghani and K. Parding
}
\keyword{utilities}
