% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dsensemble}
\alias{summary.dsensemble}
\title{Show summary of objects}
\usage{
\method{summary}{dsensemble}(x, years = seq(1990, 2090, by = 20))
}
\arguments{
\item{x}{an object of type 'DSensemble'}

\item{years}{A set of years for which to produce summary statistics}
}
\value{
A matrix containing summary statistics
}
\description{
Produce a summary table
}
\examples{

data("dse.Oslo")
summary(dse.Oslo)

}
\seealso{
summary.station summary.ncdf4
}
