% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunspots.R
\name{sunspots}
\alias{sunspots}
\title{sunspots}
\usage{
sunspots(url = "https://www.sidc.be/SILSO/DATA/SN_d_tot_V2.0.txt")
}
\arguments{
\item{url}{}
}
\value{
The call returns a zoo object
}
\description{
Reads up-to-date sunspot data from theInternet.
}
\examples{

## S3 method for class 'station'
R <- sunspots()
plot(R)

}
