% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{t2m.NCEP}
\alias{t2m.NCEP}
\alias{geoborders}
\alias{etopo5}
\alias{station.meta}
\alias{NACD}
\alias{NARP}
\alias{Oslo}
\alias{Svalbard}
\alias{bjornholt}
\alias{ferder}
\alias{vardo}
\alias{nordklim.data}
\alias{t2m.NORDKLIM}
\alias{precip.NORDKLIM}
\alias{eof.precip.ERAINT}
\alias{eof.slp.NCEP}
\alias{eof.sst.NCEP}
\alias{eof.t2m.NCEP}
\alias{eof.slp.DNMI}
\alias{eof.sst.DNMI}
\alias{eof.t2m.DNMI}
\alias{eof.t2m.NorESM.M}
\alias{precip.ERAINT}
\alias{slp.NCEP}
\alias{sst.NCEP}
\alias{slp.DNMI}
\alias{sst.DNMI}
\alias{t2m.DNMI}
\alias{t2m.NorESM.M}
\alias{slp.ERA5}
\alias{NINO3.4}
\alias{NAOI}
\alias{sunspots}
\alias{arctic.t2m.cmip5}
\alias{arctic.t2m.cmip3}
\alias{global.t2m.cmip3}
\alias{global.t2m.gcm}
\alias{global.t2m.cmip5}
\alias{scandinavia.t2m.cmip3}
\alias{scandinavia.t2m.cmip5}
\alias{IPCC.AR5.Table.9.A.1}
\alias{gcmresolution}
\alias{dse.ferder}
\alias{dse.Svalbard}
\alias{dse.Oslo}
\alias{imilast.M03}
\alias{storms}
\alias{mu.eq.f.tx}
\title{Sample data}
\usage{
t2m.NCEP(lon = NULL, lat = NULL, anomaly = FALSE, latest = FALSE,
  url = "ftp://ftp.cdc.noaa.gov/Datasets/ncep.reanalysis.derived/surface/air.mon.mean.nc",
  verbose = FALSE)
}
\arguments{
\item{lon}{longitude range c(lin.min,lon.max)}

\item{lat}{latitude range}

\item{anomaly}{TRUE: return anomaly}

\item{latest}{if TRUE check if a newer version can be downloaded}

\item{url}{source of data}

\item{verbose}{if TRUE print progress}
}
\value{
Numeric vectors/matrices with a set of attributes describing the
data.
}
\description{
The object \code{geoborders} contains data on coastlines and borders, used
in the methods \code{\link{map}}.
}
\details{
\code{etopo5} is a 5-minute gridded elevation data set provided by NOAA as described in 
"Data Announcement 88-MGG-02, Digital relief of the Surface of the Earth. NOAA, National Geophysical Data Center, Boulder, Colorado, 1988."

The object \code{station.meta} contains meta data for various sources of station data
(NACD, NARP, NORDKLIM, ECAD, GHCN, and METNO) used in the methods \code{\link{station}}.
\code{ISO03} contains country codes.

\code{NACD}, \code{NARP}, and \code{nordklim.data} contain staion data from Northern Europe from
the North Atlantic Climatological Dataset (NACD),
the Nordic Arctic Research Programme (NARP),
and the NORDKLIM project, respectively,
which are used in the methods \code{\link{station}}.

The temperature and precipitation data from NORDKLIM are also avaialble
as \code{station} objects in \code{t2m.NORDKLIM} and \code{precip.NORDKLIM}.

\code{Oslo} and \code{Svalbard} are historic reconstructions of temperature from Oslo (1837-2018)
and Svalbard (1898-2018) provided by Dr. Nordli, Met Norway.

\code{ferder} and \code{vardo} are time series of temperature and \code{bjornholt} of precipitation from stations in Norway,
downloaded from the Met Norway data archive. Ferder and Bjornholt are located near Olso
while Vardo is located in Northern Norway.

Samples of re-analyses are provided but to reduce the data size they have been stored as 20 EOFS (30 for precipitation)
To reconstruct the fields, use the functions \code{precip.ERAINT}, \code{slp.NCEP}, \code{t2m.NCEP}, \code{sst.NCEP},
\code{slp.DNMI}, \code{sst.DNMI}, and \code{t2m.DNMI}. 
The data compression facilitated by the EOFs can provide 80-90\% of the variance in the data.
ESD uses the large-scale features from these reanalyses, and hence this information loss may be acceptable for downscaling work.

A reduced copy of the NorESM (M RCP 4.5) is also provided for the examples
and demonstrations on how the downscaling can be implemented. Note:
downscaling for end-users should never be based on one GCM simulation alone.

\code{slp.ERA5} provides a small sample of 6-hourly ERA5 sea level pressure data from the North Atlantic
from September 30 to October 10 of 2016, used to test the \code{\link{CCI}} method.

Some data sets (\code{NINO3.4}, \code{NAOI} ) come with a 'frozen' version in the package,
but there are also functions that read the most recent version of these indeces from the Internet
with functions \code{\link{NINO3.4}} and \code{\link{NAO}}.

The python script \code{py.script} is used in the function \code{\link{ERA5.CDS}} to download ERA5 data from the Climate Data Store.
}
\examples{

data(Oslo)
year <- as.numeric( format(index(Oslo), '\%Y') ) 
plot(aggregate(Oslo, by=year,FUN='mean', na.rm = FALSE), new=FALSE)

data(etopo5)
z <- subset(etopo5,is=list(lon=c(-10,30),lat=c(40,60)))
map(z, new=FALSE)

}
\seealso{
\code{\link{aggregate.area}} \code{\link{as.4seasons}},
\code{\link{annual}}
}
\author{
R.E. Benestad
}
\keyword{datasets}
