% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{t2m.Pr}
\alias{t2m.Pr}
\title{Various formulas, equations and transforms.}
\usage{
t2m.Pr(x, x0 = 10, na.rm = TRUE)
}
\arguments{
\item{x}{a data object}

\item{x0}{a threshold value}

\item{na.rm}{See \code{\link{mean}}.}
}
\value{
A probability
}
\description{
\code{t2m.Pr}: rough estimate of the probability of more than x0 of rain
based on a normal distribution.
}
\seealso{
C.C.eq precip.vul t2m.vul precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
