% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clausiusclaperyon.R
\name{t2m.vul}
\alias{t2m.vul}
\title{Various formulas, equations and transforms.}
\usage{
t2m.vul(x, x0 = 30, is = 1)
}
\arguments{
\item{x}{a data object}

\item{x0}{a threshold value}

\item{is}{which of the spell results [1,2]}
}
\value{
an index for the vulnerability to temperature
}
\description{
\code{t2m.vul}: an index for the vulnerability to temperature defined as the
mean spell length for heat waves with temperatures exceeding 30C (default).
}
\seealso{
t2m.vul precip.rv precip.Pr t2m.Pr NE
}
\author{
R. Benestad
}
\keyword{parameter,element,Clausius-Clapeyron}
