% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frostAPI.R
\name{test.API}
\alias{test.API}
\title{Test function for \code{frostID} which retrieves data from the frost.met.no}
\usage{
test.API(keyfile = "~/.FrostAPI.key")
}
\arguments{
\item{keyfile}{location of API key}
}
\description{
Test function for \code{frostID} which retrieves data from the frost.met.no
}
