% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS.R, R/DS.local.R
\name{test.ds.field}
\alias{test.ds.field}
\title{Test function for DS.field}
\usage{
test.ds.field(x, verbose = FALSE)

test.ds.field(x, verbose = FALSE)
}
\arguments{
\item{x}{a \code{ds} \code{eof} object}

\item{verbose}{a boolean; if TRUE print information on progress}
}
\value{
a \code{field} object with the difference between the original field and the field
reconstructed from the independent downscaled principle components from cross-validation

a \code{field} object with the difference between the original field and the field
reconstructed from the independent downscaled principle components from cross-validation
}
\description{
Test function for DS.field

Test function for DS.field
}
