% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WG.R
\name{test.WG.fwmu.day.precip}
\alias{test.WG.fwmu.day.precip}
\title{This function tests the WG for precipitation:
Quantile-quantile plots of wet-day amounts
Number of wet days}
\usage{
test.WG.fwmu.day.precip(x = NULL, verbose = TRUE)
}
\description{
This function tests the WG for precipitation:
Quantile-quantile plots of wet-day amounts
Number of wet days
}
