% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdensity.R
\name{trackdensity}
\alias{trackdensity}
\title{export}
\usage{
trackdensity(
  lons,
  lats,
  track = NULL,
  dx = NULL,
  dy = NULL,
  radius = 5e+05,
  type = "track",
  verbose = FALSE
)
}
\description{
export
}
