% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.events.R
\name{trackstats}
\alias{trackstats}
\title{Calculate trajectory statistics}
\usage{
trackstats(x, verbose = FALSE)
}
\arguments{
\item{x}{an \code{events} object}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
an \code{events} object with statistics describing the trajectories
}
\description{
The function enumerates the trajectories ("trajectory"),
adds time steps and the total number of steps in a trajecotry ("trackcount"),
length of trajectory (in km): "tracklength"),
and if the distance "dx" beween time steps exists the
length of the trajectory from start to end ("tracklength") is also calculated.
}
