% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trajectory.R
\name{trajectory2field}
\alias{trajectory2field}
\alias{as.field.trajectory}
\title{Transform an input object into a \code{field} object}
\usage{
trajectory2field(x, dt = "month", dx = 2, dy = 2, radius = 5e+05,
  it = NULL, is = NULL, verbose = FALSE)
}
\arguments{
\item{x}{a \code{trajectory} object}

\item{dt}{frequency of output: 'month', 'season', 'quarter' (same as 'season') or 'year'}

\item{dx}{resolution in longitude direction (unit: degrees)}

\item{dy}{resolution in latitude direction (unit: degrees)}

\item{radius}{radius within which to look for trajectories for each grid point (unit: m)}

\item{it}{a time index, e.g., a range of years: c(1984,2019)}

\item{is}{a spatial index, e.g., a list with longitude and latitude ranges: list(lon=c(0,45), lat=c(45,70))}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{field} object
}
\description{
Transform a \code{trajectory} object into a \code{field} object by aggregating it in time and space.
}
\seealso{
as.field as.trajectory CCI track.events
}
