% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trajectory.R
\name{trajectory2station}
\alias{trajectory2station}
\title{Transform an input object into a \code{station} object}
\usage{
trajectory2station(
  x,
  it = NULL,
  is = NULL,
  param = NULL,
  FUN = "count",
  longname = NULL,
  unit = NULL,
  loc = NULL
)
}
\arguments{
\item{x}{a \code{trajectory} object}

\item{it}{a time index, e.g., a range of years: c(1984,2019)}

\item{is}{a spatial index, e.g., a list with longitude and latitude ranges: list(lon=c(0,45), lat=c(45,70))}

\item{param}{a characteristic of the trajectories (to see options: \code{colnames(x)})}

\item{FUN}{a function. If 'count' return number of trajectories, otherwise apply \code{FUN} to \code{param}}

\item{longname}{variable name}

\item{unit}{name of unit}

\item{loc}{name of location/region}

\item{a}{\code{station} object}
}
\description{
Transform a \code{trajectory} object into a \code{station} object by aggregating it in time and space.
}
\seealso{
as.station as.station.trajectory
}
