% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.dsensemble.R
\name{trend.dsensemble}
\alias{trend.dsensemble}
\alias{}
\title{trend.dsensemble}
\usage{
\method{trend}{dsensemble}(x, ..., it = NULL, verbose = FALSE, plot = FALSE, eof = FALSE)
}
\arguments{
\item{x}{an object of type 'dsensemble'}

\item{it}{time index (see \code{\link{subset}})}

\item{verbose}{if TRUE print progress}

\item{plot}{if TRUE show results visually on a map}

\item{eof}{if TRUE and if the dsensemble object contains a spatially aggregated pattern, x$eof, 
expand with x$eof into an ensemble of field objects to which the trend analysis is applied. 
If FALSE, expand using the original pattern x$pca into an ensemble of station objects to be aggregated.}
}
\description{
The function \code{trend.dsensemble} is used to calculate trends of dsensemble objects and summarise the results in terms of ensemble statistics: the minimum (min), fifth percentile (q5), median, mean, 95th percentile (q95) and maximum trend (max), as well as the number and fraction of ensemble members with a positive (n.pos, f.pos) and negative (n.neg, f.neg) trend. The results of the trend analysis can be displayed visually with the function \code{trendplot.dsensemble}, which is called from \code{trend.dsensemble} if the argument plot=TRUE.
}
\seealso{
trendplot.dsensemble map.dsensemble aggregate.dsensemble
}
