% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.ncdf4.station.R
\name{update.ncdf4.station}
\alias{update.ncdf4.station}
\title{update.ncdf4.station}
\usage{
\method{update}{ncdf4.station}(x, file, verbose = TRUE)
}
\arguments{
\item{x}{station object}

\item{fname}{file name of the netCDF file with stations to be updated}

\item{vebose}{For diagnostics}
}
\description{
The function adds new days of station data to an existing netCDF with daily station data
and then updates the summary statistics.
}
\author{
R.E. Benestad
}
