% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.transform.R
\name{UTM2LatLon}
\alias{UTM2LatLon}
\title{Coordinate transformations}
\usage{
UTM2LatLon(x, y, zone, southhemi = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{The x coordinates (easting)}

\item{y}{The y coordinates (northing)}

\item{zone}{UTM zone}

\item{southhemi}{if TRUE we are at the southern hemisphere}

\item{verbose}{If TRUE, print out diagnosics}
}
\description{
Transform UTM (Universal Transverse Mercator) coordinates to latitude and longitude
}
\seealso{
LatLon2UTM
}
\author{
K. Tunheim
}
