% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{vec}
\alias{vec}
\title{Draw vectors on maps for esd objects}
\usage{
vec(x, y, it = NULL, a = NULL, r = 1, ix = NULL, iy = NULL,
  new = TRUE, nx = 150, ny = 80, projection = "lonlat",
  lonR = NULL, latR = NULL, axiR = 0, verbose = FALSE,
  length = NULL, ...)
}
\arguments{
\item{x}{the zonal comonents of the vectors}

\item{y}{the meridional comonents of the vectors}

\item{it}{see \code{\link{subset}}}

\item{a}{used  to scale the length of the arrows}

\item{r}{used to make a 3D effect of plotting the arrows up in the air.}

\item{new}{TRUE: create a new graphic device.}

\item{nx}{is the number of vectors/points along the x-axis}

\item{ny}{is the number of vectors/points along the y-axis}

\item{projection}{Projections: c("lonlat","sphere","np","sp") - the latter
gives stereographic views from the North and south poles.}

\item{lonR}{Only for the spherical projection}

\item{latR}{Only for the spherical projection}

\item{axiR}{Only for the spherical projection}

\item{xlim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections.}

\item{ylim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections.}

\item{n}{The number of colour breaks in the color bar}

\item{breaks}{graphics setting - see \code{\link{image}}}

\item{type}{graphics setting - colour shading or contour}

\item{gridlines}{Only for the lon-lat projection}

\item{is}{see \code{\link{subset}}}
}
\description{
Plot arrows for e.g. a wind field or ocean currents.
}
\examples{
\dontrun{
sst <- sst.NCEP()
download.file('ftp://ftp.cdc.noaa.gov/Datasets/ncep.reanalysis.derived/surface/uwnd.mon.mean.nc','ncep_u.nc')
download.file('ftp://ftp.cdc.noaa.gov/Datasets/ncep.reanalysis.derived/surface/vwnd.mon.mean.nc','ncep_v.nc')
u <- retrieve('ncep_u.nc')
v <- retrieve('ncep_v.nc')
## Make a map of SSt and wind vectors:
map(sst)
vec(u,v,a=1.5,length=0.03,new=FALSE)

## Make a fancy map
map(sst,projection='sphere')
vec(u,v,a=0.5,r=1.1,length=0.03,projection='sphere',new=FALSE)
}


}
\seealso{
map
}
