% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\alias{vis.station}
\alias{vis.pca}
\title{InfoGraphics}
\usage{
vis(x, ...)
}
\arguments{
\item{x}{an input object of class 'DSensemble'}

\item{\dots}{additional arguments}

\item{img}{a 'raster' object, or an object that can be coerced to one by 'as.raster', to be used as background}

\item{it}{see \code{\link{subset}}}

\item{col}{color}

\item{n}{number of breaks in color scale}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{vis} shows the annual and seasonal evolution of a time series, similar to \code{\link{seasevol}}.
}
\examples{
data(Oslo)
vis(Oslo)

}
\seealso{
vis.trends wheel cumugram visprob conf graph
diagram scatter plot map
}
