% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.trends.R
\name{vis.trends}
\alias{vis.trends}
\title{Visualise trends for multiple overlapping periods}
\usage{
\method{vis}{trends}(
  x,
  ...,
  unitlabel = "unit",
  varlabel = "",
  is = 1,
  pmax = 0.01,
  minlen = 15,
  lwd = NA,
  vmax = NA,
  new = TRUE,
  show.significance = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{the 'x' argument provides the time series for which the trend
analysis is performed. Only zoo objects are accepted.}

\item{unitlabel}{unit of x.}

\item{varlabel}{name of x.}

\item{is}{spatial index for subsetting station data}

\item{pmax}{maximum p-value of trends marked as significant.}

\item{minlen}{minimum time interval to calculate trends for in units of
years.}

\item{lwd}{width of lines}

\item{vmax}{upper limit of trend scale.}

\item{new}{if TRUE plot in new window}

\item{show.significance}{TRUE to mark statistically significant trends.}

\item{verbose}{TRUE or FALSE.}
}
\description{
Produce a plot showing trends for multiple periods within a time series. The
strength of the trend is represented by the color scale and significant
trends are marked with black borders.
}
\examples{

t <- seq(as.Date("1955-01-01"),as.Date("2004-12-31"),by=1)
x <- zoo(sample(seq(-30,30,1e-1),length(t),rep=TRUE),order.by=t)
vis.trends(x, show.significance=FALSE, new=FALSE)

data(Oslo)
vis.trends(Oslo, unitlabel="oC", varlabel = "Temperature",
  pmax = 1e-2, minlen = 40, new=FALSE)
vis.trends(subset(Oslo,it='jja'), unitlabel="oC",
  varlabel = "Temperature JJA",
  pmax = 1e-3, vmax=0.5, minlen = 40, new=FALSE)
vis.trends(subset(Oslo,it='mam'), unitlabel="oC",
  varlabel = "Temperature MAM",
  pmax = 1e-3, vmax=0.5, minlen = 40, new=FALSE)

}
\author{
Kajsa Parding
}
\keyword{trend}
