% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visprob.R
\name{visprob}
\alias{visprob}
\alias{visprob.station}
\alias{visprob.station.precip}
\title{InfoGraphics}
\usage{
visprob(x, ...)
}
\arguments{
\item{x}{an input object of class 'station'}

\item{\dots}{additional arguments}

\item{y}{an input object of class 'station'}

\item{dy}{relative width of lines}

\item{threshold}{threshold defining a precipitation event}

\item{breaks}{breaks in historgram}

\item{pdf}{a boolean; if TRUE add pdfs estimated from wet-day mean}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{visprob} displays the probability density function (PDF) of a precipitation time series (\code{x}) for each year.
If only one time series is provided (\code{y}=NULL), the color of the PDFs represent the year.
If a second time series is provided, the color scale shows the annual mean value of \code{y}.
}
\examples{
data(bjornholt)
visprob(bjornholt)

}
\seealso{
wheel cumugram climvar graph conf vis diagram scatter plot map
}
