% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel.R
\name{wheel}
\alias{wheel}
\title{InfoGraphics}
\usage{
wheel(x, ...)
}
\arguments{
\item{x}{an input object of class 'station' or 'spell'}

\item{\dots}{additional arguments}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{wheel} shows the seasonal cycle with different colors for different years. If the input is a \code{station} object, \code{wheel} calls the function \code{\link{wheel.station}} and if the input is \code{spell} object, it calls \code{\link{wheel.spell}}
}
\examples{
data(bjornholt)
wheel(bjornholt, new=FALSE)

}
\seealso{
wheel.station wheel.spell graph visprob conf vis diagram cumugram scatter plot map
}
