% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel.R
\name{wheel.spell}
\alias{wheel.spell}
\title{S3 method of \code{wheel} for \code{'spell'} objects}
\usage{
\method{wheel}{spell}(x, y = NULL, new = TRUE, lwd = 2, col = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{an input object of class 'spell'}

\item{y}{an input object of class 'spell'}

\item{new}{if new create new graphic device}

\item{lwd}{relative line width}

\item{col}{color of line}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}
}
\description{
\code{wheel} shows the seasonal cycle with different colors for different years.
}
\examples{

data(bjornholt)
wetspell <- spell(bjornholt, threshold=.1)
wheel(wetspell, new=FALSE)

}
\seealso{
wheel wheel.station
}
