% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel.R
\name{wheel.station}
\alias{wheel.station}
\title{S3 method of \code{wheel} for \code{'station'} objects}
\usage{
\method{wheel}{station}(
  x,
  y = NULL,
  new = TRUE,
  lwd = 2,
  col = NULL,
  type = NULL,
  bg = "grey90",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{an input object of class 'station'}

\item{y}{an input object of class 'station'}

\item{new}{if new create new graphic device}

\item{lwd}{relative line width}

\item{col}{color of line}

\item{type}{'spiky' or 'flowy'}

\item{bg}{background color}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{\dots}{additional arguments}
}
\description{
\code{wheel} shows the seasonal cycle with different colors for different years.
}
\examples{
data(bjornholt)
wheel(bjornholt, new=FALSE)

}
\seealso{
wheel wheel.spell graph visprob conf vis diagram cumugram scatter plot map
}
