% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2ncdf.R
\name{write2ncdf4.eof}
\alias{write2ncdf4.eof}
\title{Unfinished function that doesn't do anything.}
\usage{
\method{write2ncdf4}{eof}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{input object of class 'dsensemble'}

\item{\dots}{additional arguments}

\item{verbose}{if TRUE print progress}
}
\description{
Unfinished function that doesn't do anything.
}
\seealso{
write2ncdf4
}
