% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2ncdf.R
\name{write2ncdf4.list}
\alias{write2ncdf4.list}
\alias{write2ncdf4.field}
\title{Saves climate data as netCDF.}
\usage{
\method{write2ncdf4}{list}(
  x,
  ...,
  file = "field.nc",
  prec = "short",
  scale = 0.1,
  offset = NULL,
  torg = "1970-01-01",
  missval = -999,
  verbose = FALSE
)
}
\arguments{
\item{x}{data object}

\item{\dots}{additional arguments}

\item{file}{file name}

\item{prec}{Precision: see \code{\link[ncdf4]{ncvar_def}}}

\item{scale}{Sets the atttribute 'scale_factor' which is used to scale
(multiply) the values stored (to save space may be represented as 'short').}

\item{offset}{Sets the attribute 'add_offset' which is added to the values
stored (to save space may be represented as 'short').}

\item{torg}{Time origin}

\item{missval}{Missing value: see \code{\link[ncdf4]{ncvar_def}}}

\item{verbose}{if TRUE print progress}
}
\value{
None
}
\description{
Method to save data as netCDF, making sure to include the data
structure and meta-data (attributes). The code tries to follow the netCDf
'CF' convention. The method is built on the \code{ncdf4} package.
}
\seealso{
write2ncdf4
}
\keyword{ncdf4}
\keyword{netcdf}
\keyword{save}
