% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2ncdf.R
\name{write2ncdf4}
\alias{write2ncdf4}
\alias{write2ncdf4.station}
\alias{write2ncdf4.field}
\alias{write2ncdf4.list}
\alias{write2ncdf4.eof}
\alias{write2ncdf4.pca}
\alias{write2ncdf4.dsensemble}
\title{Saves climate data as netCDF.}
\usage{
write2ncdf4(x, ...)
}
\arguments{
\item{x}{data object}

\item{file}{file name}

\item{prec}{Precision: see \code{\link[ncdf4]{ncvar_def}}}

\item{missval}{Missing value: see \code{\link[ncdf4]{ncvar_def}}}

\item{offset}{Sets the attribute 'add_offset' which is added to the values
stored (to save space may be represented as 'short').}

\item{scale}{Sets the atttribute 'scale_factor' which is used to scale
(multiply) the values stored (to save space may be represented as 'short').}

\item{torg}{Time origin}

\item{verbose}{TRUE - clutter the screen.}
}
\value{
A "zoo" "field" object with additional attributes used for further
processing.
}
\description{
Method to save station data as netCDF, making sure to include the data
structure and meta-data (attributes). The code tries to follow the netCDf
'CF' convention. The method is built on the \code{\link{ncdf4}} package.
}
\examples{

nacd <- station(src='nacd')
X <- annual(nacd)
write2ncdf4(X,file='test.nc')

}
\seealso{
\code{\link{test.retrieve.ncdf4}}.
}
\author{
R.E. Benestad
}
\keyword{ncdf4}
\keyword{netcdf}
\keyword{save}
