% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattributes.R
\name{ylab}
\alias{ylab}
\title{Create a label for plots}
\usage{
ylab(x)
}
\arguments{
\item{x}{an input object}
}
\value{
a character string
}
\description{
Using the attributes of an object, put together a character string with 
the variable name and unit to be used as label in plots and maps.
}
