% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.area.R
\name{aggregate.area}
\alias{aggregate.area}
\title{aggregate}
\usage{
\method{aggregate}{area}(x, ..., is = NULL, it = NULL, FUN = "sum",
  na.rm = TRUE, smallx = FALSE, verbose = FALSE, a = 6378,
  threshold = NULL)
}
\arguments{
\item{x}{A \code{\link{station}} object}

\item{is}{A list or data.frame providing space index, see \code{\link{subset}}}

\item{it}{A list or data.frame providing time index, see \code{\link{subset}}}

\item{FUN}{A function, e.g., 'sum' or 'mean'}

\item{na.rm}{a boolean; if TRUE ignore NA, see \code{\link{mean}}}

\item{smallx}{a boolean defaulting to FALSE}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{a}{radius of earth (unit: km)}

\item{threshold}{threshold to be used if FUN is 'area','exceedance', or 'lessthan'}
}
\value{
The call returns a station object
}
\description{
The aggregation functions are based on the S3 method for \code{zoo} objects,
but takes care of extra house keeping, such as attributes with meta data.
}
\details{
\code{aggregate.area} is used for aggregating spatial statistics, such as
the global mean or the global area of some phenomenon.
}
\examples{

## S3 method for class 'station'
data(Svalbard)
x <- aggregate(Svalbard, month, FUN='mean', na.rm=TRUE)
plot(x)

## S3 method for class 'field'
slp <- slp.DNMI()
y <- aggregate(slp, year, FUN='mean', na.rm=FALSE)

## Aggregate area
w <- aggregate.area(y)
plot(w)

}
\seealso{
aggregate aggregate.size
}
\author{
R.E. Benestad
}
\keyword{utilities}
