% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.annual.R
\name{annual}
\alias{annual}
\alias{as.annual}
\alias{as.annual.default}
\alias{as.annual.numeric}
\alias{as.annual.integer}
\alias{as.annual.yearqtr}
\alias{as.annual.station}
\alias{as.annual.spell}
\alias{annual.zoo}
\alias{annual.default}
\alias{annual.dsensemble}
\alias{annual.station}
\alias{annual.spell}
\alias{annual.field}
\alias{annual.eof}
\alias{as.monthly}
\alias{as.monthly.default}
\alias{as.monthly.station}
\alias{as.monthly.field}
\alias{as.4seasons}
\alias{as.4seasons.default}
\alias{as.4seasons.day}
\alias{as.4seasons.station}
\alias{as.4seasons.spell}
\alias{as.4seasons.field}
\alias{as.4seasons.dsensemble}
\alias{as.seasons}
\alias{as.daily}
\alias{as.OctMar}
\title{Conversion to esd objects.}
\usage{
annual(x, ...)
}
\arguments{
\item{x}{an input object of, e.g., class 'station', 'field'}

\item{\dots}{additional argument}

\item{FUN}{a function, see \code{\link{aggregate.zoo}}}

\item{nmin}{Minimum number of data points (e.g. days or months) with valid
data accepted for annual estimate. NULL demands complete years.}

\item{start}{makes it possible to estimate annual aggregated statistics that start into the year. 
E.g. start='Sep' starts the year on September 1st. Other allowed formats are start='MM-DD'.}

\item{format}{'numeric' or 'character'}

\item{na.rm}{a boolean; if TRUE, ignore NA - see see \code{\link{mean}}}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
Same class as x
}
\description{
\code{as.annual} and \code{annual} aggregates time series into annual values (e.g. means).
}
\details{
\code{as.monthly} aggregates time series into monthly values (e.g. means).

\code{as.daily} aggregates time series into daily values (e.g. means).

\code{as.4seasons} aggregates to four seasons ('djf': December-February, 'mam': March-May, 'jja': June-August, 'son': September-November)

\code{as.seasons} aggregates to a user defined season with input arguments 'start' and 'end' giving the dates. To select march to september, use either start='03-01' and end='09-30' or start = 3 and end = 9.

\code{as.OctMar} aggregates to the season October to March, which is the rainy season in parts of Africa.
}
\examples{
data(ferder)
plot(annual(ferder,FUN="min"))
plot(annual(ferder,FUN="IQR",na.rm=TRUE))
plot(annual(ferder))
lines(annual(ferder,start='Jul'))

data(bjornholt)
plot(as.4seasons(bjornholt,threshold=1,FUN="exceedance"))

}
\seealso{
aggregate
}
\keyword{utilities}
