% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.field.R
\name{as.field}
\alias{as.field}
\title{Coerce input to a \code{field} object}
\usage{
as.field(x, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}
}
\value{
a \code{field} object
}
\description{
Transform an input object into the esd class \code{field}. 
\code{as.field} is an S3 method and will redirect to a fitting function depending on the output. 
The way in which the transformation is performed depends on the type of input data.
}
\details{
\code{as.field.events} redirects to \code{\link{events2field}}.
\code{as.field.trajectory} redirects to \code{\link{trajectory2field}}.
}
\examples{
# how to generate a new field object.
year <- sort(rep(1991:2000,12))
month <- rep(1:12,length(1991:2000))
n <-length(year)
lon <- seq(-30,40,by=5); nx <- length(lon)
lat <- seq(40,70,by=5); ny <- length(lat)
# Time dimension should come first, space second.
y <- matrix(rnorm(nx*ny*n),n,nx*ny)
index <- as.Date(paste(year,month,1,sep="-"))
Y <- as.field(y,index=index,lon=lon,lat=lat,param="noise",unit="none")
map(Y)

}
\seealso{
as.field.default as.field.zoo as.field.eof as.field.comb as.field.field as.field.ds as.field.station as.field.events as.field.trajectory as.field.dsensemble.eof as.field.matrix
}
