% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.pca.R
\name{as.pca}
\alias{as.pca}
\title{Coerce input to a \code{station} object}
\usage{
as.pca(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}
}
\value{
a \code{pca} object
}
\description{
Transform an input object into the esd class \code{pca}. 
\code{as.pca} is an S3 method and will redirect to a fitting function depending on the output. 
The way in which the transformation is performed depends on the type of input data.
}
\seealso{
as.pca.ds as.pca.station PCA
}
