% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.dsensemble.pca}
\alias{as.station.dsensemble.pca}
\title{Coerce input into a station object}
\usage{
\method{as.station}{dsensemble.pca}(
  x,
  ...,
  is = NULL,
  ip = NULL,
  nmin.m = 1,
  nmin.t = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{a dsensemble pca object}

\item{\dots}{additional arguments}

\item{is}{A list or data.frame providing space index, e.g. a list of longitude and
latitude range like list(lon=c(0,60), lat=c(35,60)).}

\item{ip}{selection of patterns in PCA or EOF (used for e.g. filtering the data)}

\item{nmin.m}{Threshold used to remove ensemble members with few available values. Ensemble members with valid data for less than nmin.m time steps are removed.}

\item{nmin.t}{Threshold used to remove time steps with few available values. Time steps with valid data for less than nmin.t ensemble members are removed.}

\item{verbose}{if TRUE print progress}
}
\description{
Transform a \code{dsensemble pca} object into a \code{dsensemble station} object
by extracting the results model wise and using the downscaled principle components
to reconstruct station time series.
}
\seealso{
as.station as.station.dsensemble DSensemble
}
