% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station.dsensemble}
\alias{as.station.dsensemble}
\title{Coerce input to a \code{station} object}
\usage{
\method{as.station}{dsensemble}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{the input object of class \code{dsensemble}}

\item{...}{other arguments}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
a \code{station} object or a \code{dsensemble} \code{list} or \code{dsensemble} \code{station} object
}
\description{
Coerce input to a \code{station} object
}
\seealso{
as.station as.station.dsensemble.pca as.station.dsensemble.station DSensemble PCA
}
