% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.station.R
\name{as.station}
\alias{as.station}
\title{Coerce input to a \code{station} object}
\usage{
as.station(x, ...)
}
\arguments{
\item{x}{the input object}

\item{...}{other arguments}
}
\value{
a \code{station} object
}
\description{
Transform an input object into the esd class \code{station}. 
\code{as.station} is an S3 method and will redirect to a fitting function depending on the output. 
The way in which the transformation is performed depends on the type of input data.
}
\seealso{
as.station.zoo as.station.data.frame as.station.ds as.station.pca
}
