% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.station.R
\name{barplot.station}
\alias{barplot.station}
\title{create a barplot}
\usage{
\method{barplot}{station}(
  height,
  ...,
  col.above = "red",
  col.below = "blue",
  threshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{height}{a 'station' object}

\item{\dots}{additional arguments}

\item{col.above}{color for values below threshold. Default: blue.}

\item{threshold}{threshold - midline of plot}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
create a barplot
}
