% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian2sphere.R
\name{cartesian2sphere}
\alias{cartesian2sphere}
\title{Convert Cartesian coordinates to spherical coordinates}
\usage{
cartesian2sphere(Lon, Lat, lonR = NULL, latR = NULL, verbose = FALSE)
}
\arguments{
\item{lonR}{Center longitude of viewing angle}

\item{latR}{Center latitude of viewing angle}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{lon}{A numeric vector of longitudes}

\item{lat}{A numeric vector of latitudes}
}
\description{
Convert Cartesian coordinates to spherical coordinates
}
