% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climvar.R
\name{climvar}
\alias{climvar}
\title{InfoGraphics}
\usage{
climvar(x, FUN = "sd", plot = TRUE, ...)
}
\arguments{
\item{x}{an input object of class 'station'}

\item{FUN}{a function}

\item{\dots}{additional arguments}

\item{it}{A list or data.frame providing time index, e.g. month}

\item{start}{year and month, e.g., '-01-01' to start in january}

\item{prog}{a boolean; if TRUE show prognosis for end of year in cumugram}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{main}{main title}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{climvar} estimates how the variance varies with season
in terms of the inter-annual variability of daily standard deviation
}
\examples{
data(bjornholt)
cumugram(bjornholt)

}
\seealso{
wheel graph visprob conf vis diagram cumugram scatter plot map
}
