% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale.R
\name{col.bar}
\alias{col.bar}
\alias{colbar}
\title{Display a color bar object on an existing plot.}
\usage{
col.bar(breaks, horiz = TRUE, pch = 21, v = 1, h = 1, col = col,
  cex = 2, cex.lab = 0.6, cex.axis = 0.9, type = "r",
  verbose = FALSE, vl = 0.5, border = FALSE, ...)
}
\arguments{
\item{breaks}{A numeric vector of breakpoints for the colours}

\item{horiz}{a boolean; if TRUE add horizontal color bar, else add vertical color bar}

\item{pch}{see \code{\link{par}}}

\item{v}{Vertical space between color bar points}

\item{h}{horizontal space between color bar points}

\item{col}{see \code{\link{par}}}

\item{cex}{see \code{\link{par}}}

\item{cex.lab}{see \code{\link{par}}}

\item{type}{r : rectangular shape , p : for points}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{vl}{a numerical specifying the relative placement of the vertical lines}

\item{border}{a boolean; if TRUE show color bar borders}

\item{\dots}{Additional graphical parameters to be passed on}

\item{pal}{palette: c("bwr",rwb","faint.bwr","faint.rwb","rainbow","gray.colors","heat.colors","terrain.colors","topo.colors","cm.colors","grmg","brbu","budor","budrd","bugr","bugy","buor","buorr","bu","rd","cat","cold","warm")}
}
\description{
Add a color bar or color points into an exisiting plot or map
}
\seealso{
colbar.ini
}
