% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.events.R
\name{count.events}
\alias{count.events}
\title{Aggregate an \code{events} object in time and space}
\usage{
count.events(x, by.trajectory = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{the input object of class \code{events}}

\item{by.trajectory}{a boolean; if TRUE count each cyclone trajectory only once per time step}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{...}{other arguments}
}
\value{
a \code{station} object: number of events per month
}
\description{
Transform an \code{events} object (output of \code{CCI} including information about cyclones or anticyclones) 
into a \code{station} object by calculating the number of events per month.
}
\seealso{
CCI
}
