% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{count}
\alias{count}
\alias{wetfreq}
\alias{wetmean}
\alias{nevents}
\alias{exceedance}
\alias{exceedance.default}
\alias{exceedance.station}
\alias{exceedance.field}
\title{Count statistics}
\usage{
count(x, threshold = 1, fraction = FALSE, ...)
}
\arguments{
\item{x}{station or field object}

\item{threshold}{threshold value}

\item{fraction}{TRUE: divide the number of counts by number of samples}

\item{upper}{upper limit for maximum length - ignore any above this because they are likely erronoeus}

\item{FUN}{function}
}
\value{
Station or field objects
}
\description{
Returns a count of cases with values over s threshold value.
\code{exceedance} estimates statistics (default is mean) for peak-over-threshold, e.g. mean-over-threshold. 
Similar functions nclude: 
\code{nevents} returns the number of events with exceeding values (e.g. the
number of rainy days X > 1 mm/day).
\code{wetfreq} returns the wet-day frequency (a fraction) and \code{wetmean} returns the wet-day mean.
\code{qqgeom} produces a quantile-quantile plot of streak statistics comparing the
empirical quantiles with the distribution function quantiles (see \code{\link[stats]{qgeom}}).
}
\examples{

number of warm days:
data(ferder)
plot(as.seasons(ferder,FUN='count',threshold=20), new=FALSE)

Mild winter days - number of days in the winter season with
above freezing temperatures
try(coldwinterdays(ferder))

Quantile-quantile plot
qqgeom(ferder, treshold=1, pois=TRUE)

}
\seealso{
spell hotsummerdays coldwinterdays nwetdays plot
}
\keyword{utilities}
