% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\alias{diagnose.comb}
\alias{diagnose.eof}
\alias{diagnose.comb.eof}
\alias{diagnose.mvr}
\alias{diagnose.cca}
\alias{diagnose.ds}
\alias{diagnose.station}
\alias{diagnose.distr}
\alias{diagnose.matrix}
\alias{diagnose.dsensemble}
\alias{diagnose.ds.pca}
\title{Diagnose}
\usage{
diagnose(x, ...)
}
\arguments{
\item{x}{data object}

\item{\dots}{additional arguments}

\item{it}{teporal selection - see \code{\link{subset}}}

\item{plot}{if TRUE, plot results}

\item{plot.type}{type of plot}

\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{new}{if TRUE plot in new window}

\item{xlim}{range of x-axis}

\item{xlim}{range of y-axis}

\item{alpha}{factor modifying the opacity alpha; typically in [0,1]}

\item{map.show}{if TRUE show map}

\item{xrange}{longitude range to display in map}

\item{yrange}{latitude range to display in map}

\item{main}{main label in plot}

\item{sub}{smaller label (subtitle) in plot}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{probs}{quantile to display in plot, e.g., probs=0.95 gives a diagnosis
of the 95th percentile of the data.}
}
\value{
A 'diag' object containing test results
}
\description{
Diagnose and examine combined fields, MVR, and CCA results. applies some
tests to check for consistency.
}
\details{
The method \code{diagnose.comb.eof} which estimates the difference in the
mean for the PCs of the calibration data and GCMs over a common period in
addition to the ratio of standard deviations and lag-one autocorrelation.
The x-axis shows the difference in the mean of the segments in the PCs
representing the different data souces, the y-axis shows difference in
standard deviation and the size of the symbols the difference in the
autocorrelation (open symbols if the autocorrelation have different signs).

\code{climvar} estimates the climatological variance, e.g. how the
inter-annual variance varies with seasons.
}
\examples{

t2m <- t2m.DNMI(lon=c(-40,40),lat=c(30,70))
T2m <- t2m.NorESM.M(lon=c(-40,40),lat=c(30,70))
# Combine in time to compute common EOFs:
X <- combine(t2m,T2m)
diagnose(X)

ceof <- EOF(X,it="jan")
plot(diagnose(ceof))

slp <- slp.NCEP(lat=c(-40,40),anomaly=TRUE)
sst <- sst.NCEP(lat=c(-40,40),anomaly=TRUE)
eof.1 <- EOF(slp,it="jan")
eof.2 <- EOF(sst,it="jan")
cca <- CCA(eof.1,eof.2)
diagnose(cca)


}
\author{
R.E. Benestad
}
\keyword{utilities}
