% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{MODIS and GIMMS Download Chain}
\usage{
download(type = c("MODIS", "GIMMS"), ...)
}
\arguments{
\item{type}{\code{character}. Currently available options are "MODIS"
(default) and "GIMMS".}

\item{...}{Additional arguments passed to the underlying download functions
for MODIS (\code{\link{runGdal}}; except for 'product' and 'collection', with
the latter set to \code{"006"}) and GIMMS (\code{\link{downloadGimms}};
except for 'version', which is set to \code{1L}) depending on 'type'.}
}
\value{
A \code{character} vector of local filenames.
}
\description{
This is one of the core functions of the \strong{exploratories} package as it
invokes the product download of MODIS or GIMMS in preparation for
preprocessing and the application of the EOT algorithm.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{runGdal}}, \code{\link{downloadGimms}}.
}

