% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ele2param}
\alias{ele2param}
\alias{esd2ele}
\title{Dictionary and conversion tools between esd element identifier and variables
names and specifications.}
\usage{
ele2param(ele = NULL, src = NULL)
}
\arguments{
\item{src}{A character string for the acronym of the data source. The data
sources are :}

\item{param, ele}{Parameter or element identifier. There are several core
parameters or elements as well as a number of additional parameters. The
parameters or elements are :}
}
\value{
A meta data matrix object with the glossary of the different
variables or element identifiers as originally defined by each data source
}
\description{
Converts between esd element/parameter identifier and variable names from
different data sources.
}
\examples{

# Eg.1 # Display the glossary of paramerters or element identifiers for 'GHCND' data source.
print(ele2param(ele=NULL,src='GHCND'))
# Eg.2 # Display the glossary of parameters or element identifiers for all data sources. 
print(ele2param())
# Eg.3 # Convert mean temperature parameter (param) to esd element (ele).
ele <- esd2ele(param='t2m')
print(ele)


}
\author{
A. Mezghani, MET Norway
}
\keyword{parameter,element}
