% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.events.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate trajectories}
\usage{
enumerate(x, param = "trajectory", verbose = FALSE)
}
\arguments{
\item{x}{an \code{events} object}

\item{param}{parameter name of the trajectory numbers. Default: trajectory.}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
an \code{events} object with statistics describing the trajectories
}
\description{
The function enumerates the trajectories ("trajectory"). It is used in the function \code{"trackstats"} and can be applied to re-enumerate cyclone trajectories in an \code{events} object after selecting a subset using the function \code{"subset"}.
}
