% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_CDS.R
\name{ERA5.CDS}
\alias{ERA5.CDS}
\title{R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to}
\usage{
ERA5.CDS(
  param = "total_precipitation",
  it = 1979:2018,
  varnm = NULL,
  AREA = "['-90','-180','90','180']",
  FNAME = "'ERA5_XXX_YYYY.nc'",
  FUN = "monsum",
  path = "~/Downloads/",
  verbose = TRUE
)
}
\arguments{
\item{param}{variable name in CDS call, e.g. 'total_precipitation', '2m_temperature', 'mean_sea_level_pressure',
'10m_u_component_of_wind', '10m_v_component_of_wind', 'relative_humidity', 'dewpoint_depression', 'snow_depth'}

\item{it}{the years to extract.}

\item{varnm}{variable name for local data file.}

\item{AREA}{the area/region to extract [south,west,north,east]}

\item{FNAME}{the name of the local files for storing the data}

\item{FUN}{the function for CDO to aggregate the data, eg 'monsum', 'daymean',monmean', 'yearsum',
'yearmax', etc. If NULL, then leave the data as they are (e.g. daily data).}

\item{path}{The path where the data are stored. Can be a symbolic link.}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to
}
