% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_CDS.R
\name{ERA5.CDS}
\alias{ERA5.CDS}
\title{R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to}
\usage{
ERA5.CDS(
  param = "total_precipitation",
  it = 1979:2018,
  varnm = "tp",
  AREA = "['-90','-180','90','180']",
  FNAME = "'ERA5_XXX_YYYY.nc'",
  FUN = "monsum",
  path = "~/Downloads/",
  verbose = TRUE
)
}
\arguments{
\item{param}{variable name in CDS call, e.g. 'total_precipitation', '2m_temperature', or 'mean_sea_level_pressure'}

\item{it}{the years to extract.}

\item{varnm}{variable name for local data file.}

\item{AREA}{the area/region to extract [south,west,north,east]}

\item{FNAME}{the name of the local files for storing the data}

\item{FUN}{the function for CDO to aggregate the data, eg 'monsum', 'daymean',monmean', 'yearsum', 'yearmax', etc. If NULL, then leave the data as they are (e.g. daily data).}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to
}
\examples{
ERA5.CDS(X='2m_temperature',varnm='t2m',it=2015:2018,AREA="['50','0','60','10']",
         FUN='daymean')
ERA5.CDS(X='total_precipitation',varnm='tp',it=2018,AREA="['0','50','10','60']",
         FUN='yearsum')
ERA5.CDS(X='mean_sea_level_pressure',varnm='slp',it=2018,AREA="['40','-50','60','30']",
         FUN='monmean')

}
