% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_CDS.R
\name{ERA5.CDS}
\alias{ERA5.CDS}
\title{R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to}
\usage{
ERA5.CDS(
  param = "total_precipitation",
  it = 1979:2018,
  varnm = NULL,
  lon = c(-180, 180),
  lat = c(-90, 90),
  FNAME = "'ERA5_XXX_YYYY.nc'",
  FUN = "monsum",
  path = "~/Downloads/",
  verbose = TRUE
)
}
\arguments{
\item{param}{variable name in CDS call, e.g. 'total_precipitation', '2m_temperature', 'mean_sea_level_pressure',
'10m_u_component_of_wind', '10m_v_component_of_wind', 'relative_humidity', 'dewpoint_depression', 'snow_depth'}

\item{it}{the years to extract.}

\item{varnm}{variable name for local data file.}

\item{lon}{longitude of the area/region to extract.}

\item{lat}{latitude of the area/region to extract.}

\item{FNAME}{the name of the local files for storing the data}

\item{FUN}{the function for CDO to aggregate the data, eg 'monsum', 'daymean',monmean', 'yearsum',
'yearmax', etc. If NULL, then leave the data as they are (e.g. daily data).}

\item{path}{The path where the data are stored. Can be a symbolic link.}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
R-script that downloads daily data from the Copernicus Climate Data Store (CDS) using
the CDS set-up and python scripts through the API. The files will be stored
as netCDF files. This script assumes that CDO and python are installed: 
https://www.unidata.ucar.edu/software/netcdf/workshops/most-recent/third_party/CDO.html.
It only works on Linux platforms...
See https://cds.climate.copernicus.eu/api-how-to
}
\examples{
\dontrun{
ERA5.CDS(param='2m_temperature',varnm='t2m',it=2015:2018,lon=c(0,10), lat=c(50,60),
         FUN='daymean')
ERA5.CDS(param='total_precipitation',varnm='tp',it=2018,lon=c(50,60),lat=c(0,10),
         FUN='yearsum')
ERA5.CDS(param='mean_sea_level_pressure',varnm='slp',it=2018,lon=c(-50,30),lat=c(40,60),
         FUN='monmean')
}
}
