% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Visualize EOT Validation Results}
\usage{
evaluate(pred, resp, by = 24L, times = 10L, size = 5L, ...)
}
\arguments{
\item{pred}{\code{Raster*} object. Predictor data for EOT analysis.}

\item{resp}{\code{Raster*} object. Response data for EOT analysis.}

\item{by}{\code{integer}. Increment of the sequence passed to
\code{\link{seq}}.}

\item{times}{\code{integer}. Determines the number of
\code{\link{evaluateEOT}} runs.}

\item{size}{\code{integer}. Sample size passed to \code{\link{sample}}.}

\item{...}{Additional arguments passed to \code{\link{write.csv}}.}
}
\value{
A \code{data.frame}.
}
\description{
Evaluate performance of EOT-based spatial resampling \emph{n} times and
subsequently visualize derived error and regression metrics.
}
\author{
Florian Detsch
}

