% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.dsensemble.R
\name{expandpca}
\alias{expandpca}
\title{Expand PCA to obtain station data}
\usage{
expandpca(x, it = NULL, FUN = NULL, FUNX = "mean", verbose = FALSE,
  anomaly = FALSE, test = FALSE)
}
\description{
Expand PCA to obtain station data
}
