% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHCND.R
\name{ghcnd.meta}
\alias{ghcnd.meta}
\alias{ghcnd.station}
\title{Fetch station data from GHND and convert into \code{station} objects.}
\usage{
ghcnd.meta(
  param = NULL,
  src = "ghcnd",
  path = "data.GHCND",
  url = "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily",
  save.file = FALSE,
  verbose = TRUE,
  force = TRUE
)
}
\arguments{
\item{url}{lURL of the database}

\item{x}{a 'stationmeta' object}
}
\description{
Transform an input object into the esd class \code{station}. 
\code{ghcnd.meta} reads the inventory and metadata of the stations (slow due to big data volume - requires Internet access).
}
\details{
\code{ghcnd.station} reads the raw data (slow due to big data volume - requires Internet access).
}
\examples{
meta <- ghcnd.meta()
m <- subset(meta,lon=c(24,35),lat=c(-30,-25),verbose=TRUE)
y <- ghcnd.station(subset(m,is=1:10),verbose=TRUE)

}
\seealso{
station as.station
}
