% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHCND.R, R/ghcnd.R
\name{ghcnd.meta}
\alias{ghcnd.meta}
\alias{ghcnd.station}
\alias{ghcnd.data}
\alias{ghcnm.meta}
\alias{ghcnm.data}
\title{Fetch station data from GHND and convert into \code{station} objects.}
\usage{
ghcnd.meta(
  param = NULL,
  src = "ghcnd",
  path = "data.GHCND",
  url = "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily",
  save.file = FALSE,
  verbose = TRUE,
  force = TRUE
)

ghcnd.meta(
  param = NULL,
  src = "ghcnd",
  path = "data.GHCND",
  url = "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily",
  save.file = FALSE,
  verbose = TRUE,
  force = TRUE
)
}
\arguments{
\item{param}{climate variable}

\item{src}{source of data}

\item{path}{path to directory where to save data}

\item{url}{url to data source}

\item{save.file}{a boolean; if TRUE save data or metadata}

\item{verbose}{a boolean; if TRUE print information about progress}

\item{force}{a boolean; if TRUE overwrite old file}

\item{x}{a 'stationmeta' object}
}
\description{
Transform an input object into the esd class \code{station}. 
\code{ghcnd.meta} reads the inventory and metadata of the stations (slow due to big data volume - requires Internet access).

\code{ghcnd.meta} and \code{ghncm.meta} read and organize metadata of daily (ghcnd) and monthly (ghcnm) GHCN data.
\code{ghncd.data} and \code{ghcnd.data} read daily and monthly mean GHCN data from NOAA (\url{ftp.ncdc.noaa.gov}).
}
\details{
\code{ghcnd.station} reads the raw data (slow due to big data volume - requires Internet access).
}
\examples{
meta <- ghcnd.meta()
m <- subset(meta,lon=c(24,35),lat=c(-30,-25),verbose=TRUE)
y <- ghcnd.station(subset(m,is=1:10),verbose=TRUE)
}
\seealso{
station as.station
}
