% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHCND.R
\name{GHCND.meta}
\alias{GHCND.meta}
\alias{GHCND.station}
\title{Fetch station data from GHND and convert into \code{station} objects.}
\usage{
GHCND.meta(
  is = NULL,
 
    url = "https://www.ncei.noaa.gov/data/global-historical-climatology-network-daily/doc/ghcnd-stations.txt"
)
}
\arguments{
\item{url}{lURL of the database}

\item{x}{a 'stationmeta' object}
}
\description{
Transform an input object into the esd class \code{station}. 
\code{GHCND.meta} reads the inventory and metadata of the stations (slow due to big data volume - requires Internet access).
}
\details{
\code{GHCND.station} reads the raw data (slow due to big data volume - requires Internet access).
}
\examples{
meta <- ghcnd.meta(is=list(cntr=c('Mozambique','Zimbabwe','Malawi')))
m <- subset(meta,lon=c(24,35),lat=c(-30,-25),verbose=TRUE)
y <- ghcnd.station(subset(m,is=1:10),verbose=TRUE)

}
\seealso{
station as.station ghcnd.meta ghcnd.station
}
