% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmap.R
\name{gridmap}
\alias{gridmap}
\alias{gridmap.default}
\alias{gridmap.station}
\alias{gridmap.pca}
\alias{gridstations}
\title{Creates a gridded map}
\usage{
gridmap(
  Y,
  FUN = "mean",
  colbar = list(pal = "t2m"),
  project = "lonlat",
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  verbose = FALSE,
  plot = FALSE,
  new = TRUE
)
}
\arguments{
\item{Y}{A station object or a PCA object.}

\item{FUN}{A function or name of a function, e.g, "mean" or "trend"}

\item{colbar}{A list specifying the color bar, e.g., list(col="precip",
breaks=seq(1,10), rev=FALSE)}

\item{project}{projection: "lonlat" or "sphere"}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{zlim}{range of color axis}

\item{verbose}{if TRUE print information about progress}

\item{plot}{if TRUE display results as plots}

\item{new}{if TRUE plot in new window}
}
\description{
A function that uses \code{LatticeKrieg} and elevation data to grid station
based data and present a map.
}
\examples{

data("precip.NORDKLIM")
precip.gp <- gridmap(precip.NORDKLIM, plot=TRUE)
map(precip.gp)
}
