% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{history.stamp}
\alias{history.stamp}
\title{This function adds a stamp in the history of x
with 'sys.call', 'date()', and 'src (source)}
\usage{
history.stamp(x = NULL, y = NULL, verbose = FALSE, ...)
}
\description{
This function adds a stamp in the history of x
with 'sys.call', 'date()', and 'src (source)
}
