% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howsimilar.R
\name{howsimilar}
\alias{howsimilar}
\title{howsimilar}
\usage{
howsimilar(x, y, plot = "cor", FUN = NULL, regress = FALSE,
  verbose = FALSE)
}
\arguments{
\item{x}{A \code{\link{field}} object representing the reanalysis}

\item{y}{A \code{\link{station}} object}

\item{plot}{a boolean; if TRUE ignore NA, see \code{\link{mean}}}

\item{regress}{a boolean defaulting to FALSE}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\value{
The call returns a station object
}
\description{
field-station comparison
Extract time series from field objects with the same coordinates as the stations
Estimate correlation, RMSE, max and min for the extracted series. Store as 
attributes in the extracted station object. Use regression to calibrate to adjust for the
fact that a grid-box volume is not the same as a point measurement.
Rasmus E. Benestad
}
\details{
\code{howsimilar} is used for assessing reanalyses against station data.
}
\examples{
test.howsimilar()

}
\seealso{
regrid
}
\author{
R.E. Benestad
}
\keyword{utilities}
