% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.transform.R
\name{LatLon2UTM}
\alias{LatLon2UTM}
\alias{UTM2LatLon}
\title{Coordinate transformations}
\usage{
LatLon2UTM(lat, lon, zone, verbose = FALSE)
}
\arguments{
\item{lat}{A vector containing latitudes (unit: degrees east)}

\item{lon}{A vector containing longitude (unit: degrees north/south)}

\item{zone}{UTM zone}

\item{verbose}{If TRUE, print out diagnosics}

\item{x}{The x coordinates (easting)}

\item{y}{The y coordinates (northing)}
}
\description{
Transform UTM (Universal Transverse Mercator) coordinates to/from latitude and longitude
}
\author{
K. Tunheim
}
