% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd-package.R
\docType{data}
\name{manual}
\alias{manual}
\alias{ABC4ESD}
\alias{esd.tips}
\alias{esd.issues}
\alias{downscaling.about}
\title{Manual for esd}
\description{
Help, assistance and manuals for \code{esd} and empirical-statistical
downsclaing (ESD).
}
\examples{

manual()

}
\keyword{datasets}
