\name{manual}
\alias{manual}
\alias{ABC4ESD}
\alias{esd.tips}
\alias{esd.issues}
\alias{downscaling.about}
\docType{data}
\title{Manual for esd}
\description{
  Help, assistance and manuals for \code{esd} and empirical-statistical
  downsclaing (ESD). 
}
\usage{
manual(url='https://ndownloader.figshare.com/files/2126237',browser='firefox')
ABC4ESD(url='http://www.gvc2.gu.se/ngeo/rcg/edu/esd.pdf',browser='firefox')
esd.tips(url='https://github.com/metno/esd/wiki',browser='firefox')
esd.issues(url='https://github.com/metno/esd/issues',browser='firefox')
downscaling.about(url=paste('http://climatescience.oxfordre.com/view/10.1093/acrefore/',
                            '9780190228620.001.0001/acrefore-9780190228620-e-27',sep=''),
                  browser='firefox')
}
\examples{
manual()
}
\keyword{datasets}
