% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{manual}
\alias{manual}
\alias{ABC4ESD}
\alias{esd.tips}
\alias{esd.issues}
\alias{downscaling.about}
\alias{element.kode}
\title{Help and assistance}
\usage{
manual(url = "https://ndownloader.figshare.com/files/2126237",
  browser = "firefox")
}
\description{
Help and assistance
}
