% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map.default}
\alias{map.matrix}
\alias{map.data.frame}
\alias{map.station}
\alias{map.stationmeta}
\alias{map.stationsummary}
\alias{map.comb}
\alias{map.eof}
\alias{map.ds}
\alias{map.field}
\alias{map.corfield}
\alias{map.cca}
\alias{map.events}
\alias{map.mvr}
\alias{map.pca}
\alias{map.array}
\alias{map.trend}
\alias{map.mvcomb}
\alias{lonlatprojection}
\alias{rotM}
\alias{map2sphere}
\title{Plot maps for esd objects}
\usage{
map(x, ...)
}
\arguments{
\item{x}{the object to be plotted; in \code{rotM}, x holds a vector of
x-coordinates.}

\item{...}{further arguments passed to or from other methods.}

\item{FUN}{The function to be applied on x before mapping (e.g. \code{mean})}

\item{colbar}{The colour scales defined through \code{colscal}. Users can
specify the colour `pal'*ette (`pal'), the number of breaks (`n'), values of
`breaks', and the position of the color bar from the main plot (`pos'). The
`rev' argument, will produce a reversed color bar if set to TRUE. The other
arguments (`type',`h' and `v') are more specific to \code{col.bar} and are
used only if argument `fancy' is set to TRUE (not yet finished). \code{colbar=NULL}
is used if the colourbar is not to be shown. Also use \code{colbar=NULL} to present several 
maps in one figure (e.g. with \code{par(mfcol=c(2,2))}).}

\item{lab}{\code{'default'} to show a lable saying what variable (unit) and time period. 
\code{'simple'} to just use \code{varid(x)}, and \code{'unit'} to show variable and unit. 
Other strings will be used as the label for the plot and \code{NULL} shows no label.}

\item{it}{see \code{\link{subset}}}

\item{is}{see \code{\link{subset}}}

\item{new}{TRUE: create a new graphic device.}

\item{projection}{Projections: c("lonlat","sphere","np","sp") - the latter
gives stereographic views from the North and south poles. Other types of projections 
are also possible based on a wrapper function for \code{oce::mapPlot} 
(e.g. \code{projection="+proj=moll"}) - see the details provided for 
\code{oce::mapPlot} for available projections.}

\item{xlim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections. Default \code{xlim=NULL} is the same as \code{c(-1,+1)} and refers to coordinates
on a unit sphere (radius = 1). If plotting should be limited to a range of longitudes and latitudes
then \code{subset(.)} can be used prior to \code{map}.}

\item{ylim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections. See \code{xlim}.}

\item{n}{The number of colour breaks in the color bar}

\item{breaks}{graphics setting - see \code{\link{image}}}

\item{gridlines}{Only for the lon-lat projection}

\item{lonR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{latR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{axiR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{style}{Only for the spherical projection used by \code{map2sphere} to apply night shade effect. c('plain','night')}

\item{y}{a vector of y coordinates}

\item{z}{a vector of z coordinates}

\item{ip}{Selects which pattern (see \code{\link{EOF}}, \code{\link{CCA}}) to plot}

\item{geography}{TRUE: plot geographical features}

\item{angle}{for hatching}

\item{a}{used in \code{\link{vec}} to scale the length of the arrows}

\item{r}{used in \code{\link{vec}} to make a 3D effect of plotting the arrows up in the air.}

\item{ix}{used to subset points for plotting errors}

\item{iy}{used to subset points for plotting errors}

\item{colorbar}{Show the color bar in the map (default TRUE). If FALSE, the
colorbar is not added into the map (ignored).}

\item{cex}{Size of symbols.}

\item{cex0}{Scaling of symbols if \code{cex} is defined by a variable with
different size for different locations.}

\item{cex.subset}{...}

\item{add.text}{Add abbreviated location names.}

\item{full.names}{Show the full name of the location.}

\item{showall}{Default is set to FALSE}

\item{showaxis}{If set to FALSE, the axis are not displayed in the plot.}

\item{fancy}{If set to true, will use \code{\link{col.bar}} instead of
\code{image} to produce the colour bar}

\item{text}{If TRUE, display text info on the map.The default is set to
FALSE}

\item{show.val}{Display the values of 'x' or 'FUN(x)' on top of the coloured
map.}

\item{legend.shrink}{If set, the size of the color bar is shrinked (values
between 0 and 1)}

\item{land}{if TRUE mask land, else mask ocean}

\item{what}{What to map: ['eof','field] for EOF pattern or the field}

\item{type}{- default c('fill','contour')
recovered from the EOFs.}

\item{fig}{see \code{\link{par}}}

\item{add}{set add=TRUE if adding the map as a subplot}

\item{nbins}{number of bins/colour categories}
}
\value{
A field object
}
\description{
Make map of geophysical data. These plot functions are S3 methods for esd
objects.
}
\examples{

# Select stations in ss and map the geographical location 
# of the selected stations with a zoom on Norway.
ss <- select.station(cntr="NORWAY",param="precip",src="GHCND")
map(ss, col="blue",bg="lightblue",xlim = c(-10,30) , ylim = c(50,70), new=FALSE)

## Get NACD data and map the mean values
y <- station.nacd()
map(y,FUN='mean',colbar=list(pal="t2m",n=10), cex=2, new=FALSE)

# Examples of cyclone maps (map.events)
data(storms)
# Subset cyclones from the start of January 2016 lasting a minimum of 10 time steps 
x <- subset(storms,it=c("2016-01-01","2016-01-07"),ic=list(param="trackcount",pmin=10))
# Map with points and lines showing the cyclone centers and trajectories
map(x, type=c("trajectory","points"), col="blue")
## Map with only the trajectory and start and end points
map(x, type=c("trajectory","start","end"), col="red")
## Map showing the cyclone depth (slp at center) as a color scale (rd = red scale)
map(x, param="pcent", type=c('trajectory','start'), 
    colbar=list(pal="rd", rev=TRUE, breaks=seq(980,1000,2)), 
    alpha=0.9, new=FALSE)

# Example showing two maps in one figure:
\dontrun{
rr <- retrieve('~/data/data.ECAD/rr_ens_mean_0.1deg_reg.nc',it=c(2000,2001))
it <- 195
par(mfcol=c(1,2))
attr(rr,'variable') <- 'precip'
map(rr,FUN='mean',new=FALSE,type='fill',colbar=NULL,lab=paste(range(index(rr)),collapse=' - '))
map(subset(rr,it=it),FUN='mean',new=FALSE,type='fill',colbar=NULL,lab=as.character(index(rr)[it]))
}

## Example: plotting maps with different projections based on the function oce::mapPlot
t2m <- t2m.NCEP()
map(t2m,projection="+proj=moll")
}
\seealso{
map.trajectory vec map.dsensemble

\code{\link{plot.station}}, \code{\link{showmaps}}
}
\keyword{map}
