% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map.default}
\alias{map.matrix}
\alias{map.data.frame}
\alias{map.station}
\alias{map.stationmeta}
\alias{map.stationsummary}
\alias{map.comb}
\alias{map.eof}
\alias{map.ds}
\alias{map.dsensemble}
\alias{map.field}
\alias{map.corfield}
\alias{map.cca}
\alias{map.events}
\alias{map.trajectory}
\alias{map.mvr}
\alias{map.pca}
\alias{map.trend}
\alias{lonlatprojection}
\alias{rotM}
\alias{gridbox}
\alias{map2sphere}
\alias{vec}
\alias{mask}
\title{Plot maps for esd objects}
\usage{
map(x, it = NULL, is = NULL, new = FALSE, ...)
}
\arguments{
\item{x}{the object to be plotted; in \code{rotM}, x holds a vector of
x-coordinates.}

\item{it}{see \code{\link{subset}}}

\item{is}{see \code{\link{subset}}}

\item{new}{TRUE: create a new graphic device.}

\item{...}{further arguments passed to or from other methods.}

\item{FUN}{The function to be applied on x before mapping (e.g. \code{mean})}

\item{colbar}{The colour scales defined through \code{colscal}. Users can
specify the colour `pal'*ette (`pal'), the number of breaks (`n'), values of
`breaks', and the position of the color bar from the main plot (`pos'). The
`rev' argument, will produce a reversed color bar if set to TRUE. The other
arguments (`type',`h' and `v') are more specific to \code{col.bar} and are
used only if argument `fancy' is set to TRUE (not yet finished).}

\item{projection}{Projections: c("lonlat","sphere","np","sp") - the latter
gives stereographic views from the North and south poles.}

\item{xlim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections.}

\item{ylim}{see \code{\link{plot}} - only used for 'lonlat' and 'sphere'
projections.}

\item{n}{The number of colour breaks in the color bar}

\item{breaks}{graphics setting - see \code{\link{image}}}

\item{type}{graphics setting - colour shading or contour}

\item{gridlines}{Only for the lon-lat projection}

\item{lonR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{latR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{axiR}{Only for the spherical projection used by \code{map2sphere} to change viewing angle}

\item{y}{a vector of y coordinates}

\item{z}{a vector of z coordinates}

\item{ip}{Selects which pattern (see \code{\link{EOF}}, \code{\link{CCA}}) to plot}

\item{geography}{TRUE: plot geographical features}

\item{angle}{for hatching}

\item{a}{used in \code{\link{vec}} to scale the length of the arrows}

\item{r}{used in \code{\link{vec}} to make a 3D effect of plotting the arrows up in the air.}

\item{ix}{used to subset points for plotting errors}

\item{iy}{used to subset points for plotting errors}

\item{colorbar}{Show the color bar in the map (default TRUE). If FALSE, the
colorbar is not added into the map (ignored).}

\item{cex}{Size of symbols.}

\item{cex0}{Scaling of symbols if \code{cex} is defined by a variable with
different size for different locations.}

\item{cex.subset}{...}

\item{add.text}{Add abbreviated location names.}

\item{full.names}{Show the full name of the location.}

\item{showall}{Default is set to FALSE}

\item{showaxis}{If set to FALSE, the axis are not displayed in the plot.}

\item{fancy}{If set to true, will use \code{\link{col.bar}} instead of
\code{image} to produce the colour bar}

\item{text}{If TRUE, display text info on the map.The default is set to
FALSE}

\item{show.val}{Display the values of 'x' or 'FUN(x)' on top of the coloured
map.}

\item{legend.shrink}{If set, the size of the color bar is shrinked (values
between 0 and 1)}

\item{land}{if TRUE mask land, else mask ocean}

\item{what}{What to map: ['eof','field] for EOF pattern or the field
recovered from the EOFs.}

\item{fig}{see \code{\link{par}}}

\item{nbins}{number of bins/colour categories}
}
\value{
A field object
}
\description{
Make map of geophysical data. These plot functions are S3 methods for esd
objects.
}
\examples{

# dontrun because of this problem: "Error in plot.new() : figure margins too large"
# Select stations in ss and map the geographical location 
# of the selected stations with a zoom on Norway.
ss <- select.station(cntr="NORWAY",param="precip",src="GHCND")
graphics.off()
map(ss, col="blue",bg="lightblue",xlim = c(-10,30) , ylim = c(50,70), new=FALSE)

## Get NACD data and map the mean values
y <- station.nacd()
map(y,FUN='mean',colbar=list(pal=varid(y),n=10), cex=2, new=FALSE)

# Examples of cyclone maps (map.events)
data(storms)
# Subset cyclones from the start of January 2016 lasting a minimum of 10 time steps 
x <- subset(storms,it=c("2016-01-01","2016-01-07"),ic=list(param="trackcount",pmin=10))
# Map with points and lines showing the cyclone centers and trajectories
map(x, type=c("trajectory","points"), col="blue")
## Map with only the trajectory and start and end points
map(x, type=c("trajectory","start","end"), col="red")
## Map showing the cyclone depth (slp at center) as a color scale (rd = red scale)
map(x, param="pcent", type=c('trajectory','start'), 
    colbar=list(pal="rd", rev=TRUE, breaks=seq(980,1000,2)), 
    alpha=0.9, new=FALSE)

}
\seealso{
\code{\link{plot.station}}
}
\author{
R.E. Benestad
}
\keyword{map}
