% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.trajectory.R
\name{map.trajectory}
\alias{map.trajectory}
\alias{hexbin.trajectory}
\alias{sunflower.trajectory}
\alias{map.pca.trajectory}
\title{Plot trajectory maps}
\usage{
\method{map}{trajectory}(x, it = NULL, is = NULL,
  type = "trajectory", param = NA, projection = "lonlat",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{the trajectory object to be plotted.}

\item{it}{A list or data.frame providing time index, e.g. month}

\item{is}{A list or data.frame providing space index, e.g. station record}

\item{type}{type of map: 'paths' shows trajectories; 'density' shows the
spatial density of the trajectories; 'colors' shows colored trajectories where
the colorscale represents 'param'; 'anomaly' or 'shape' show  only the
longitude and latitude displacement of the trajectories}

\item{projection}{Projections: c("lonlat","sphere","np","sp") - the latter
gives stereographic views from the North and south poles.}

\item{parameter}{to display as a color scale}

\item{col}{color of trajectories}

\item{colmap}{Colour scales, either as an output from \code{\link[grDevices]{rgb}} or a
single character string 'bwr' (blue-white-red) or 'rwb' ('red-white-blue')}

\item{new}{TRUE: create a new graphics device}

\item{xlim}{see \code{\link{plot}} - only used for 'lonlat' projection}

\item{ylim}{see \code{\link{plot}} - only used for 'lonlat' projection}

\item{main}{an overall title for the plot}

\item{lonR}{Only for the spherical projection - see \code{\link{map2sphere}}}

\item{latR}{Only for the spherical projection - see \code{\link{map2sphere}}}

\item{leg}{logical. If TRUE, legend is shown.}

\item{alpha}{factor modifying the opacity alpha; typically in [0,1]}
}
\description{
\code{map.trajectory} is an \code{S3} method for making different types of trajectory maps.
By default, \code{map.trajectory} shows individual trajectories, but the
number density of trajectories can also be visualised by using the argument \code{type='denisty'}.
}
\details{
The functions \code{hexbin.trajectory} and \code{sunflower.trajectory}
produce alternative versions of trajectory density maps.
}
\examples{

# plot storm tracks zoomed in on the north Atlantic and northern Europe
data(imilast.M03)
map.trajectory(imilast.M03,col="blue",alpha=0.1,
          projection='latlon',xlim=c(-60,60),ylim=c(30,90),
          new=FALSE)

# spherical projection
map.trajectory(imilast.M03,col="blue",alpha=0.1,projection='sphere',new=FALSE)

# plot number density for grid boxes of width 2 degrees and height 1 degree
hexbin.trajectory(imilast.M03,xlim=c(-60,60),ylim=c(30,90),dx=2,dy=1,new=FALSE)
sunflower.trajectory(imilast.M03,xlim=c(-60,60),ylim=c(30,90),dx=2,dy=1,new=FALSE)

\dontrun{
# calculate cyclone density, takes a little while
cdens <- as.field(imilast.M03)
map(cdens,new=FALSE)
}

}
\seealso{
\code{\link{map} \link{map.events}}
}
\keyword{map}
\keyword{trajectory}
