% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metno.frost.R
\name{metno.frost.meta.day}
\alias{metno.frost.meta.day}
\alias{metno.frost.meta.month}
\alias{metno.frost.meta.minute}
\title{Download METNO station metadata using frost.met.no}
\usage{
metno.frost.meta.day(
  param = c("t2m", "precip", "tmin", "tmax", "slp", "pon", "pox", "fg", "fx"),
  save2file = FALSE,
  path = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{param}{Vector of parameters}

\item{save2file}{if TRUE, save metadata in a local file}

\item{verbose}{if TRUE, print diagnostics}

\item{\dots}{additional arguments}

\item{url}{The URL to the webpage to request new client credentials at Frost API}
}
\value{
A meta data matrix object for all stations in METNO's collection
that have measured any of the given parameters. Start and end time are included.
}
\description{
Where there are multiple measuring periods registered for the parameter,
only the earliest start time and the latest end time are used.
}
\examples{

\dontrun{
# Fetch metadata for all stations and measuring periods of the t2m parameter
metno.frost.meta.day(param=c('t2m'))
# Fetch metadata for all stations and measuring periods of all available parameters
metno.frost.meta.month()
}

}
\author{
K. Tunheim
}
\keyword{parameter,metadata,metno,norway,frost}
