% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadCRUT4.R
\name{NASAgiss}
\alias{NASAgiss}
\title{Download GISS Sea Surface Temperature data from NASA}
\usage{
NASAgiss(
  url = "http://data.giss.nasa.gov/gistemp/tabledata_v3/GLB.Ts+dSST.txt",
  plot = FALSE
)
}
\arguments{
\item{url}{url}

\item{plot}{a boolean; if TRUE show resutls in plot}
}
\description{
Download GISS Sea Surface Temperature data from NASA
}
