% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.trajectory.R
\name{PCA.trajectory}
\alias{PCA.trajectory}
\alias{plot.pca.trajectory}
\title{Principle component analysis of trajectory objects.}
\usage{
\method{PCA}{trajectory}(X, ..., neofs = 20, param = c("lon", "lat"),
  anomaly = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a 'trajectory' object}

\item{\dots}{additional arguments}

\item{neofs}{number of EOF patterns to include}

\item{param}{parameters to include in principle component analysis.}

\item{anomaly}{logical. If TRUE, subtract the first latitude/longitude from
each trajectory.}

\item{verbose}{TRUE - clutter the screen with messages}
}
\description{
Computes principal component analysis for trajectory data, e.g., storm
tracks.  Add some reference and details about the method.  The PCA is based
on \code{\link{svd}}.
}
\examples{
# Simple EOF for annual mean SST:
data(imilast.M03)
x <- subset(imilast.M03,is=list(lon=c(-20,20),lat=c(50,70)))
# PCA of longitude and latitude
pca <- PCA(x,param=c('lon','lat'))
plot(pca)
map(pca,projection='latlon')

# latitude only
pca <- PCA(x,param=c('lat'))
plot(pca)

}
\keyword{multivariate}
\keyword{spatial}
\keyword{ts}
