% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessMODIS.R
\docType{methods}
\name{preprocessMODIS}
\alias{preprocessMODIS}
\alias{preprocessMODIS,list-method}
\title{MODIS Preprocessing Chain}
\usage{
\S4method{preprocessMODIS}{list}(x, vi = c("NDVI", "EVI"), dsn = getwd(),
  ext = NULL, interval = "fortnight", whit = TRUE, cores = 1L, ...)
}
\arguments{
\item{x}{A named \code{list} of local MODIS SDS layers per product as 
returned by \code{\link{runGdal}}.}

\item{vi}{\code{character}. Vegetation index under consideration. Currently 
available options are "NDVI" (default) and "EVI".}

\item{dsn}{\code{character}. Target folder for file output.}

\item{ext}{\code{Extent}, or any object from which an \code{Extent} can be
extracted, see \code{\link[raster]{crop}}. If missing, layer clipping is 
skipped.}

\item{interval}{See \code{\link{temporalComposite}}.}

\item{whit}{\code{logical}. \code{TRUE} (default) invokes gap-filling via 
modified Whittaker smoothing, see \code{\link{whittaker.raster}}.}

\item{cores}{\code{integer}. Number of cores for parallel processing.}

\item{...}{Additional arguments passed to \code{\link{whittaker.raster}}
(except for 'vi' and 'outDirPath').}
}
\value{
A preprocessed \code{Raster*} object.
}
\description{
MODIS preprocessing chain including optional layer clipping, two-fold 
quality-control, fortnightly composite creation, and subsequent combination 
of Terra and Aqua imagery into half-monthly maximum value composites .
}
\seealso{
\code{\link[raster]{crop}}, \code{\link{whittaker.raster}}.
}
\author{
Florian Detsch
}
