% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.thredds.met.no.R
\name{radar}
\alias{radar}
\title{Function to read radar data from thredds.met.no}
\usage{
radar(
  url = "https://thredds.met.no/thredds/catalog/remotesensingradaraccr/",
  lons = c(9.5, 11.5),
  lats = c(59, 61),
  param = "lwe_precipitation_rate",
  FUN = "sum",
  it = 2010:2019,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{url}{URL for the data on thredds.met.no}

\item{lons}{Longitude selection - =NULL reads all}

\item{lats}{Latitude selection - =NULL reads all}

\item{param}{Variable name}

\item{FUN}{Function for daily aggregation. =NULL gives raw data}

\item{it}{Intex time - the years to select}

\item{verbose}{write out diagnostics}

\item{plot}{plot the results while reading.}
}
\description{
Function to read radar data from thredds.met.no
}
\examples{
Z <- radar(lons = c(10.5,11), lats = c(59.5,60))
z <- radar(it=2010)

}
\seealso{
station.thredds, meta.thredds
}
