% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.storms.R
\name{read.imilast}
\alias{read.imilast}
\alias{read.hurdat2}
\title{Read cyclone data}
\usage{
read.imilast(fname, path = NULL, verbose = FALSE)
}
\arguments{
\item{fname}{filename (for \code{read.hurdat}, filename can also be a url)}

\item{path}{path to file}

\item{verbose}{a boolean; If FALSE, do not display comments (silent mode).
If TRUE, display extra information on progress.}
}
\value{
An "events" "data.frame" object containing the date, time, lon, and
lat, as well as additional information (e.g., trajectory number, slp or
other measure of storm strength) of the cyclones.
}
\description{
Methods for reading cyclone data.
}
\details{
\code{read.imilast} reads data from files following the standards of the IMILAST project
(Neu et al., 2013, IMILAST: A Community Effort to Intercompare Extratropical Cyclone Detection
and Tracking Algorithms. Bull. Amer. Meteor. Soc., 94, 529-547,
\url{https://doi.org/10.1175/BAMS-D-11-00154.1}).

\code{read.hurdat2} reads data from the Atlantic hurricane database
(\url{http://www.nhc.noaa.gov/data/#hurdat})
}
\author{
K. Parding
}
\keyword{IMILAST}
\keyword{cyclones}
\keyword{hurdat2}
\keyword{storms}
