% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.ESGF.R
\name{retrieve.ESGF}
\alias{retrieve.ESGF}
\title{Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api}
\usage{
\method{retrieve}{ESGF}(
  param = "tas",
  verbose = FALSE,
  ...,
  url = "https://esgf-data3.ceda.ac.uk/thredds/dodsC",
  CMIP = "esg_cmip6/CMIP6",
  MIP = "ScenarioMIP",
  inst = "MOHC",
  model = "UKESM1-0-LL",
  scen = "ssp585",
  ID = "r2i1p1f2",
  type = "Amon",
  grid = "gn",
  version = "v20190507",
  period = "2015-2100"
)
}
\arguments{
\item{param}{Name of parameter}

\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{url}{The base URL for ESGF}

\item{CMIP}{Which generation of CMIP}

\item{MIP}{Name of activity}

\item{inst}{Institution ID}

\item{model}{Name of climate model}

\item{scen}{Name of scenario}

\item{ID}{The name of the run}

\item{type}{The table ID}

\item{grid}{Grid ID}

\item{version}{Version}

\item{period}{The time segment of the file}
}
\value{
A "zoo" "field" object with additional attributes used for further
processing.
}
\description{
Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api
}
\examples{
retrieve.ESGF(lon=c(-30,40),lat=c(50,70),period='205001-210012') -> X
map(X)
}
\seealso{
retrieve
}
