% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.ESGF.R
\name{retrieve.ESGF}
\alias{retrieve.ESGF}
\alias{meta.ESGF}
\title{Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api
meta.ESGF returns a data.frame with the model metadata and the OpenDAP URL that 
can be used with retrieve.ESGF.}
\usage{
\method{retrieve}{ESGF}(im = 1, meta = NULL, verbose = FALSE, ...)
}
\arguments{
\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}

\item{param}{Name of parameter}

\item{url}{The base URL for ESGF}

\item{n}{Number of models to read - NULL reads everything}

\item{expid}{Name of the MIP experiment ('historical' for historical run)}

\item{mip}{CMIP5 or CMIP6}

\item{freq}{Frequency of data}
}
\value{
A data.frame for \code{meta.ESGF} and a "zoo" "field" object with additional attributes used for further
processing for +code{retrieve.ESGF}.
}
\description{
Retrieve CMIP data directly from the Earth System Grid Federation (ESGF)
https://earthsystemcog.org/projects/cog/esgf_search_restful_api
meta.ESGF returns a data.frame with the model metadata and the OpenDAP URL that 
can be used with retrieve.ESGF.
}
\examples{
\dontrun{
meta <- meta.ESGF(n=3)
X <- retrieve.ESGF(im=3,lon=c(-30,40),lat=c(50,70),meta=meta)
map(X)
}
}
\seealso{
retrieve
}
