% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.station}
\alias{select.station}
\title{Select from meta data base}
\usage{
\method{select}{station}(
  x = NULL,
  ...,
  loc = NULL,
  param = NULL,
  ele = NULL,
  stid = NULL,
  lon = NULL,
  lat = NULL,
  alt = NULL,
  cntr = NULL,
  src = NULL,
  it = NULL,
  nmin = NULL,
  user = "external",
  verbose = FALSE
)
}
\description{
Function that searches the meta data base for the requested station data
Search priority: ID, name, coordinates, altitude, country,...
Can return several matches
}
