% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell.R
\name{spell}
\alias{spell}
\alias{spell.default}
\alias{spell.station}
\alias{hist.spell}
\alias{wetfreq}
\alias{wetmean}
\alias{nevents}
\alias{exceedance}
\alias{exceedance.default}
\alias{exceedance.station}
\alias{exceedance.field}
\alias{HDD}
\alias{CDD}
\alias{GDD}
\alias{qqgeom}
\title{Spell statistics}
\usage{
spell(x, threshold, ...)
}
\arguments{
\item{x}{station or field object}

\item{threshold}{threshold value}

\item{upper}{upper limit for maximum length - ignore any above this because they are likely erronoeus}

\item{fraction}{TRUE: divide the number of counts by number of samples}

\item{FUN}{function}
}
\value{
Station or field objects
}
\description{
Statistics of spell durations (consecutive wet and dry days), e.g. dry and
wet periods or duration of extremes. Similar functions include:
\code{count} returns a count of cases with values over s threshold value.
\code{exceedance} estimates statistics for peak-over-threshold, and
\code{nevents} returns the number of events with exceeding values (e.g. the
number of rainy days X > 1 mm/day).
\code{wetfreq} returns the wet-day frequency (a fraction) and \code{wetmean} returns the wet-day mean.
\code{CDD}: Cooling degree day
\code{GDD}: Growing degree days (\url{http://en.wikipedia.org/wiki/Growing_degree-day})
\code{HDD}: Heating degree day
\code{qqgeom} produces a quantile-quantile plot of streak statistics comparing the
empirical quantiles with the distribution function quantiles (see \code{\link[stats]{qgeom}}).
}
\examples{

# Example 1 : 
data(bjornholt)
x <- spell(bjornholt, threshold=.1)
x.ann <- annual(x, FUN="max")
plot(x.ann, plot.type="multiple",new=FALSE)
# Example 2 :
plot(x, new=FALSE)

# Growing degree days:
data(ferder)
plot(as.seasons(ferder,FUN='GDD'), new=FALSE)

# Mild winter days - number of days in the winter season with
# above freezing temperatures
try(coldwinterdays(ferder))

# Quantile-quantile plot
qqgeom(ferder, treshold=1, pois=TRUE)

}
\seealso{
hotsummerdays coldwinterdays coldspells heatwavespells nwetdays plot
}
\keyword{utilities}
