% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSA.R
\name{SSA}
\alias{SSA}
\title{Singular Spectrum Analysis}
\usage{
SSA(x, m = 12, plot = TRUE, main = "SSA analysis", sub = "",
  anom = TRUE, ip = 1, verbose = FALSE)
}
\arguments{
\item{x}{A station or eof object.}

\item{m}{Window length.}

\item{plot}{Flag: plot the diagnostics.}

\item{main}{main title (see \code{link{plot}}).}

\item{sub}{subtitle (see \code{link{plot}}).}

\item{anom}{TRUE if analysis on anomalies}

\item{ip}{If x is an eof-object, which PC to use.}

\item{verbose}{Print out diagnostics.}
}
\value{
A SSA object: An \code{link{svd}} object with additional parameters:
m (window length), nt (original length of series), Nm (effective length of
series= nt - m), anom (FLAG for use of anomaly), param (name of parameter,
typically 'precip' or 't2m'), station (the station object to which SSA is
applied).
}
\description{
After von Storch & Zwiers (1999), Statistical Analysis in Climate Research,
p. 312
}
\keyword{manip}
