% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{stnr}
\alias{stnr}
\title{MetNo meta data function}
\usage{
stnr(navn = NULL, lon = NULL, lat = NULL, max.dist = 10,
  alt = NULL, Fylke = NULL, Kommune = NULL, fy = NULL, ty = NULL,
  ny = NULL, param = "TAM", plot = FALSE, print = FALSE)
}
\description{
MetNo meta data function
}
