% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dsensemble}
\alias{summary.dsensemble}
\title{Show summary of objects}
\usage{
\method{summary}{dsensemble}(object, ..., years = seq(1990, 2090, by =
  20), verbose = FALSE)
}
\arguments{
\item{object}{an object of type 'DSensemble'}

\item{\dots}{additional arguments}

\item{years}{A set of years for which to produce summary statistics}

\item{verbose}{if TRUE print progress}
}
\value{
A matrix containing summary statistics
}
\description{
Produce a summary table
}
\examples{

data("dse.Oslo")
summary(dse.Oslo)

}
\seealso{
summary.station summary.ncdf4
}
