% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ncdf4}
\alias{summary.ncdf4}
\title{Summary of netcdf file}
\usage{
\method{summary}{ncdf4}(object, ..., verbose = TRUE)
}
\arguments{
\item{verbose}{a boolean; if TRUE print information about progress}

\item{ncfile}{filename of netcdf file}
}
\description{
Summary of netcdf file
}
\seealso{
retrieve check.ncdf4
}
