% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.station}
\alias{summary.station}
\title{Show summary of objects}
\usage{
\method{summary}{station}(x, im = 1:12)
}
\arguments{
\item{x}{an object of type 'station'}

\item{im}{The order of months in the table. Use \code{im=c(10:12,1:9)} for Oct-Sep.}
}
\value{
A matrix containing summary statistics
}
\description{
Produce a summary table
}
\examples{

data("Oslo")
summary(Oslo)

}
\seealso{
summary.dsensemble summary.ncdf4
}
