% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate.dsensemble}
\alias{validate.eof.comb}
\alias{validate.default}
\alias{validate.cca}
\alias{validate.ds}
\alias{validate.eof}
\alias{validate.eof.field}
\alias{validate.pca}
\title{Validate}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{esd object to be validated}

\item{\dots}{other arguments}

\item{conf.int}{confidence interval}

\item{colbar}{for plotting. See \code{colbar}}

\item{plot}{if TRUE produce plot}

\item{verbose}{if TRUE print progress}
}
\description{
The method \code{validate}
}
\examples{

  slp1 <- slp.DNMI(lon=c(-50,50),lat=c(30,70))
  slp2 <- slp.DNMI(lon=c(-50,50),lat=c(30,70))
  slpcomb <- combine(slp1,slp2)
  eofcomb <- EOF(slpcomb)
  validate(eofcomb, new=FALSE)

}
