% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WG.R
\name{WG}
\alias{WG}
\alias{WG.station}
\alias{WG.fwmu.day.precip}
\alias{WG.FT.day.t2m}
\alias{WG.pca.day.t2m.precip}
\alias{FTscramble}
\title{Weather generators for conditioned on simulated climate aggregated
statistics.}
\usage{
WG(x, ...)
}
\arguments{
\item{x}{station object}

\item{\dots}{additional arguments}

\item{option}{Define the type of WG}

\item{amean}{annual mean values. If NULL, use those estimated from x; if NA,
estimate using \code{\link{DSensemble.t2m}}, or if provided, assume a
'dsensemble' object.}

\item{asd}{annual standard deviation. If NULL, use those estimated from x;
if NA, estimate using \code{\link{DSensemble.t2m}}, or if provided, assume a
'dsensemble' object.}

\item{t}{Time axis. If null, use the same as x or the last interval of same
length as x from downscaled results.}

\item{ip}{passed on to \code{\link{DSensemble.t2m}}}

\item{select}{passed on to \code{\link{DSensemble.t2m}}}

\item{lon}{passed on to \code{\link{DSensemble.t2m}}}

\item{lat}{passed on to \code{\link{DSensemble.t2m}}}

\item{plot}{if TRUE, plot results}

\item{biascorrect}{passed on to \code{\link{DSensemble.t2m}}}

\item{verbose}{passed on to \code{\link{DSensemble.t2m}}}

\item{mu}{annual wet-mean values. If NULL, use those estimated from x; if
NA, estimate using \code{\link{DSensemble.t2m}}, or if provided, assume a
'dsensemble' object.}

\item{fw}{annual wet-day frequency. If NULL, use those estimated from x; if
NA, estimate using \code{\link{DSensemble.t2m}}, or if provided, assume a
'dsensemble' object.}

\item{ndd}{annual mean dry spell length. If NULL, use those estimated from
x; if NA, estimate using \code{\link{DSensemble.t2m}}, or if provided,
assume a 'dsensemble' object.}

\item{threshold}{Definition of a rainy day.}

\item{method}{Assume a gemoetric or a poisson distribution. Can also define
ownth methods.}

\item{t2m}{station object with temperature}

\item{precip}{station object with precipitation.}

\item{ndbr}{Number of}

\item{n.spells.year}{= c('fw','spell') if 'fw' then estimate number of spells according to 365.25 else estimate number of events from \code{\link{spell}}.}

\item{alpha.scaling}{TRUE scale the low-probability events according to alpha in DOI:10.1088/1748-9326/ab2bb2}

\item{alpha}{values for alpha-scaling}

\item{ensure.fw}{TRUE then WG tries to ensure that fw of simulations match those of observations or prescribed by adding or substracting wet days.}
}
\description{
Weather generators for conditional simulation of daily temperature and/or
precipitation, given mean and/or standard deviation. The family of WG
functions produce stochastic time series with similar characteristics as the
station series provided (if none if provided, it will use either ferder or
bjornholt provided by the esd-package). Here characteristics means similar
mean value, standard deviation, and spectral properties. \code{FTscramble}
takes the Fourier components (doing a Fourier Transform - FT) of a series
and reassigns random phase to each frequency and then returns a new series
through an inverse FT. The FT scrambling is used for temperature, but not
for precipitation that is non-Gaussian and involves sporadic events with
rain. For precipitation, a different approach is used, taking the wet-day
frequency of each year and using the wet-day mean and ranomly generated
exponentially distributed numbers to provide similar aggregated annual
statistics as the station or predicted though downscaling. The precipitation
WG can also take into account the number of consequtive number-of-dry-days
statistics, using either a Poisson or a gemoetric distribution.
}
\details{
The weather generater produces a series with similar length as the provided
sample data, but with shifted dates according to specified scenarios for
annual mean mean/standard deviation/wet-day mean/wet-day frequency.

\code{WG.FT.day.t2m} generates daily temperature from seasonal means and
standard deviations. It is given a sample station series, and uses
\code{FTscramble} to generate a series with random phase but similar (or
predicted - in the future) spectral characteristics. It then uses a quantile
transform to prescribe predicted mean and standard deviation, assuming the
distributions are normal. The temperal structure (power spectrum) is
therefore similar as the sample provided.

\code{WG.fwmu.day.precip} has been designed to use with downscaled results for 
annual wet-day frequency and annual wet-day mean precipitation. 
It uses the annual wet-day mean precipitation and the wet-day
frequency as input, and takes a sample station of daily values to
simulate stochastic numbers of consecutive wet days, based on its annual mean
number of consecutive wet days. It also uses the mean annual cycle of wet-day frequency
as well as the wet-day mean precipitation to guide the timing of wet days and amounts, 
and hence tries to mimic rain seasons. If not specified, it is taken from the sample data 
after being phase scrambled/shuffled (\code{FTscramble} - a bit like a deck of cards). 
The number of wet events per year is estimated from the cumulative sum of wet days that 
needs to match 366 times the wet-day frequency. The annual wet-day frequency is a
lso taken to be phase scrambled estimates from the sample data unless specifically 
specified. The daily amount is taken from stochastic values generated with 
\code{\link{rexp}} scaled for the tail according to alpha in 
(described in DOI: 10.1088/1748-9326/ab2bb2) as in \code{\link{day2IDF}}. 
The number of consecutive wet days can be approximated by a geometric distribution 
(\code{\link{rgeom}}), and the annual mean number was estimated from the sample series.
\code{test.WG.fwmu.day.precip} presents diagnostics of tests of \code{WG.fwmu.day.precip}.
}
\examples{
## Temperature
data(ferder)
x <- WG(ferder)
## Plot the results
plot(merge(ferder,x), xlab='', ylab=c('Obs T2m','WG T2m'), col='blue', main=paste(loc(x),' Obs/WG'))

## Daily precipitation
data(bjornholt)
z <- WG(bjornholt)
## Plot the results
plot(merge(bjornholt,z), xlab='', ylab=c('Obs precip','WG precip'), col='blue', main=paste(loc(z),' Obs/WG'))
sz <- sort(coredata(z))
sy <- sort(coredata(bjornholt))
## Use WG to 'simulate' climate change
z2 <- WG(bjornholt, mu=annual(bjornholt, FUN='wetmean') + 2)
sz2 <- sort(coredata(z2))

## Plot the comparison of quantiles
plot(sy, sz, pch=19, cex=0.7, main='QQ-plot', xlab='Observations', ylab='WG')
grid()
lines(c(0, max(sy,sz,na.rm=TRUE)), c(0,max(sy,sz,na.rm=TRUE)), lty=2, col='red')
points(sy, sz2, col='blue', cex=0.7)

## Test the WG
test.WG.fwmu.day.precip()
}
\author{
R.E. Benestad
}
\keyword{manip}
