% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel.R
\name{wheel}
\alias{wheel}
\alias{wheel.station}
\alias{wheel.spell}
\title{InfoGraphics}
\usage{
wheel(x, ...)
}
\arguments{
\item{x}{an input object of class 'station' or 'spell'}

\item{\dots}{additional arguments}

\item{y}{an input object of class 'station' or 'spell'}

\item{new}{if new create new graphic device}

\item{lwd}{relative line width}

\item{col}{color of line}

\item{type}{'spiky' or 'flowy'}

\item{bg}{background color}

\item{verbose}{a boolean; if TRUE print information about progress}
}
\description{
Various functions for visual display of data and statistics
}
\details{
\code{wheel} shows the seasonal cycle with different colors for different years
}
\examples{
data(bjornholt)
wheel(bjornholt, new=FALSE)

}
\seealso{
graph visprob conf vis diagram cumugram scatter plot map
}
