% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\alias{TGW}
\alias{geostrophicwind}
\alias{geostrophicwind.station}
\alias{geostrophicwind.field}
\title{Wind analysis}
\usage{
windrose(x, saw = 10, max.scale = NULL, main = NULL,
  cols = c("grey90", "yellow", "green", "red", "blue", "darkgreen",
  "darkred", "magenta", "black"), param = c("u", "v"), simple = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{station object}

\item{saw}{Directional resolution in degrees}

\item{max.scale}{scaling factor for windrose}

\item{cols}{a vector defining colors for plot}

\item{param}{Name of the variables representing zonal and meridional wind}

\item{simple}{Only plot the windrose, not an additional historgram for
windspeed}

\item{verbose}{if TRUE print information about progress}

\item{thrshld}{Threshold for windspeed}
}
\description{
A function that plots windroses from station objects which contain both the
zonal and meridional components. These are stored seperately as if they were
two different station records.
}
\note{
Adapted from clim.pact
}
\examples{

\dontrun{
slp <- station(param='slp',cntr='Denmark',src='ecad')
uv <- TGW(subset(slp,is=c(1,3,11)))
UV <- geostrophicwind(slp)
windrose(uv)
map(UV,FUN='q95')
}

}
\seealso{
\code{link{geostrophicwind}}
}
\author{
R.E. Benestad
}
