% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2ncdf.R
\name{write2ncdf4}
\alias{write2ncdf4}
\title{Saves climate data as netCDF.}
\usage{
write2ncdf4(x, ...)
}
\arguments{
\item{x}{data object}

\item{\dots}{additional arguments}
}
\value{
None
}
\description{
Method to save station data as netCDF, making sure to include the data
structure and meta-data (attributes). The code tries to follow the netCDf
'CF' convention. The method is built on the \code{ncdf4} package.
}
\examples{

nacd <- station(src='nacd')
X <- annual(nacd)
write2ncdf4(X,file='test.nc')

}
\seealso{
write2ncdf4.station 
write2ncdf4.field write2ncdf4.list write2ncdf4.station write2ncdf4.eof
write2ncdf4.pca write2ncdf4.dsensemble
}
\keyword{ncdf4}
\keyword{netcdf}
\keyword{save}
