% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2ncdf.R
\name{write2ncdf4.station}
\alias{write2ncdf4.station}
\title{Saves climate data as netCDF.}
\usage{
\method{write2ncdf4}{station}(
  x,
  ...,
  file = "station.nc",
  prec = "short",
  offset = 0,
  missval = -99,
  it = NULL,
  stid = NULL,
  append = FALSE,
  scale = 0.1,
  torg = "1899-12-31",
  stid_unlim = FALSE,
  namelength = 24,
  nmin = 30,
  verbose = FALSE,
  doi = "NA",
  namingauthority = "NA",
  processinglevel = "NA",
  creatortype = "NA",
  creatoremail = "NA",
  institution = "NA",
  publishername = "NA",
  publisheremail = "NA",
  publisherurl = "NA",
  project = "NA",
  nspc = 300
)
}
\arguments{
\item{x}{data object}

\item{\dots}{additional arguments}

\item{file}{file name}

\item{prec}{Precision: see \code{\link[ncdf4]{ncvar_def}}}

\item{offset}{Sets the attribute 'add_offset' which is added to the values
stored (to save space may be represented as 'short').}

\item{missval}{Missing value: see \code{\link[ncdf4]{ncvar_def}}}

\item{it}{a time index, see \code{\link{subset}}}

\item{stid}{station id}

\item{append}{a Boolean; if TRUE append output to existing file}

\item{scale}{Sets the attribute 'scale_factor' which is used to scale
(multiply) the values stored (to save space may be represented as 'short').}

\item{torg}{Time origin}

\item{stid_unlim}{a Boolean; if TRUE the stid dimension is unlimited}

\item{namelength}{a numeric specifying the number of characters in dimension and variable names}

\item{nmin}{Only calculate summary statistics for stations with nmin years of data (e.g. 30 years).}

\item{verbose}{TRUE - clutter the screen.}

\item{doi}{- Data ID. All the following arguments are meant to accommodate for the convention described 
at \url{https://adc.met.no/node/4}}

\item{nspc}{- the size of chunks of data processes sequentially in order to limit the need of computer memory. 
Smaller number requires less memory.}
}
\value{
None
}
\description{
Method to save station data as netCDF, making sure to include the data
structure and meta-data (attributes). The code tries to follow the netCDf
'CF' convention. The method is built on the \code{ncdf4} package.
}
\details{
To save space, the values are saved as short (16-bit signed integer that
can hold values between -32768 and 32767).
(see NC_SHORT in \url{https://www.unidata.ucar.edu/software/netcdf/docs/data_type.html}).
}
\seealso{
write2ncdf4
}
\keyword{ncdf4}
\keyword{netcdf}
\keyword{save}
