% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{year}
\alias{year}
\alias{month}
\alias{season}
\alias{season.default}
\title{Conversion to esd objects.}
\usage{
year(x)
}
\arguments{
\item{x}{an object of, e.g., class 'station', 'field', or 'zoo', or a date}

\item{format}{for season, set the format of the output 'character' or 'numeric'}
}
\value{
a numeric for \code{year}, \code{month}, and \code{day}; A numeric or character for \code{season}
}
\description{
\code{year}, \code{month}, \code{day}, \code{season} return the years, months, days, and seasons
associated with the data.
}
\examples{
data(bjornholt)
year(bjornholt)
month(bjornholt)
day(bjornholt)
season(bjornholt)
season(bjornholt, format="numeric")

}
\keyword{utilities}
